### 2021-11-22 #3942 @whedon
[PRE REVIEW]: epwshiftr: Create future EnergyPlus Weather files using CMIP6 data
**Submitting author:** @hongyuanjia (<a href="http://orcid.org/0000-0002-0075-8183">Hongyuan Jia</a>)
**Repository:** <a href="https://github.com/ideas-lab-nus/epwshiftr" target ="_blank">https://github.com/ideas-lab-nus/epwshiftr</a>
**Version:** v0.1.3
