+++ pkgdown.yml
@@ -3,5 +3,5 @@ pkgdown: 1.6.1
 pkgdown_sha: ~
 articles:
   paper: paper.html
-last_built: 2021-11-22T11:11Z
+last_built: 2021-11-22T15:19Z
 
