### 2021-11-29 #3942 @hongyuanjia
Hi @KristinaRiemer.

First, thank you for accepting as the editor for my submission!

I went through [the list of JOSS reviewers](https://bit.ly/joss-reviewers) but did not find any R experts that work on this topic. I list folks that work on a similar topic but use python ([znicholls](https://github.com/znicholls)) or R experts that may be a good match for this review, based on their GitHub record. See: https://github.com/openjournals/joss-reviews/issues/3942#issuecomment-975627753

Also, I searched all available CRAN packages and found developers below who worked on similar topics and maybe a good suit for this review. But since they are not on the JOSS reviewer list, I am not sure whether they are available/willing for doing this review.

- Michael Sumner ([mdsumner](https://github.com/mdsumner)): Spatial expert and the developer of a number of famous R spatial packages.
- Michael Matiu ([mitmat](https://github.com/mitmat)):  [eurocordexr](https://cran.r-project.org/package=eurocordexr) maintainer.
- Blake Seers ([blasee](https://github.com/blasee)): [clifro](https://cran.r-project.org/package=clifro) maintainer
- Bartosz Czernecki ([bczernecki](https://github.com/bczernecki)): [climate](https://cran.r-project.org/package=climate) maitainer
- Hernando Cortina ([cortinah](https://github.com/cortinah)): [hockeystick](https://cran.r-project.org/package=hockeystick) maintainer
