### 2021-11-29 #3942 @KristinaRiemer
Luckily, we are not limited to the JOSS reviewer list, that list is just a good place to start. Anyone who feels they have the right skill set (and no conflict of interest with the authors) can be a reviewer.

I've reached out to a few folks and I'll use your list too. Thanks!
