### 2021-12-31 #4030 @whedon
[REVIEW]: epwshiftr: Create future EnergyPlus Weather files using CMIP6 data
**Submitting author:** @hongyuanjia (<a href="http://orcid.org/0000-0002-0075-8183">Hongyuan Jia</a>)
**Repository:** https://github.com/ideas-lab-nus/epwshiftr
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.3
