### 2021-12-31 #3942 @KristinaRiemer
Hi @mitmat and @bczernecki! Thank you again for agreeing to review this JOSS submission. The review itself will happen in the GitHub issue linked to above. Feel free to let me know if you have questions about this process or the submission there.
