### 2022-01-05 #4030 @KristinaRiemer
Hi @hongyuanjia, you might want to review the "What should my paper contain?" section of the [JOSS author guide](https://joss.readthedocs.io/en/latest/submitting.html) to ensure your paper has the required sections in it.
