### 2022-01-07 #4030 @hongyuanjia
Hi @KristinaRiemer, thanks for the remainder. I will update the paper structure accordingly based on the guide.
