### 2022-01-12 #4030 @mitmat
Hi @hongyuanjia,

I started to review the package. I got to test the functionality until
https://github.com/ideas-lab-nus/epwshiftr#extract-cmip6-output-data
but since I do not have EnergyPlus, I cannot continue. Is it possible to put some example  epw file somewhere, so I can continue? Maybe even the one from the example, so I can check it out...
