### 2022-01-13 #4030 @hongyuanjia
Hi @mitmat, thanks for the review. EnergyPlus is an open-source building energy simulation. It provides free available worldwide weather data at https://energyplus.net/weather. The one I used in the example can be downloaded at the EnergyPlus GitHub repo: https://raw.githubusercontent.com/NREL/EnergyPlus/develop/weather/USA_CA_San.Francisco.Intl.AP.724940_TMY3.epw.

Please let me know if you encounter any issue about accessing the data.
