### 2022-01-14 #4030 @mitmat
I got through the examples in the repo readme. Package works like a charm. Already the first part of managing the many files is solved nicely. I have no real issues with the functionality, but the documentation needs to be improved! (see also the issues in the repo).

Comments to the paper will come later…

**License:**

You have two license files in the repo, the one is MIT (as specified on CRAN, too), and the other is a personal copyright (non-permissive, I guess?). For JOSS, an open license is required, so please check this.

**Contribution and authorship:**

From the github commit history, it’s clear that the main author has made major contributions to the package. The co-author’s contribution is not clear: his extremely few commits are not of substantial nature, so please clarify the author list and contributions.

**Functionality**

While using the package, one likely will have to work with many files. I like the approach you’ve taken to manage a file index. However, I suggest to explain better how you set the custom directory, and maybe give some hints on why one should do this.

In `init_cmip6_index()`, if you select year 2050, why do you also get the previous and next year?

What is `future_epw()` creating? The files seem to have the same years, but the values are changed.

How do you deal with non-standard calendars in GCMs? Noleap or 360?

**Documentation**

The `esgf_query()` and `init_cmip6_index()` functions have in theory a lot of options to query the ESGF servers. However, your package is designed for CMIP6, so maybe better to not allow all values to changed?

BTW, out of curiosity, are you aware of any packages in R that query ESGF servers? And if yes, do they have the option of login? I know of something in python, but have not seen anything for R yet.
