### 2022-01-14 ideas-lab-nus/epwshiftr#28 @mitmat
joss-review pkg readme
- [ ] Missing statement of need.
- [ ] Please consider adding an example on how to download the CMIP6 files, if you already have the URL and data directory (https://github.com/ideas-lab-nus/epwshiftr#manage-cmip6-output-files).
- [ ] Also for the epw file, you can add the URL you gave me in https://github.com/openjournals/joss-reviews/issues/4030#issuecomment-1011701997 I just saw, you have in the example section of `match_coord()` the code how to download an epw file. Maybe just put the same in the get started section.
- [x] The contribute section needs to be improved. See also: https://joss.readthedocs.io/en/latest/review_criteria.html#community-guidelines You can also check how other packages in JOSS have dealt with this.
