### 2022-01-14 ideas-lab-nus/epwshiftr#29 @mitmat
joss-review documentation of return value (data.table) likely wrong/outdated
I found the the description of the data.table output in the help differs from what I get, when I run the two functions below. Please update.

- [ ] `extract_data()`
- [ ] `morphing_epw()`
