### 2022-01-14 ideas-lab-nus/epwshiftr#30 @mitmat
joss-review Usage examples
The usage examples are often not run, and require auxiliary data that is not in the packages. Do you think it’s possible to add some data to the package, so the examples can be run? I know that putting raw GCM data in there is not possible, otherwise the package size will explode. But maybe crop some netcdf file to a little extent, and put some epw file, and then you can have real examples that can be run? What do you think?
