### 2022-01-14 ideas-lab-nus/epwshiftr#31 @mitmat
joss-review Description of morphing
The term *morphing* is used in building simulations, but rarely in other fields. Can you update the documentation in `morphing_epw()` and explain what you mean by stretching and shifting? Maybe even with some formulas? You can still keep the reference to the paper, but some basic info would be very useful for users.

Otherwise, you could also mention the terms *bias adjustment* and *downscaling*, since these two are more common in other disciplines.
