### 2022-01-18 #4030 @mitmat
@hongyuanjia @KristinaRiemer

I finished the review. Please find in a post above the comments to the package, and some new issues in the pkg repo.

As **addon to above**:

The function `summary_database()` provides in its summary start and end dates, but no info, whether you have complete coverage of all the years in between. Is this intended so?

For the morphing methods, or generally when working with climate data, one needs to use climatological averages (i.e. 30 year averages). Maybe you could give the users a warning if they fail to do so, when calling the morphing function?

Finally, regarding the **paper**:

The package is nice in a way that it provides the full process from managing climate model data, performing bias adjustment, and giving output in the desired format.

Section on “Statement of need” is missing

I wonder how it compares to existing packages, which of course do not have all the functionality, but maybe parts? Is there anything in R for manging CMIP6 (or CMIP5, for that matter), or other ESGF data? Bias adjustment and downscaling is an active topic, I’m sure some packages exist? (-> missing “state of the field”)

Some disclaimer on using climate data would be nice (maybe also in the github repo). It’s great that the climate data is freely available, but still users need to be made aware of what climate model data can or cannot do. See e.g. the well-made guide from EURO-CORDEX https://www.euro-cordex.net/imperia/md/content/csc/cordex/guidance_for_euro-cordex_climate_projections_data_use__2021-02_1_.pdf

The term morphing is typical for building simulations, but rarely used outside of the field. You might want to mention the term “bias adjustment” or “downscaling”, which are commonly used. https://hypeweb.smhi.se/what-is-bias-adjustment/

Minor comments:

L13: GCM usually refers to General Circulation Model, not Global Climate model
L18: “will be used” should be in the past, since CMIP6 has already been used in the IPCC AR6.
L20: which existing tools?
L21-23: I do not agree. There are a lot of tools in the wild. Maybe not specifically targeted to BES, but for “general” climate research, there are many existing tools, one is e.g. https://github.com/SantanderMetGroup/climate4R and others.

And the last thing, out of personal curiosity, not necessarily relevant for the review:
Is there any chance to extract information directly from nc files, without needing to download everything? Since you only extract point information, this could save a lot of space and bandwidth. Do you if that is possible?
