### 2022-01-22 #4030 @hongyuanjia
Thank you @mitmat for taking the time to give such thorough reviews of epwshiftr! I will start to address all your comments and issues opened in the epwshiftr repo in the following days.

> I got through the examples in the repo readme. Package works like a charm. Already the first part of managing the many files is solved nicely. I have no real issues with the functionality, but the documentation needs to be improved! (see also the issues in the repo).
I agree that the documentation needs improvements, especially for the description of the algorithm used to statistical downscaling climate data.

> #### License:
>
> You have two license files in the repo, the one is MIT (as specified on CRAN, too), and the other is a personal copyright (non-permissive, I guess?). For JOSS, an open license is required, so please check this.

epwshiftr is released under MIT license. Unlike GPL license, MIT license is a template that requires additional details to be complete in the `LICENSE` file. CRAN requires packages to specify the `License` to be `MIT + file LICENSE` where the `LICENSE` contains only two fields: the year and copyright holder, which will be filled into the MIT template, see https://cran.r-project.org/web/licenses/MIT. CRAN does not allow packages to include the full text of standard licenses like MIT.

As described in https://r-pkgs.org/license.html#key-files, it is common practice to include a copy of the license file in open-source software. Since CRAN does not permit to include a copy of standard licenses in packages, the full copy of the license is included in `.Rbuildignore` to make sure this file is not sent to CRAN. So that's why there are 2 "license" files in the repo:

- `LICENSE`: The one required by CRAN which includes the year and copyright holder data for the MIT license.
- `LICENSE.md`: The full MIT license.

> #### Contribution and authorship:
>
> From the github commit history, it’s clear that the main author has made major contributions to the package. The co-author’s contribution is not clear: his extremely few commits are not of substantial nature, so please clarify the author list and contributions.

The epwshiftr was developed during my postdoc and was supported by a research project of Adrian, the co-author. Even Adrian did not commit to the repo that much, but he contributed a lot to the development underneath via conceptualization and supervision through lots of regular discussions on project direction and design.

> Functionality
>
> While using the package, one likely will have to work with many files. I like the approach you’ve taken to manage a file index. However, I suggest to explain better how you set the custom directory, and maybe giving some hints on why one should do this.

Thanks for the suggestion. The user can use the option `epwshiftr.dir` to control where the file index is saved. The default data storage is the OS-specific user data directory using the `rappdirs::user_data_dir()` in order to comply with [CRAN policies](https://cran.r-project.org/web/packages/policies.html). There is also a function `get_data_dir()` to retrieve the current data directory. I agree that some hints may be needed to explain when to change the data directory, especially the case when one may want to keep file index and climate data in one place. I will add some documentation for this.

> In `init_cmip6_index()`, if you select year 2050, why do you also get the previous and next year?

This relates to the way that epwshiftr handles datetime data in NetCDF files, as you also mentioned in the following question:

> How do you deal with non-standard calendars in GCMs? Noleap or 360?

Most of output NetCDFs from CMIP6 GCMs use [Proleptic Gregorian calendar](https://en.wikipedia.org/wiki/Proleptic_Gregorian_calendar). epwshiftr uses `RNetCDF::utcal.nc()` to correctly handle the conversion. However, it does not handle Noleap or 360 calendars. Thanks for catching it. I will make a patch to epwshiftr for non-standard calendar handling.

Why I need the previous and next year is that, taking `tas_day_EC-Earth3_ssp585_r1i1p1f1_gr_20590101-20591231.nc` as an example, which uses Proleptic Gregorian calendar with unit `days since 1850-01-01 00:00:00`. The range of the date time entry is [76336.5, 76700.5], which represents `2059-01-01 12:00:00 UTC` and `2059-12-31 12:00:00 UTC`. In order to get a full year of 2059, we need the previous and next half day.

> What is `future_epw()` creating? The files seem to have the same years, but the values are changed.

`future_epw()` is used to format the morphed climate data into the [EnergyPlus Weather file (EPW)](https://bigladdersoftware.com/epx/docs/9-6/auxiliary-programs/energyplus-weather-file-epw-data-dictionary.html#energyplus-weather-file-epw-data-dictionary) which can be directly used by large number of building performance simulation software. The mostly widely used data format for EPW is probably TMY3 ([Typical Meteorological Year](https://bigladdersoftware.com/epx/docs/9-6/auxiliary-programs/source-weather-data-formats.html#source-weather-data-formats)). To calculate a TMY, a multiyear data set is analyzed and 12 months are chosen from that time frame that best represents the median conditions. For example, a TMY developed from a set of data for the years 1998–2005 might use data from 2000 for January, 2003 for February, 1999 for March, and so on. So when inserting morphed climate data back into the EPW file, I kept the year value unchanged to keep the typical year information. A comment is inserted in the `COMMENTS 1` field giving some details about the data, e.g. `This climate change adapted weather file, which bases on 'ssp126-2080s-INM-CM5-0' ensemble data...`.
