### 2022-01-24 ideas-lab-nus/epwshiftr#32 @hongyuanjia
Cannot handle non-standard calendar, e.g., 360_day and noleap 
See https://github.com/openjournals/joss-reviews/issues/4030#issuecomment-1019049222
