### 2022-01-24 d3e3177
[CI] Use r-lib/actions v2 (#34)
