+++ pkgdown.yml
@@ -1,7 +1,7 @@
-pandoc: 2.7.3
-pkgdown: 1.6.1
+pandoc: 2.14.2
+pkgdown: 2.0.2
 pkgdown_sha: ~
 articles:
   paper: paper.html
-last_built: 2021-11-22T15:19Z
+last_built: 2022-01-24T09:31Z
 
