### 2022-02-01 #4030 @KristinaRiemer
It looks like @hongyuanjia has started to respond to @mitmat's comments, thank you for your thoroughness with those! And I agree with the conclusion about authorship; intellectual contributions are more than sufficient for authorship, see the [JOSS authorship guidelines](https://joss.readthedocs.io/en/latest/submitting.html#authorship).

@bczernecki it looks like you've started your review, how is it going?
