### 2022-02-02 #4030 @bczernecki
@hongyuanjia @KristinaRiemer

I have quite thoroughly investigated content of the `epwshiftr` R package and it looks quite promising. However, IMHO there's still a big space for further improvements which needs to be addressed before accepting the software for publication in JOSS and recommending the software for wider audience.

Most of my concerns and comments stay in line with everything that was previously suggested by @mitmat, i.e.
- LICENSE
- contribution of co-author. I have checked that in Git repo there's around 10 lines of code that were touched by the co-author. To be honest - it looks like ghost-writing and I would strongly recommend to clearly address contribution of both authors and remove co-authors with contribution that is less than ~5%
- Documentation (more details later on)
- Unit / Integration tests - Currently `testthat` returns around 60% of code coverage. This is definitely not enough for good quality software that can be used in scientific / commercial / engineering purposes. In most companies the lower limit is set to 80-85%

Other comments are related to the problems that I found while working with the examplary workflow suggested by README file:

- in the "create a CMIP6 output file index" `idx <- init_cmip6_index` source argument is not clear by reading built-in documentation - I think there should be some extra step that would give a hint which GCMs are available as these mentioned in the code might become outdated quite soon. It is also not clear whether only the mentioned GCMs are supported or user can provide its own (playing with the code show that the latter option is applicable in here), and therefore I would highly recommend to add an extra step that will tell an user which GCMs can be used here.

- According to the `README` file, section: Manage CMIP6 output files: "_You have to download CMIP6 output file by yourself using your preferable methods or tools. The download url can be found in the file_url column in the index."_ Frankly speaking I don't understand why this capability is not supported by the package? I tried to do it with the build-in `download.file` command and it worked for me:

```
> for (i in seq_along(idx$file_url)){
+   download.file(idx$file_url[i], basename(idx$file_url[i]))
+ }
attempting URL address 'http://esgf3.dkrz.de/thredds/fileServer/cmip6/ScenarioMIP/DKRZ/MPI-ESM1-2-HR/ssp585/r1i1p1f1/day/hurs/gn/v20190710/hurs_day_MPI-ESM1-2-HR_ssp585_r1i1p1f1_gn_20450101-20491231.nc'
Content type 'application/x-netcdf' length 323106447 bytes (308.1 MB)
==================================================
downloaded 308.1 MB

```

I think developer should write a solution that basically checks whether downloading in R is possible and if not then provide message or even recommend external packages for this purposes that can be implemented with epwshiftr. Moreover, it is not clear to me where this NetCDF files should be downloaded (into working directory or tmp folder that was created before?). I could find it in the next step, but would be good to know it in advance

- Line:  `epw <- file.path(eplusr::eplus_config(8.8)$dir, "WeatherData/USA_CA_San.Francisco.Intl.AP.724940_TMY3.epw")` gives warning: `Failed to find configuration data of EnergyPlus v8.8 `
What I had to do on MacOS: `eplusr::install_eplus("latest")` which basically resolved the problem. I think it should be clearly defined in the README file about this mandatory step. Some extra code that would detect whether it is needed or not would be also beneficial. In my case I ended up with:
`epw <- file.path(eplusr::eplus_config(9.4)$dir, "WeatherData/USA_CA_San.Francisco.Intl.AP.724940_TMY3.epw")`, so it looks like defualt versioning might be different for different operating systems.

- Line: `coord <- match_coord(epw, threshold = list(lon = 1, lat = 1), max_num = 1)` - as far as I understood the code, it assumes that distance on longitude and latitude is the same which is not true (except equator). It has to be corrected e.g. by utilizing euclidean or spheroid formulas (e.g. https://github.com/bczernecki/climate/blob/dev/R/spheroid_dist.R)

- **there's sth wrong in parsing coordinates !!!** Please take a look that `knitr::kable(head(data$data))` - it gives longitude 301.875, which in reality should be [1] -122.4; I am wondering whether the data taken for calculation in next steps is accurrate

- It is not clear to me whether any bias correction method is applied and if so which one is utilized in the package. It is utterly important as some data converted in the package do not follow gaussian distribution and it would require some extra domain knowledge. I would add extra message about bias correction or lack of such a method. If there's not bias correction then I would highly recommend utilizing one of available packages.
