### 2022-02-11 #4030 @KristinaRiemer
@hongyuanjia have you had a chance to look over @bczernecki's comments yet?

@bczernecki, with regards to coauthorship, our [JOSS authorship guidelines](https://joss.readthedocs.io/en/latest/submitting.html#authorship) err on the side of letting the authors decide who has made meaningful enough non-code contributions to warrant authorship. It sounds like those contributions were sufficient in this case.

About the test coverage, maybe it would be more useful to determine if the current set of tests cover the most important or error-prone functionality of the software, as opposed to just a threshold for test coverage percent? Are there parts of this code that currently lack tests but could really benefit from them?
