### 2022-02-13 ideas-lab-nus/epwshiftr#35 @hongyuanjia
Code coverage dropped because of morphing tests were not run
The code coverage dropped down to ~63%. This is because that morphing related tests were not run as the `.fst` file from test results of `extract_data()` were not ready
