### 2022-02-13 76ff383
[test] Make sure tests data is ready before morphing tests (#36)
