### 2022-02-13 #4030 @hongyuanjia
@KristinaRiemer Sorry for the late. I just came back to work from the Chinese New Year holiday. I will start to address @bczernecki comments in the following week.

### License
Thank you @KristinaRiemer for further elaboration on the coauthorship requirement of JOSS.
Please see https://github.com/openjournals/joss-reviews/issues/4030#issuecomment-1019049222. Adrian contributed a lot to the development underneath via conceptualization and supervision through lots of regular discussions on project direction and design. I believe his contribution is sufficient enough to be on the author list.

### Unit tests
Actually, the test coverage of epwshiftr is 100%. The drop of coverage to ~63% is due to the undesired test execution order. Please see https://github.com/ideas-lab-nus/epwshiftr/issues/35. This has been fixed in https://github.com/ideas-lab-nus/epwshiftr/pull/36.

### Available GCMs (`source`)
Thanks for the suggestions. `init_cmip6_index()` is calling `esgf_query()` underneath, which will send queries using [ESGF search RESTful API](https://esgf.github.io/esg-search/ESGF_Search_RESTful_API.html). The `source` argument indicates the name of GCM ([General Circulation Models](https://www.ipcc-data.org/guidelines/pages/gcm_guide.html)) as specified by the ESGF API. Currently there are ~110 different GCMs included in various CMIP6 activity. The full list of GCMs can be found [online](https://pcmdi.llnl.gov/CMIP6/ArchiveStatistics/esgf_data_holdings/).  I will update the documentation to include the webpage of the GCM full list.

Since the GCM outputs listed all follow the [CMIP standards](https://docs.google.com/document/d/1os9rZ11U0ajY7F8FWtgU4B49KcB59aFlBVGfLC4ahXs/edit), the parsing functionalities in epwshiftr should work fine on all the GCMs. The criteria for selecting the default 11 GCMs were the resolution scale (daily), projection variables (giving outputs most related to climate data for building energy simulation), and the generation of all 4 experiments of activity `ScenarioMIP`. This follows the [same approach](https://www.weathershift.com/Morphing%20Climate%20Data.pdf) for the commercial service [WeatherShift](https://www.weathershift.com/). Actually, the initial motivation of epwshiftr development was creating a free open-source package that can replicate similar outputs as WeatherShift and [CCWorldWeatherGen](https://energy.soton.ac.uk/climate-change-world-weather-file-generator-for-world-wide-weather-data-ccworldweathergen/) for building energy modelers, but updated for the latest CMIP6 data.

### Downloading GCM outputs

#### Why not provide functionality to download using the package itself?

I truly agree it would be useful if epwshiftr could provide functionality to download netCDF files for GCMs. However, I find that it is hard to maintain a fast, easy-to-use and reliable approach to download them because of the following reasons:

**File size**
All CMIP6 output has been written to netCDF files with one variable stored per file. The size of a single GCM output highly depends on the date period simulated, grid resolution, and time resolution. The file you mentioned is for humidity data for a single emission scenario for only 5y (2045 ~ 2049). Some GCM, e.g. `MRI-ESM2-0`, compiles 50y data into one netCDF and the file size explosive grown to ~2.6G. The total file size of humidity data from `MRI-ESM2-0` for all 4 emission scenarios for 2020-2100 is >32G. Let alone that it is normal to consider various climate variables from multiple GCMs. The file size can easily go up to hundreds of GBs.
**Unstable data node**
Moreover, CMIP6 model outputs are hosted on a [collection of nodes](https://esgf-node.llnl.gov/status/) across the world and some nodes could down which makes certain data inaccessible at some point. epwshiftr provides [`get_data_node()`](https://ideas-lab-nus.github.io/epwshiftr/reference/get_data_node.html) to help the user retrieve the data node status, including ping responses.

I will update the package document to describe more on why the decision was made to let the user handle the downloading itself.

#### Where to put the netCDF files
Actually, the location of the netCDF files does not matter except that all files should be placed in a single folder (can be in different subdirectories). epwshiftr lets the user to use the `dir` argument in [`summary_database()`](https://ideas-lab-nus.github.io/epwshiftr/reference/summary_database.html) to specify where it should scan the netCDF files the user downloaded against the index file created using `init_cmip6_index()`. I will update the example in README to provide more informative guide on this.

### EPW file in the example in README
The example in the README focuses on the basic workflow of using epwshiftr and hides a lot of other information including what [EnergyPlus](https://energyplus.net/) is and where to get EnergyPlus weather files (EPW). When I wrote the example, I assumed that the target users of epwshiftr package were building energy modelers who have installed and use EnergyPlus. Sorry for the trouble when trying to access the EPW file included in the example.

Every version of EnergyPlus comes along with several example EPW files. They do not change across different versions.
`eplusr::eplus_config(8.8)$dir` was used to get the installation folder of EnergyPlus v8.8. The code `epw <- file.path(eplusr::eplus_config(8.8)$dir, "WeatherData/USA_CA_San.Francisco.Intl.AP.724940_TMY3.epw")` wouldn't work since you did not have EnergyPlus of that version installed. `eplusr::install_eplus("latest")` will install the latest EnergyPlus version {eplusr} package supported. At that time, the latest supported version was v9.4. So that's why `eplusr::eplus_config(9.4)$dir` worked. As I mentioned in https://github.com/openjournals/joss-reviews/issues/4030#issuecomment-1011701997, EnergyPlus provides free available worldwide weather data at energyplus.net/weather. The one I used in the example can be downloaded at the EnergyPlus GitHub repo. I will update the example to directly use that link.

The EPW file is a mandatory input for [`match_coord()`](https://ideas-lab-nus.github.io/epwshiftr/reference/match_coord.html). Despite an EPW file path, the `epw` argument can also be a regular expression, e.g. `"los angeles.*tmy3", to find and download matched EPW files in the [EnergyPlus weather database](https://energyplus.net/weather) and [OneBuilding.org](https://climate.onebuilding.org/). Using this approach did not require an installation of EnergyPlus at all. I documented this but did not include the usage of it to keep the example short. I may update the example to demonstrate this.

### Grid distance calculation
> Line: coord <- match_coord(epw, threshold = list(lon = 1, lat = 1), max_num = 1) - as far as I understood the code, it assumes that distance on longitude and latitude is the same which is not true (except equator). It has to be corrected e.g. by utilizing euclidean or spheroid formulas (e.g. https://github.com/bczernecki/climate/blob/dev/R/spheroid_dist.R)

Thanks a lot for catching this and thanks for providing the approach from the {climate} package!
I found the blog [Accessing netCDF Data by Coordinates](https://www.unidata.ucar.edu/blogs/developer/en/entry/accessing_netcdf_data_by_coordinates) from Unidata to be a very helpful resource on describing common approaches. I will update the package to correct this.

### Coordinate parsing
Thanks for mentioning this. Actually, I believe epwshiftr has already correctly handled this case, please see https://github.com/ideas-lab-nus/epwshiftr/blob/master/R/coord.R#L205-L208 during coordinate matching in `match_coord()`. Currently coordinate data was only used when finding "closest" grid points and extracting data of them.
