### 2022-02-17 #4030 @hongyuanjia
@KristinaRiemer @mitmat @bczernecki I have opened issues in the epwshiftr repo to be addressed as per your comments. I will address them one by one. Once all are solved, I will request to process the review.

- [ ] https://github.com/ideas-lab-nus/epwshiftr/issues/32
- [ ] https://github.com/ideas-lab-nus/epwshiftr/issues/37
- [ ] https://github.com/ideas-lab-nus/epwshiftr/issues/38
- [ ] Provide a `dir` parameter to let the user specify where to load the file index. https://github.com/ideas-lab-nus/epwshiftr/issues/19
- [ ] Update README to directly use an EPW file from EnergyPlus GitHub repo. https://github.com/ideas-lab-nus/epwshiftr/issues/28
- [ ] Add the statement of need in README. https://github.com/ideas-lab-nus/epwshiftr/issues/28
- [ ] Update the description of output for  `extract_data()` and `morphing_epw()`. https://github.com/ideas-lab-nus/epwshiftr/issues/29
- [ ] Update contribution information in README to provide guidelines about how to contribute and report issues. https://github.com/ideas-lab-nus/epwshiftr/issues/28
- [ ] https://github.com/ideas-lab-nus/epwshiftr/issues/39
- [ ] https://github.com/ideas-lab-nus/epwshiftr/issues/30
- [ ] https://github.com/ideas-lab-nus/epwshiftr/issues/31
- [ ] https://github.com/ideas-lab-nus/epwshiftr/issues/40
- [ ] https://github.com/ideas-lab-nus/epwshiftr/issues/41
- [ ] https://github.com/ideas-lab-nus/epwshiftr/issues/42
- [ ] https://github.com/ideas-lab-nus/epwshiftr/issues/43
