### 2022-03-19 bb7398f
[feat] Detect missing files in `summary_database()`
