### 2022-03-19 #4030 @hongyuanjia
@KristinaRiemer Sorry for the late. Unfortunately, I am slowly moving forward. Right now I mainly focus on addressing https://github.com/ideas-lab-nus/epwshiftr/issues/40. Will address reviewer comments related to features (https://github.com/ideas-lab-nus/epwshiftr/issues/32,  ideas-lab-nus/epwshiftr#19, https://github.com/ideas-lab-nus/epwshiftr/issues/43) and bugs (https://github.com/ideas-lab-nus/epwshiftr/issues/39) during this weekend. I will try my best to solve documentation-related issues next week.
