+++ man/summary_database.Rd
@@ -5,7 +5,7 @@
 \title{Summary CMIP6 model output file status}
 \usage{
 summary_database(
-  dir = get_data_dir(),
+  dir,
   by = c("activity", "experiment", "variant", "frequency", "variable", "source",
     "resolution"),
   mult = c("skip", "latest"),
@@ -44,9 +44,9 @@ stored in different directories. Default: \code{FALSE}.}
 \item{miss}{Actions when matched files in the previous summary do not exist
 when running current summary. Only applicable when \code{append} is set to
 \code{TRUE}. If \code{"keep"}, the metadata for the missing output files will
-be kept. If \code{"overwrite"} and the missing output files are found in
-current directory, existing metadata will be overwriten based on the
-newly found files. Default: \code{"overwrite"}.}
+be kept. If \code{"overwrite"}, existing metadata of those output will be
+first removed from the output file index and overwriten based on the
+newly matched files if possible. Default: \code{"keep"}.}
 
 \item{recursive}{If \code{TRUE}, scan recursively into directories. Default:
 \code{FALSE}.}
@@ -76,10 +76,11 @@ columns plus:\tabular{lll}{
 
 Also 2 extra \code{\link[data.table:data.table]{data.table::data.table()}} are attached as \strong{attributes}:
 \itemize{
-\item \code{not_found}: A \code{\link[data.table:data.table]{data.table::data.table()}} that contains meta data for those
-CMIP6 output files that are listed in current CMIP6 output file index but
-not cannot be found in current database.
-\item \code{not_matched}: A \code{\link[data.table:data.table]{data.table::data.table()}} that contains meta data for
+\item \code{not_found}: A \code{\link[data.table:data.table]{data.table::data.table()}} that contains metadata for those
+CMIP6 outputs that are listed in current CMIP6 output file index but the
+existing file paths are not valid now and cannot be found in current
+database.
+\item \code{not_matched}: A \code{\link[data.table:data.table]{data.table::data.table()}} that contains metadata for
 those CMIP6 output files that are found in current database but not listed
 in current CMIP6 output file index.
 }
@@ -102,6 +103,11 @@ files in the same place.
 \code{datetime_end} global attributes of each NetCDF file to match against the
 output file index. So the names of NetCDF files do not necessarily follow the
 CMIP6 file name encoding.
+
+\code{summary_database()} will append 5 columns in the CMIP6 ouput file index:
+\itemize{
+\item \code{file_path}: the full path of matched NetCDF file for every case.
+}
 
 \code{summary_database()} uses \link[future.apply:future_lapply]{future.apply}
 underneath to speed up the data processing if applicable. You can use your
