### 2022-03-19 efc0fef
[fix] Use spheroid formula for grid distance calculation
