+++ tests/testthat/test-coord.R
@@ -41,50 +41,25 @@ test_that("match_nc_coord()", {
         expect_error(match_nc_coord(path, loc$latitude, loc$longitude, threshold = list(lat = 1.0, lon = 0.01)))
 
         # can specify maximum matched number
-        expect_is(matched <- match_nc_coord(path, loc$latitude, loc$longitude, max_num = 1L), "list")
-        expect_equal(names(matched), c("lat", "lon"))
-        expect_equal(names(matched$lat), c("index", "value", "dis", "which"))
-        expect_equal(length(matched$lat$index), 1L)
-        expect_equal(length(matched$lat$value), 1L)
-        expect_equal(length(matched$lat$dis), 1L)
-        expect_equal(length(matched$lat$which), 1L)
-        expect_equal(names(matched$lon), c("index", "value", "dis", "which"))
-        expect_equal(length(matched$lon$index), 1L)
-        expect_equal(length(matched$lon$value), 1L)
-        expect_equal(length(matched$lon$dis), 1L)
-        expect_equal(length(matched$lon$which), 1L)
+        expect_is(matched <- match_nc_coord(path, loc$latitude, loc$longitude, max_num = 1L), "data.table")
+        expect_equal(names(matched), c("index", "ind_lon", "ind_lat", "lon", "lat", "dist"))
+        expect_equal(nrow(matched), 1L)
 
         # can match multiple values
-        expect_is(matched <- match_nc_coord(path, loc$latitude, loc$longitude), "list")
-        expect_equal(names(matched), c("lat", "lon"))
-        expect_equal(names(matched$lat), c("index", "value", "dis", "which"))
-        expect_equal(length(matched$lat$index), 2L)
-        expect_equal(length(matched$lat$value), 2L)
-        expect_equal(length(matched$lat$dis), 2L)
-        expect_equal(length(matched$lat$which), 2L)
-        expect_equal(names(matched$lon), c("index", "value", "dis", "which"))
-        expect_equal(length(matched$lon$index), 3L)
-        expect_equal(length(matched$lon$value), 3L)
-        expect_equal(length(matched$lon$dis), 3L)
-        expect_equal(length(matched$lon$which), 3L)
+        expect_is(matched <- match_nc_coord(path, loc$latitude, loc$longitude, threshold = list(lat = 1.0, lon = 1.0)), "data.table")
+        expect_equal(names(matched), c("index", "ind_lon", "ind_lat", "lon", "lat", "dist"))
+        expect_equal(nrow(matched), 6L)
 
         # can change EPW longitude to [0, 360] range
         loc <- eplusr:::WEATHER_DB[title == "USA_NY_New.York.City-Central.Park.744860_TMY2"]
-        expect_is(matched <- match_nc_coord(path, loc$latitude, loc$longitude), "list")
-        expect_equal(names(matched), c("lat", "lon"))
-        expect_equal(names(matched$lat), c("index", "value", "dis", "which"))
-        expect_equal(length(matched$lat$index), 3L)
-        expect_equal(length(matched$lat$value), 3L)
-        expect_equal(length(matched$lat$dis), 3L)
-        expect_equal(length(matched$lat$which), 3L)
-        expect_equal(names(matched$lon), c("index", "value", "dis", "which"))
-        expect_equal(length(matched$lon$index), 3L)
-        expect_equal(length(matched$lon$value), 3L)
-        expect_equal(length(matched$lon$dis), 3L)
-        expect_equal(length(matched$lon$which), 3L)
+        expect_is(matched <- match_nc_coord(path, loc$latitude, loc$longitude, max_num = 1), "data.table")
[…]
