+++ DESCRIPTION
@@ -47,11 +47,11 @@ BugReports: https://github.com/ideas-lab-nus/epwshiftr/issues
 Roxygen: list(markdown = TRUE)
 RoxygenNote: 7.1.2
 Collate:
-    'epwshiftr-package.R'
-    'utils.R'
-    'esgf.R'
-    'netcdf.R'
     'coord.R'
+    'utils.R'
+    'epwshiftr-package.R'
+    'esgf.R'
     'morph.R'
+    'netcdf.R'
 VignetteBuilder: knitr
 Config/testthat/start-first: utils, esgf, coord, netcdf, morph
