### 2022-03-19 e96c7eb
[fix] Only remove `index` column when warning mes has been issued
