### 2022-03-19 84ef756
[test] Fix `match_coord()` tests
