### 2022-03-19 2d21103
[refactor] Refactor `get_nc_meta()` to use new coord structure
