### 2022-03-19 724c5ae
[test] Update tests for `extract_data()`
