### 2022-03-19 2ab28e8
[doc] Correct `extract_data()`
