### 2022-03-19 12741ff
[doc] Update `extract_data()` doc
