### 2022-03-19 2f2ba56
[doc] Update NEWS
