### 2022-03-19 bf3911d
[doc] Fix typos in NEWS
