### 2022-03-20 393a661
[CRAN] Solve undefined global variables
