### 2022-03-20 0f31475
[fix] Use distances when calculating location mean
