+++ tests/testthat/test-morph.R
@@ -96,7 +96,7 @@ test_that("morphing_from_mean()", {
         )
         expect_equal(names(res),
             c("activity_drs", "experiment_id", "institution_id", "source_id",
-              "member_id", "table_id", "lon", "lat", "interval", "datetime",
+              "member_id", "table_id", "lon", "lat", "dist", "interval", "datetime",
               "year", "month", "day", "hour", "minute", "dry_bulb_temperature",
               "delta", "alpha"
             )
@@ -117,7 +117,7 @@ test_that("morphing_from_mean()", {
         )
         expect_equal(names(res),
             c("activity_drs", "experiment_id", "institution_id", "source_id",
-              "member_id", "table_id", "lon", "lat", "interval", "datetime",
+              "member_id", "table_id", "lon", "lat", "dist", "interval", "datetime",
               "year", "month", "day", "hour", "minute", "dry_bulb_temperature",
               "delta", "alpha"
             )
@@ -137,7 +137,7 @@ test_that("morphing_from_mean()", {
         )
         expect_equal(names(res),
             c("activity_drs", "experiment_id", "institution_id", "source_id",
-              "member_id", "table_id", "lon", "lat", "interval", "datetime",
+              "member_id", "table_id", "lon", "lat", "dist", "interval", "datetime",
               "year", "month", "day", "hour", "minute", "dry_bulb_temperature",
               "delta", "alpha"
             )
@@ -157,7 +157,7 @@ test_that("morphing_from_mean()", {
         )
         expect_equal(names(res),
             c("activity_drs", "experiment_id", "institution_id", "source_id",
-              "member_id", "table_id", "lon", "lat", "interval", "datetime",
+              "member_id", "table_id", "lon", "lat", "dist", "interval", "datetime",
               "year", "month", "day", "hour", "minute", "dry_bulb_temperature",
               "delta", "alpha"
             )
@@ -177,7 +177,7 @@ test_that("morphing_from_mean()", {
         )
         expect_equal(names(res),
             c("activity_drs", "experiment_id", "institution_id", "source_id",
-              "member_id", "table_id", "lon", "lat", "interval", "datetime",
+              "member_id", "table_id", "lon", "lat", "dist", "interval", "datetime",
               "year", "month", "day", "hour", "minute", "dry_bulb_temperature",
               "delta", "alpha"
             )
