### 2022-03-20 34cd7b2
[fix] Make sure `extract_data()` and `morphing_epw()` return same metadata
