+++ R/morph.R
@@ -137,22 +137,31 @@ remove_units <- function (data, var) {
 #' | 11   | `total_cover`  | [data.table::data.table()] | Derived         | Data of total sky cover after morphing                            |
 #' | 12   | `opaque_cover` | [data.table::data.table()] | Derived         | Data of opaque sky cover after morphing                           |
 #'
-#' Each [data.table::data.table()] listed above contains x columns
+#' Each [data.table::data.table()] listed above contains 19 columns below or an
+#' empty [data.table::data.table()] if the corresponding variables cannot be
+#' found in the input `epw_cmip6_data` object.
 #'
-#' | No.  | Column            | Type      | Description                                                                            |
-#' | ---: | -----             | -----     | -----                                                                                  |
-#' | 1    | `activity_drs`    | Character | Activity DRS (Data Reference Syntax)                                                   |
-#' | 2    | `institution_id`  | Character | Institution identifier                                                                 |
-#' | 3    | `source_id`       | Character | Model identifier                                                                       |
-#' | 4    | `experiment_id`   | Character | Root experiment identifier                                                             |
-#' | 5    | `member_id`       | Character | A compound construction from `sub_experiment_id` and `variant_label`                   |
-#' | 6    | `table_id`        | Character | Table identifier                                                                       |
-#' | 7    | `lat`             | Double    | The **averaged** values of input latitude                                              |
-#' | 8    | `lon`             | Double    | The **averaged** values of input longitude                                             |
-#' | 9    | `interval`        | Factor    | The year value of data morphed                                                         |
-#' | 10   | **Variable Name** | Double    | The morphed data, where `Variable Name` is the corresponding EPW weather variable name |
-#' | 11   | `delta`           | Double    | The shift factor. Will be `NA` for derivied values                                     |
-#' | 12   | `alpha`           | Double    | The stretch factor. Will be `NA` for derivied values                                   |
+#' | No.  | Column            | Type      | Description                                                                                           |
+#' | ---: | -----             | -----     | -----                                                                                                 |
+#' | 1    | `activity_drs`    | Character | Activity DRS (Data Reference Syntax)                                                                  |
+#' | 2    | `institution_id`  | Character | Institution identifier                                                                                |
+#' | 3    | `source_id`       | Character | Model identifier                                                                                      |
+#' | 4    | `experiment_id`   | Character | Root experiment identifier                                                                            |
+#' | 5    | `member_id`       | Character | A compound construction from `sub_experiment_id` and `variant_label`                                  |
+#' | 6    | `table_id`        | Character | Table identifier                                                                                      |
+#' | 7    | `lon`             | Double    | The **averaged** values of input longitude                                                            |
+#' | 8    | `lat`             | Double    | The **averaged** values of input latitude                                                             |
+#' | 9    | `dist`            | Double    | The **averaged** spherical distances in km between EPW location and grid coordinates                  |
+#' | 10   | `interval`        | Factor    | The label value used to average raw input data                                                        |
+#' | 11   | `datetime`        | POSIXct   | The datetime value with **fake year** generated by calling the `Epw$data()` method with the input EPW |
+#' | 12   | `year`            | Integer   | The **original** year of the raw EPW data                                                             |
+#' | 13   | `month`           | Integer   | The month value of the morphed data                                                                   |
+#' | 14   | `day`             | Integer   | The day of the morphed data                                                                           |
+#' | 15   | `hour`            | Integer   | The hour of the morphed data                                                                          |
+#' | 16   | `minute`          | Integer   | The minute of the morphed data                                                                        |
[…]
