+++ man/morphing_epw.Rd
@@ -42,7 +42,9 @@ An \code{epw_cmip6_morphed} object, which is basically a list of 12 elements:\ta
 }
 
 
-Each \code{\link[data.table:data.table]{data.table::data.table()}} listed above contains x columns\tabular{rlll}{
+Each \code{\link[data.table:data.table]{data.table::data.table()}} listed above contains 19 columns below or an
+empty \code{\link[data.table:data.table]{data.table::data.table()}} if the corresponding variables cannot be
+found in the input \code{epw_cmip6_data} object.\tabular{rlll}{
    No. \tab Column \tab Type \tab Description \cr
    1 \tab \code{activity_drs} \tab Character \tab Activity DRS (Data Reference Syntax) \cr
    2 \tab \code{institution_id} \tab Character \tab Institution identifier \cr
@@ -50,12 +52,19 @@ Each \code{\link[data.table:data.table]{data.table::data.table()}} listed above 
    4 \tab \code{experiment_id} \tab Character \tab Root experiment identifier \cr
    5 \tab \code{member_id} \tab Character \tab A compound construction from \code{sub_experiment_id} and \code{variant_label} \cr
    6 \tab \code{table_id} \tab Character \tab Table identifier \cr
-   7 \tab \code{lat} \tab Double \tab The \strong{averaged} values of input latitude \cr
-   8 \tab \code{lon} \tab Double \tab The \strong{averaged} values of input longitude \cr
-   9 \tab \code{interval} \tab Factor \tab The year value of data morphed \cr
-   10 \tab \strong{Variable Name} \tab Double \tab The morphed data, where \verb{Variable Name} is the corresponding EPW weather variable name \cr
-   11 \tab \code{delta} \tab Double \tab The shift factor. Will be \code{NA} for derivied values \cr
-   12 \tab \code{alpha} \tab Double \tab The stretch factor. Will be \code{NA} for derivied values \cr
+   7 \tab \code{lon} \tab Double \tab The \strong{averaged} values of input longitude \cr
+   8 \tab \code{lat} \tab Double \tab The \strong{averaged} values of input latitude \cr
+   9 \tab \code{dist} \tab Double \tab The \strong{averaged} spherical distances in km between EPW location and grid coordinates \cr
+   10 \tab \code{interval} \tab Factor \tab The label value used to average raw input data \cr
+   11 \tab \code{datetime} \tab POSIXct \tab The datetime value with \strong{fake year} generated by calling the \code{Epw$data()} method with the input EPW \cr
+   12 \tab \code{year} \tab Integer \tab The \strong{original} year of the raw EPW data \cr
+   13 \tab \code{month} \tab Integer \tab The month value of the morphed data \cr
+   14 \tab \code{day} \tab Integer \tab The day of the morphed data \cr
+   15 \tab \code{hour} \tab Integer \tab The hour of the morphed data \cr
+   16 \tab \code{minute} \tab Integer \tab The minute of the morphed data \cr
+   17 \tab \strong{Variable Name} \tab Double \tab The morphed data, where \verb{Variable Name} is the corresponding EPW weather variable name \cr
+   18 \tab \code{delta} \tab Double \tab The shift factor. Will be \code{NA} for derivied values \cr
+   19 \tab \code{alpha} \tab Double \tab The stretch factor. Will be \code{NA} for derivied values \cr
 }
 }
 \description{
