### 2022-03-20 24a8600
[doc] Update NEWS
