### 2022-03-20 8ca4335
[fix] Use spheroid formula for grid distance calculation (#45)
