+++ R/netcdf.R
@@ -313,7 +313,7 @@ summary_database <- function (
 #' @importFrom data.table as.data.table set setattr setcolorder
 #' @importFrom RNetCDF utcal.nc
 #' @importFrom units set_units
-get_nc_data <- function (x, lats, lons, years, unit = TRUE) {
+get_nc_data <- function (x, coord, years, unit = TRUE) {
     assert_flag(unit)
     assert_integerish(years, lower = 1900, unique = TRUE, sorted = TRUE, any.missing = FALSE, null.ok = TRUE)
 
@@ -338,15 +338,12 @@ get_nc_data <- function (x, lats, lons, years, unit = TRUE) {
 
     # match time
     time <- match_nc_time(nc, years)
-
-    which_lat <- lats$which
-    which_lon <- lons$which
     which_time <- time$which
 
     # get dimensions
     dims <- get_nc_dims(nc)[J(c("lon", "lat", "time")), on = "name"][order(id)][, id := .I]
-    dims[J("lon"), on = "name", `:=`(value = list(which_lon))]
-    dims[J("lat"), on = "name", `:=`(value = list(which_lat))]
+    dims[J("lon"), on = "name", `:=`(value = list(coord$ind_lon))]
+    dims[J("lat"), on = "name", `:=`(value = list(coord$ind_lat))]
 
     # find correct dimentions
     ord <- list(
@@ -376,9 +373,9 @@ get_nc_data <- function (x, lats, lons, years, unit = TRUE) {
     dt <- rbindlist(use.names = TRUE, lapply(seq_along(which_time), function (i) {
         if (!length(which_time[[i]])) {
             return(data.table(
-                value = double(),
-                lon = double(), lat = double(),
-                datetime = Sys.time()[0]
+                index = integer(),
+                lon = double(), lat = double(), dist = double(),
+                datetime = Sys.time()[0], value = double()
             ))
         }
 
@@ -390,16 +387,27 @@ get_nc_data <- function (x, lats, lons, years, unit = TRUE) {
         # longitude, latitude, time).
         dt <- as.data.table(RNetCDF::var.get.nc(nc, var,
             c(min(dims$value[[1L]]), min(dims$value[[2L]]), min(dims$value[[3L]])),
-            c(length(dims$value[[1L]]), length(dims$value[[2L]]), length(dims$value[[3L]])),
+            c(length(unique(dims$value[[1L]])), length(unique(dims$value[[2L]])), length(dims$value[[3L]])),
             collapse = FALSE)
         )
 
-        setnames(dt, c(dims$name, "value"))
-        setnames(dt, "time", "datetime")
-
-        set(dt, NULL, "lon", lons$value[dt$lon])
-        set(dt, NULL, "lat", lats$value[dt$lat])
+        setnames(dt, c(paste0("ord_", dims$name), "value"))
+        setnames(dt, "ord_time", "datetime")
         set(dt, NULL, "datetime", time$datetime[[i]][dt$datetime])
+
+        # only keep the matched points
+        ind_lon <- ind_lat <- ord_lon <- ord_lat <- .GRP <- NULL # to make CRAN check happy
+        # NOTE: make a copy of input coord here, otherwise there will be a shallow
+        # copy warning
+        # should be cheap to do so, since the coord is normally a small DT
+        co <- copy(coord)
+        co[order(ind_lon), ord_lon := .GRP, by = "ind_lon"]
+        co[order(ind_lat), ord_lat := .GRP, by = "ind_lat"]
+        res <- dt[co, on = c("ord_lon", "ord_lat")]
+        set(res, NULL, c("ord_lon", "ord_lat", "ind_lon", "ind_lat"), NULL)
[…]
