### 2022-03-20 21b65e5
[fix] Use spheroid formula for grid distance calculation (#45)
