+++ R/coord.R
@@ -66,13 +66,34 @@ match_location_coord <- function (path, dict, threshold = list(lon = 1.0, lat = 
 # match_coord {{{
 #' Match coordinates of input EPW in the CMIP6 output file database
 #'
-#' `match_coord()` takes an EPW and uses its longitude and latitude to match
-#' corresponding values that meet specified threshold in NetCDF files.
+#' `match_coord()` takes an EPW and uses its longitude and latitude to calculate
+#' the distance between the EPW location and the global grid points in NetCDF
+#' files.
 #'
 #' `match_coord()` uses [future.apply][future.apply::future_lapply()]
 #' underneath. You can use your preferable future backend to
 #' speed up data extraction in parallel. By default, `match_coord()` uses
 #' `future::sequential` backend, which runs things in sequential.
+#'
+#' @section Geographical distance calculation:
+#'
+#' `match_coord()` calculates the geographical distances based formulas of
+#' spherical trigonometry:
+#'
+#' \deqn{
+#' \begin{align}
+#' &\Delta{X}=\cos(\phi_2)\cos(\lambda_2) - \cos(\phi_1)\cos(\lambda_1);\\
+#' &\Delta{Y}=\cos(\phi_2)\sin(\lambda_2) - \cos(\phi_1)\sin(\lambda_1);\\
+#' &\Delta{Z}=\sin(\phi_2) - \sin(\phi_1);\\
+#' &C_h=\sqrt{(\Delta{X})^2 + (\Delta{Y})^2 + (\Delta{Z})^2}.
+#' \end{align}
+#' }
+#'
+#' where \eqn{phi} is the latitude and \eqn{lambda} is the longitude.  This
+#' formula treats the Earth as a sphere. The geographical distance between
+#' points on the surface of a spherical Earth is \eqn{D = RC_h}.
+#'
+#' For more details, please see this [Wikipedia](https://en.wikipedia.org/wiki/Geographical_distance#Tunnel_distance)
 #'
 #' @param epw Possible values:
 #'
@@ -84,12 +105,16 @@ match_location_coord <- function (path, dict, threshold = list(lon = 1.0, lat = 
 #'   search is case-insensitive
 #'
 #' @param threshold A list of 2 elements `lon` and `lat` specifying the
-#'        absolute distance threshold used when matching longitude and latitude.
-#'        Default: `list(lon = 1.0, lat = 1.0)`
+#'        absolute distance threshold used for subsetting longitude and
+#'        latitude value for calculating distances. If `NULL`, no subsetting is
+#'        performed and the distances between the target location and all grid
+#'        points are calculated. This is useful set the `threshold` value to
+#'        exclude some points that are absolute too far away from the target
+#'        location. Default: `list(lon = 1.0, lat = 1.0)`
 #'
-#' @param max_num The maximum number to be matched for both longitude and
-#'        latitude when `threshold` is matched. Default is `NULL`, which means
-#'        no limit
+#' @param max_num The maximum number of grid points to be matched. Default is
+#'        `NULL`, which means no number limit and the total matched grid points
+#'        are determined by the `threshold` input.
 #'
 #' @return An `epw_cmip6_coord` object, which is basically a list of 3 elements:
 #'
@@ -99,16 +124,17 @@ match_location_coord <- function (path, dict, threshold = list(lon = 1.0, lat = 
 #' * `coord`: A [data.table::data.table()] which is basically CMIP6 index
[…]
