### 2022-03-20 cb7b703
[fix] Make sure `index` column is removed in `not_found` attr
