### 2022-03-20 daa19c7
[feat] Detect missing files in `summary_database()` (#44)
