+++ R/netcdf.R
@@ -6,7 +6,7 @@ get_nc_meta <- function (file) {
     # get all attributes
     atts <- get_nc_atts(file)
 
-    # get meta data
+    # get metadata
     meta <- as.list(atts[
         J("NC_GLOBAL",
           c("mip_era", "activity_id", "institution_id", "source_id",
@@ -39,15 +39,30 @@ get_nc_meta <- function (file) {
 # summary_database {{{
 #' Summary CMIP6 model output file status
 #'
-#' `summary_database()` scan the directory specified and returns a
+#' `summary_database()` scans the directory specified and returns a
 #' [data.table()] containing summary information about all the CMIP6
 #' files available against the output file index loaded using
 #' [load_cmip6_index()].
 #'
+#' The database here can be any directory that stores the NetCDF files for CMIP6
+#' GCMs. It can be also be the same as `get_data_dir()` where epwshiftr stores
+#' the output file index, if you want to save the output file index and output
+#' files in the same place.
+#'
+#' `summary_database()` uses the `tracking_id`, `datetime_start` and
+#' `datetime_end` global attributes of each NetCDF file to match against the
+#' output file index. So the names of NetCDF files do not necessarily follow the
+#' CMIP6 file name encoding.
+#'
+#' `summary_database()` will append 5 columns in the CMIP6 ouput file index:
+#'
+#' - `file_path`: the full path of matched NetCDF file for every case.
+#'
 #' `summary_database()` uses [future.apply][future.apply::future_lapply()]
-#' underneath. You can use your preferable future backend to
-#' speed up data extraction in parallel. By default, `summary_database()` uses
-#' `future::sequential` backend, which runs things in sequential.
+#' underneath to speed up the data processing if applicable. You can use your
+#' preferable future backend to speed up data extraction in parallel. By default,
+#' `summary_database()` uses `future::sequential` backend, which runs things in
+#' sequential.
 #'
 #' @param dir A single string indcating the directory where CMIP6 model output
 #'        NetCDF files are stored.
@@ -72,6 +87,13 @@ get_nc_meta <- function (file) {
 #'        will be used. If `"skip"`, all matched files will be skip and this
 #'        case will be kept as unmatched. Default: `"skip"`.
 #'
+#' @param miss Actions when matched files in the previous summary do not exist
+#'        when running current summary. Only applicable when `append` is set to
+#'        `TRUE`. If `"keep"`, the metadata for the missing output files will
+#'        be kept. If `"overwrite"`, existing metadata of those output will be
+#'        first removed from the output file index and overwriten based on the
+#'        newly matched files if possible. Default: `"keep"`.
+#'
 #' @param recursive If `TRUE`, scan recursively into directories. Default:
 #'        `FALSE`.
 #'
@@ -81,8 +103,8 @@ get_nc_meta <- function (file) {
 #'        database file saved in [get_data_dir()] will remain unchanged.
 #'        Default: `FALSE`.
 #'
-#' @param warning If `TRUE`, warning messages will show when multiple files
-#'        match a same case. Default: `TRUE`.
+#' @param warning If `TRUE`, warning messages will show when target files are
+#'        missing or multiple files match a same case. Default: `TRUE`.
 #'
 #' @return A [data.table::data.table()] containing corresponding grouping
 #' columns plus:
@@ -97,9 +119,16 @@ get_nc_meta <- function (file) {
 #' | `dl_percent`     | Units (%)      | Total percentage of file downloaded |
 #' | `dl_size`        | Units (Mbytes) | Total size of file downloaded       |
 #'
-#' Also an attribute `not_matched` is added to the returned
-#' [data.table::data.table()] which contains meta data for those CMIP6 output
[…]
