### 2022-03-21 4fa023e
[test] Add tests for edge cases
