+++ NAMESPACE
@@ -11,6 +11,7 @@ export(match_coord)
 export(morphing_epw)
 export(set_cmip6_index)
 export(summary_database)
+importFrom(PCICt,as.PCICt)
 importFrom(RNetCDF,utcal.nc)
 importFrom(checkmate,assert_character)
 importFrom(checkmate,assert_choice)
