### 2022-03-21 e237081
[feat] Warning if input raw data is less than a decade
