+++ NEWS.md
@@ -65,6 +65,10 @@
   `"keep"`, which is the default, to do nothing about it or `"overwrite"` to
   overwrite those entries based on the newly matched NetCDF files if possible
   (#40).
+* A new parameter `warning` has been added in `morphing_epw()`. If set to `TRUE`,
+  warnings will be issued for cases with input data less than a decade (10 years)
+  . This is because using data that only covers a short period of time may not
+  be able to capture the average of future climate (#41).
 
 ## Bug fixes
 
