+++ R/morph.R
@@ -1,7 +1,7 @@
 # preprocess_morphing {{{
 #' @importFrom checkmate assert_count
 #' @importFrom data.table set
-preprocess_morphing <- function (dt, leapyear = FALSE, years = NULL, labels = NULL) {
+preprocess_morphing <- function (dt, leapyear = FALSE, years = NULL, labels = NULL, warning = FALSE) {
     # add datetime components
     data.table::set(dt, NULL, c("year", "month", "day", "hour", "minute"),
         list(data.table::year(dt$datetime),
@@ -37,6 +37,33 @@ preprocess_morphing <- function (dt, leapyear = FALSE, years = NULL, labels = NU
             y <- data.table(year = years, interval = as.factor(labels))
 
             dt[y, on = "year", interval := i.interval]
+        }
+    }
+
+    if (warning) {
+        # give a warning if the dataset only includes years less than a decade
+        rng_year <- dt[, list(years = list(sort(unique(year))), num_years = length(unique(year))),
+            by = c("variable", "table_id", "source_id", "experiment_id", "member_id")
+        ][num_years < 10L]
+
+        if (nrow(rng_year)) {
+            set(rng_year, NULL, "index_case", seq.int(nrow(rng_year)))
+
+            mes <- rng_year[, by = "index_case", {
+                head <- sprintf("#%i | For case %s_%s_%s_%s_%s:\n",
+                    .BY$index_case, variable, table_id, source_id, experiment_id, member_id
+                )
+                yrs <- sprintf("   --> [%i] %s", num_years, paste0("'", years[[1L]], "'", collapse = ", "))
+                list(message = paste0(head, yrs, collapse = "\n"))
+            }]$message
+
+            ori <- getOption("warning.length")
+            options(warning.length = 8170L)
+            on.exit(options(warning.length = ori), add = TRUE)
+
+            warning("Case(s) shown below contains CMIP data less than a decade. ",
+                "The morphed data may not be able to capture average weather of the future climate.\n",
+                paste0(mes, collapse = "\n"), call. = FALSE)
         }
     }
 
@@ -119,6 +146,10 @@ remove_units <- function (data, var) {
 #'        "shift")`. `"combined"` is only applicable to `tdb`. The default
 #'        morphing method for each variable is listed in the *Return* section.
 #'        If `NULL`, the default methods will be used. Default: `NULL`.
+#' @param warning If `TRUE`, warnings will be issued for cases with input data
+#'        less than a decade (10 years). This is because using data that only
+#'        covers a short period of time may not be able to capture the average
+#'        of future climate. Default: `FALSE`.
 #'
 #' @return An `epw_cmip6_morphed` object, which is basically a list of 12 elements:
 #'
@@ -169,7 +200,7 @@ remove_units <- function (data, var) {
 #' 26, 49–61. https://doi.org/10.1191/0143624405bt112oa
 #'
 #' @export
-morphing_epw <- function (data, years = NULL, labels = NULL, methods = NULL) {
+morphing_epw <- function (data, years = NULL, labels = NULL, methods = NULL, warning = FALSE) {
     assert_class(data, "epw_cmip6_data")
     if (is.null(methods)) {
         methods <- c(tdb = "stretch", rh = "stretch", p = "stretch",
@@ -199,7 +230,7 @@ morphing_epw <- function (data, years = NULL, labels = NULL, methods = NULL) {
         tasmin <- data_cmip[J("tasmin"), on = "variable", nomatch = NULL]
         if (!nrow(tasmax)) tasmax <- NULL
[…]
