### 2022-03-21 5d64581
[test] Add a test for #41
