### 2022-03-21 35458ca
[test] Add a test for #41 (#49)
