### 2022-03-21 45466eb
[test] Add a test for #41 (#49)
