+++ R/esgf.R
@@ -14,6 +14,10 @@
 #' Laboratory)](http://esgf-node.llnl.gov)
 #' Index Node.
 #'
+#' The core Controlled Vocabularies (CVs) for use in CMIP6, including all
+#' activities, experiment, sources (GCMs), frequencies can be found at the
+#' [WCRP-CMIP/CMIP6_CVs](https://github.com/WCRP-CMIP/CMIP6_CVs) GitHub repo.
+#'
 #' @param activity A character vector indicating activity identifiers. Default:
 #'        `"ScenarioMIP"`. Possible values:
 #'
@@ -300,6 +304,176 @@ esgf_query <- function (
     dt
 }
 # }}}
+
+build_cmip6_db <- function() {
+    all_types <- c("activity", "experiment", "sub_experiment", "institution",
+        "source", "table", "frequency", "grid_label", "realm", "source_type",
+        "req_global_atts", "resolution")
+
+    cvs <- lapply(all_types, fetch_cmip6_cv)
+}
+
+CMIP6DB <- R6::R6Class("CMIP6DB",
+    public = list(
+
+    ),
+
+    private = list(
+        url_base = "https://esgf-node.llnl.gov/search/cmip6/"
+    )
+)
+
+# download.file(url_base, "esgf_llnl.html", mode = "wb", method = "libcurl")
+# l <- readLines(url_base)
+
+fetch_cmip6_cv <- function(type, save = FALSE) {
+    assert_string(type)
+
+    if (type %in% c("activity", "experiment", "sub_experiment", "institution", "source", "table")) {
+        file_name <- sprintf("CMIP6_%s_id.json", type)
+    } else if (type %in% c("frequency", "grid_label", "realm", "source_type")) {
+        file_name <- sprintf("CMIP6_%s.json", type)
+    } else if (type == "req_global_atts") {
+        file_name <- "CMIP6_requried_global_attributes.json"
+    } else if (type == "resolution") {
+        file_name <- "CMIP6_nominal_resolution.json"
+    } else {
+        all_types <- c(
+            "activity", "experiment", "sub_experiment", "institution", "source", "table",
+            "frequency", "grid_label", "realm", "source_type",
+            "req_global_atts", "resolution"
+        )
+        stop(sprintf("Invalid CMIP6 CV type found: '%s'. Should be one of [%s]",
+            type, paste0("'", all_types, "'", collapse = ", ")
+        ))
+    }
+
+    URL_CMIP6_CV <- "https://raw.githubusercontent.com/WCRP-CMIP/CMIP6_CVs/master"
+    url_json <- file.path(URL_CMIP6_CV, file_name, fsep = "/")
+    q <- tryCatch(jsonlite::read_json(url_json, TRUE), warning = function (w) w, error = function (e) e)
+
+    # nocov start
+    if (inherits(q, "warning") || inherits(q, "error")) {
+        message(sprintf("Failed to fetch CV json file '%s'. Please check network connection.", url_json))
+        return(data.table())
+    }
+
+    switch(type,
+        activity = format_cmip6_cv_flat(q),
+        experiment = format_cmip6_cv_experiment(q),
+        frequency = format_cmip6_cv_flat(q),
+        grid_label = format_cmip6_cv_flat(q),
+        institution = format_cmip6_cv_flat(q),
+        resolution = format_cmip6_cv_simple(q),
+        realm = format_cmip6_cv_flat(q),
+        req_global_atts = format_cmip6_cv_simple(q),
+        source = format_cmip6_cv_source(q),
[…]
