### 2022-03-22 33184ca
[refactor] Remove redundant code
