+++ R/esgf.R
@@ -305,176 +305,6 @@ esgf_query <- function (
 }
 # }}}
 
-build_cmip6_db <- function() {
-    all_types <- c("activity", "experiment", "sub_experiment", "institution",
-        "source", "table", "frequency", "grid_label", "realm", "source_type",
-        "req_global_atts", "resolution")
-
-    cvs <- lapply(all_types, fetch_cmip6_cv)
-}
-
-CMIP6DB <- R6::R6Class("CMIP6DB",
-    public = list(
-
-    ),
-
-    private = list(
-        url_base = "https://esgf-node.llnl.gov/search/cmip6/"
-    )
-)
-
-# download.file(url_base, "esgf_llnl.html", mode = "wb", method = "libcurl")
-# l <- readLines(url_base)
-
-fetch_cmip6_cv <- function(type, save = FALSE) {
-    assert_string(type)
-
-    if (type %in% c("activity", "experiment", "sub_experiment", "institution", "source", "table")) {
-        file_name <- sprintf("CMIP6_%s_id.json", type)
-    } else if (type %in% c("frequency", "grid_label", "realm", "source_type")) {
-        file_name <- sprintf("CMIP6_%s.json", type)
-    } else if (type == "req_global_atts") {
-        file_name <- "CMIP6_requried_global_attributes.json"
-    } else if (type == "resolution") {
-        file_name <- "CMIP6_nominal_resolution.json"
-    } else {
-        all_types <- c(
-            "activity", "experiment", "sub_experiment", "institution", "source", "table",
-            "frequency", "grid_label", "realm", "source_type",
-            "req_global_atts", "resolution"
-        )
-        stop(sprintf("Invalid CMIP6 CV type found: '%s'. Should be one of [%s]",
-            type, paste0("'", all_types, "'", collapse = ", ")
-        ))
-    }
-
-    URL_CMIP6_CV <- "https://raw.githubusercontent.com/WCRP-CMIP/CMIP6_CVs/master"
-    url_json <- file.path(URL_CMIP6_CV, file_name, fsep = "/")
-    q <- tryCatch(jsonlite::read_json(url_json, TRUE), warning = function (w) w, error = function (e) e)
-
-    # nocov start
-    if (inherits(q, "warning") || inherits(q, "error")) {
-        message(sprintf("Failed to fetch CV json file '%s'. Please check network connection.", url_json))
-        return(data.table())
-    }
-
-    switch(type,
-        activity = format_cmip6_cv_flat(q),
-        experiment = format_cmip6_cv_experiment(q),
-        frequency = format_cmip6_cv_flat(q),
-        grid_label = format_cmip6_cv_flat(q),
-        institution = format_cmip6_cv_flat(q),
-        resolution = format_cmip6_cv_simple(q),
-        realm = format_cmip6_cv_flat(q),
-        req_global_atts = format_cmip6_cv_simple(q),
-        source = format_cmip6_cv_source(q),
-        source_type = format_cmip6_cv_flat(q),
-        sub_experiment = format_cmip6_cv_flat(q),
-        table = format_cmip6_cv_simple(q)
-    )
-
-}
-
-format_cmip6_cv_simple <- function(json) {
-    c(q$version_metadata, q[1L])
-}
-
-format_cmip6_cv_flat <- function(json) {
-    d <- data.table(names(json[[1L]]), unlist(json[[1L]]))
-    setnames(d, c(names(json)[1L], "description"))
-
-    c(json$version_metadata, list(value = as.data.table(d)))
-}
-
[…]
