### 2022-03-22 9b5754d
[doc] Document where to find CVs (#50)
