### 2022-03-22 f8b0b87
[doc] Document where to find CVs (#50)
