+++ pkgdown.yml
@@ -3,5 +3,5 @@ pkgdown: 2.0.2
 pkgdown_sha: ~
 articles:
   paper: paper.html
-last_built: 2022-03-21T14:24Z
+last_built: 2022-03-22T07:17Z
 
