+++ README.Rmd
@@ -247,6 +247,14 @@ extent permitted by law.
 
 ## Contribute
 
+If you encounter a clear bug or have questions about the usage, please file an
+issue with a minimal reproducible example on
+[GitHub](https://github.com/ideas-lab-nus/epwshiftr/issues?q=is%3Aissue+is%3Aopen+sort% 3Aupdated-desc)
+If you have a solution for an existing bug or an implementation for a missing
+feature, please send a pull request and let us review.
+
+------------------------------------------------------------------------
+
 Please note that the 'epwshiftr' project is released with a
 [Contributor Code of Conduct](https://github.com/ideas-lab-nus/epwshiftr/blob/master/.github/CODE_OF_CONDUCT.md).
 By contributing to this project, you agree to abide by its terms.
