### 2022-03-23 23b1131
[doc] Fix equation errrors
