### 2022-03-25 6e62caa
[feat] Add CMIP6Dict class
