+++ R/dict.R
@@ -1,0 +1,849 @@
+#' CMIP6 Controlled Vocabularies (CVs) and Data Request Dictionary
+#'
+#' The `CMIP6Dict` object provides functionalities to fetch the latested CMIP6
+#' Controlled Vocabularies (CVs) and Data Request information.
+#'
+#' The CMIP6 CVs defines a well-defined set of global attributes that are
+#' recorded in each CMIP6 model output, providing information necessary for
+#' interpreting the data.
+#'
+#' The CMIP6 Data Request defines all the quantities from CMIP6 simulations that
+#' should be archived. This includes both quantities of general interest needed
+#' from most of the CMIP6-endorsed model intercomparison projects (MIPs) and
+#' quantities that are more specialized and only of interest to a single
+#' endorsed MIP.
+#'
+#' The data of CVs for use in CMIP6 is stored as JSON files in a
+#' [GitHub Repo](https://github.com/WCRP-CMIP/CMIP6_CVs), while the Data Request
+#' data is stored a Microsoft Excel file (`CMIP6_MIP_tables.xlsx`) in a
+#' [Subversion repo](http://proj.badc.rl.ac.uk/svn/exarch/CMIP6dreq/trunk).
+#' `CMIP6Dict` object is able to fetch those files and parse them into
+#' [data.table][data.table::data.table]s.
+#'
+#' For more information, please see:
+#'
+#' - [CMIP6 Global Attributes, DRS, Filenames, Directory Structure, and CV's](https://docs.google.com/document/d/1h0r8RZr_f3-8egBMMh7aqLwy3snpD6_MrDz1q8n5XUk/edit)
+#' - [CMIP6 Data Request](https://wcrp-cmip.github.io/WGCM_Infrastructure_Panel/CMIP6/data_request.html)
+#'
+#' @examples
+#' # create a new CMIP6Dict object
+#' dict <- cmip6_dict()
+#'
+#' @author Hongyuan Jia
+#'
+#' @importFrom R6 R6Class
+#' @name CMIP6Dict
+#' @export
+cmip6_dict <- function() CMIP6Dict$new()
+
+#' @name CMIP6Dict
+#' @export
+CMIP6Dict <- R6::R6Class("CMIP6Dict",
+    public = list(
+        #' @description
+        #' Get the version of CVs and Data Request
+        #'
+        #' @return A list of two element:
+        #'
+        #' - `cv`: A [numeric_version] object giving the version of CVs
+        #' - `req`: A [numeric_version] object giving the version of Data
+        #'   Request
+        version = function() {
+            private$m_version
+        },
+
+        #' @description
+        #' Get the last modified time for CVs
+        #'
+        #' @return A list of 14 [DateTime][POSIXct]s:
+        #'
+        #' - `cv`: The last modified time for the whole CV collection
+        #' - `activity`: The last modified time for the CV `activity_id`
+        #' - `experiment`: The last modified time for the CV `experiment_id`
+        #' - `sub_experiment`: The last modified time for the CV `sub_experiment_id`
+        #' - `institution`: The last modified time for the CV `institution_id`
+        #' - `source`: The last modified time for the CV `source_id`
+        #' - `table`: The last modified time for the CV `table_id`
+        #' - `frequency`: The last modified time for the CV `frequency`
+        #' - `grid_label`: The last modified time for the CV `grid_label`
+        #' - `realm`: The last modified time for the CV `realm`
+        #' - `source_type`: The last modified time for the CV `source_type`
+        #' - `req_global_atts`: The last modified time for the CV
+        #'   `required_global_attributes`
+        timestamp = function() {
+            private$m_timestamps[-1L]
+        },
+
+        #' @description
+        #' Get the time when the dictionary was built
+        #'
+        #' @return A [DateTime][POSIXct]
+        built_time = function() {
[…]
