+++ man/CMIP6Dict.Rd
@@ -1,0 +1,288 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/dict.R
+\name{CMIP6Dict}
+\alias{CMIP6Dict}
+\alias{cmip6_dict}
+\title{CMIP6 Controlled Vocabularies (CVs) and Data Request Dictionary}
+\usage{
+cmip6_dict()
+}
+\description{
+The \code{CMIP6Dict} object provides functionalities to fetch the latested CMIP6
+Controlled Vocabularies (CVs) and Data Request information.
+}
+\details{
+The CMIP6 CVs defines a well-defined set of global attributes that are
+recorded in each CMIP6 model output, providing information necessary for
+interpreting the data.
+
+The CMIP6 Data Request defines all the quantities from CMIP6 simulations that
+should be archived. This includes both quantities of general interest needed
+from most of the CMIP6-endorsed model intercomparison projects (MIPs) and
+quantities that are more specialized and only of interest to a single
+endorsed MIP.
+
+The data of CVs for use in CMIP6 is stored as JSON files in a
+\href{https://github.com/WCRP-CMIP/CMIP6_CVs}{GitHub Repo}, while the Data Request
+data is stored a Microsoft Excel file (\code{CMIP6_MIP_tables.xlsx}) in a
+\href{http://proj.badc.rl.ac.uk/svn/exarch/CMIP6dreq/trunk}{Subversion repo}.
+\code{CMIP6Dict} object is able to fetch those files and parse them into
+\link[data.table:data.table]{data.table}s.
+
+For more information, please see:
+\itemize{
+\item \href{https://docs.google.com/document/d/1h0r8RZr_f3-8egBMMh7aqLwy3snpD6_MrDz1q8n5XUk/edit}{CMIP6 Global Attributes, DRS, Filenames, Directory Structure, and CV's}
+\item \href{https://wcrp-cmip.github.io/WGCM_Infrastructure_Panel/CMIP6/data_request.html}{CMIP6 Data Request}
+}
+}
+\examples{
+# create a new CMIP6Dict object
+dict <- cmip6_dict()
+
+}
+\author{
+Hongyuan Jia
+}
+\section{Methods}{
+\subsection{Public methods}{
+\itemize{
+\item \href{#method-version}{\code{CMIP6Dict$version()}}
+\item \href{#method-timestamp}{\code{CMIP6Dict$timestamp()}}
+\item \href{#method-built_time}{\code{CMIP6Dict$built_time()}}
+\item \href{#method-build}{\code{CMIP6Dict$build()}}
+\item \href{#method-list}{\code{CMIP6Dict$list()}}
+\item \href{#method-update}{\code{CMIP6Dict$update()}}
+\item \href{#method-save}{\code{CMIP6Dict$save()}}
+\item \href{#method-load}{\code{CMIP6Dict$load()}}
+\item \href{#method-print}{\code{CMIP6Dict$print()}}
+\item \href{#method-clone}{\code{CMIP6Dict$clone()}}
+}
+}
+\if{html}{\out{<hr>}}
+\if{html}{\out{<a id="method-version"></a>}}
+\if{latex}{\out{\hypertarget{method-version}{}}}
+\subsection{Method \code{version()}}{
+Get the version of CVs and Data Request
+\subsection{Usage}{
+\if{html}{\out{<div class="r">}}\preformatted{CMIP6Dict$version()}\if{html}{\out{</div>}}
+}
+
+\subsection{Returns}{
+A list of two element:
+\itemize{
+\item \code{cv}: A \link{numeric_version} object giving the version of CVs
[…]
