### 2022-03-30 9b48e9a
[feat] Add method `CMIP6Dict$is_empty()`
