+++ R/dict.R
@@ -53,6 +53,18 @@ CMIP6Dict <- R6::R6Class("CMIP6Dict",
         },
 
         #' @description
+        #' Is it an empty CMIP6Dict?
+        #'
+        #' `$is_empty()` checks if this `CMIP6Dict` is empty, i.e. the
+        #' `$build()` or `$load()` method hasn't been called yet and there is no
+        #' data of CVs and Data Request.
+        #'
+        #' @return A single logical value of `TRUE` or `FALSE`.
+        is_empty = function() {
+            is.null(private$m_version)
+        },
+
+        #' @description
         #' Get the last modified time for CVs
         #'
         #' @return A list of 14 [DateTime][POSIXct]s:
@@ -79,7 +91,7 @@ CMIP6Dict <- R6::R6Class("CMIP6Dict",
         #'
         #' @return A [DateTime][POSIXct]
         built_time = function() {
-            private$m_timestamps[names(private$m_timestamps) == "db"]
+            private$m_timestamps[names(private$m_timestamps) == "dict"]
         },
 
         #' @description
@@ -87,48 +99,48 @@ CMIP6Dict <- R6::R6Class("CMIP6Dict",
         #'
         #' @return The updated `CMIP6Dict` itself.
         build = function() {
-            db <- cmip6dict_build()
+            dict <- cmip6dict_build()
 
             private$m_version = list(
-                cv = db$cvs$activity$CV_collection_version,
-                req = db$mip_table$version
+                cv = dict$cvs$activity$CV_collection_version,
+                req = dict$mip_table$version
             )
 
             private$m_timestamps <- list(
-                cv              = db$cvs$activity$CV_collection_modified,
-                activity        = db$cvs$activity$activity_id_CV_modified,
-                experiment      = db$cvs$experiment$experiment_id_CV_modified,
-                sub_experiment  = db$cvs$sub_experiment$sub_experiment_id_CV_modified,
-                institution     = db$cvs$institution$institution_id_CV_modified,
-                source          = db$cvs$source$source_id_CV_modified,
-                table           = db$cvs$table$table_id_CV_modified,
-                frequency       = db$cvs$frequency$frequency_CV_modified,
-                grid_label      = db$cvs$grid_label$grid_label_CV_modified,
-                realm           = db$cvs$realm$realm_CV_modified,
-                source_type     = db$cvs$source_type$source_type_CV_modified,
-                req_global_atts = db$cvs$req_global_atts$required_global_attributes_CV_modified,
-                resolution      = db$cvs$resolution$nominal_resolution_CV_modified,
-                db              = db$built_time
+                cv              = dict$cvs$activity$CV_collection_modified,
+                activity        = dict$cvs$activity$activity_id_CV_modified,
+                experiment      = dict$cvs$experiment$experiment_id_CV_modified,
+                sub_experiment  = dict$cvs$sub_experiment$sub_experiment_id_CV_modified,
+                institution     = dict$cvs$institution$institution_id_CV_modified,
+                source          = dict$cvs$source$source_id_CV_modified,
+                table           = dict$cvs$table$table_id_CV_modified,
+                frequency       = dict$cvs$frequency$frequency_CV_modified,
+                grid_label      = dict$cvs$grid_label$grid_label_CV_modified,
+                realm           = dict$cvs$realm$realm_CV_modified,
+                source_type     = dict$cvs$source_type$source_type_CV_modified,
+                req_global_atts = dict$cvs$req_global_atts$required_global_attributes_CV_modified,
+                resolution      = dict$cvs$resolution$nominal_resolution_CV_modified,
+                dict              = dict$built_time
             )
 
[…]
