### 2022-03-30 d0ddc96
[feat] Save the CMIP6 into package env
