### 2022-03-30 0476c18
[doc] Add examples for `CMIP6Dict`
