### 2022-03-31 b2c9939
[refactor] Return self in `CMIP6Dict$load()`
