### 2022-03-31 882d1bb
[doc] Add `CMIP6Dict$print()` in examples
