+++ man/CMIP6Dict.Rd
@@ -80,6 +80,8 @@ dict$save()
 new_dict <- cmip6_dict()
 new_dict$load()
 
+# print will show the version summary and the last built time
+dict$print()
 }
 }
 \author{
