### 2022-03-31 5d9231b
[test] Add CMIP6Dict tests
