+++ tests/testthat/test-dict.R
@@ -1,0 +1,195 @@
+test_that("cmip6_dict()", {
+    expect_is(dict <- cmip6_dict(), "CMIP6Dict")
+    expect_is(EPWSHIFTR_ENV$dict, "CMIP6Dict")
+})
+
+test_that("$build() and $save()", {
+    dict <- cmip6_dict()
+
+    expect_is(dict$build(), "CMIP6Dict")
+
+    expect_true(file.exists(dict$save(test_path())))
+})
+
+test_that("$version()", {
+    dict <- cmip6_dict()
+    expect_null(dict$version())
+
+    expect_is(dict$load(test_path()), "CMIP6Dict")
+    expect_is(dict$version(), "list")
+    expect_equal(names(dict$version()), c("cv", "req"))
+    expect_is(dict$version()$cv, "numeric_version")
+    expect_is(dict$version()$req, "numeric_version")
+})
+
+test_that("$is_empty()", {
+    dict <- cmip6_dict()
+    expect_true(dict$is_empty())
+
+    expect_is(dict$load(test_path()), "CMIP6Dict")
+    expect_false(dict$is_empty())
+    expect_is(dict$version()$req, "numeric_version")
+})
+
+test_that("$timestamp()", {
+    dict <- cmip6_dict()
+    expect_null(dict$timestamp())
+
+    expect_is(dict$load(test_path()), "CMIP6Dict")
+    expect_is(dict$timestamp(), "list")
+    expect_equal(length(dict$timestamp()), 13L)
+    expect_equal(names(dict$timestamp()),
+        c("cv", "activity", "experiment", "sub_experiment", "institution",
+            "source", "table", "frequency", "grid_label", "realm", "source_type",
+            "req_global_atts", "resolution"
+        )
+    )
+})
+
+test_that("$built_time()", {
+    dict <- cmip6_dict()
+    expect_null(dict$built_time())
+
+    expect_is(dict$load(test_path()), "CMIP6Dict")
+    expect_is(dict$built_time(), "POSIXct")
+    expect_equal(length(dict$built_time()), 1L)
+})
+
+test_that("$list()", {
+    dict <- cmip6_dict()
+    expect_null(dict$list("activity"))
+
+    expect_is(dict$load(test_path()), "CMIP6Dict")
+
+    expect_is(dict$list("activity"), "data.table")
+    expect_equal(
+        vapply(dict$list("activity"), typeof, ""),
+        c(activity_id = "character", description = "character")
+    )
+
+    expect_is(dict$list("experiment"), "data.table")
+    expect_equal(
+        x <- vapply(dict$list("experiment"), typeof, ""),
+        c(experiment_id = "character", activity_id = "list",
+          additional_allowed_model_components = "list",
+          description = "character", end_year = "integer",
+          experiment = "character", min_number_yrs_per_sim = "integer",
+          parent_activity_id = "list", parent_experiment_id = "list",
+          required_model_components = "list", start_year = "integer",
+          sub_experiment_id = "list", tier = "integer"
+        )
+    )
+
+    expect_is(dict$list("sub_experiment"), "data.table")
+    expect_equal(
+        vapply(dict$list("sub_experiment"), typeof, ""),
+        c(sub_experiment_id = "character", description = "character")
+    )
+
+    expect_is(dict$list("institution"), "data.table")
+    expect_equal(
+        vapply(dict$list("institution"), typeof, ""),
+        c(institution_id = "character", description = "character")
+    )
+
+    expect_is(dict$list("source"), "data.table")
+    expect_equal(
+        x <- vapply(dict$list("source"), typeof, ""),
+        c(source_id = "character", activity_participation = "list",
+          cohort = "character", institution_id = "list",
[…]
