### 2022-03-31 4e02a4b
[doc] Update NEWS
