### 2022-03-31 096cdae
Revert "[CI] Add mockthat in GitHub Actions"
This reverts commit fade967778ef8eeedbdbc15596670d89a1611c2c.
