+++ .github/workflows/pr-commands.yaml
@@ -31,7 +31,7 @@ jobs:
 
       - uses: r-lib/actions/setup-r-dependencies@v2
         with:
-          extra-packages: any::roxygen2, any::mockthat
+          extra-packages: any::roxygen2
           needs: pr-document
 
       - name: Document
