### 2022-03-31 fade967
[CI] Add mockthat in GitHub Actions
