+++ .github/workflows/R-CMD-check.yaml
@@ -56,7 +56,7 @@ jobs:
 
       - uses: r-lib/actions/setup-r-dependencies@v2
         with:
-          extra-packages: any::rcmdcheck
+          extra-packages: any::rcmdcheck, any::mockthat
           needs: check
 
       - uses: r-lib/actions/check-r-package@v2
