### 2022-03-31 4bb1c18
[doc] Add `cmip6_dict()`in pkgdown yaml
