+++ _pkgdown.yml
@@ -10,6 +10,9 @@ development:
   model: auto
 
 reference:
+  - title: CMIP6 Controlled Vocabularies (CVs) and Data Request
+    contents:
+      - cmip6_dict()
   - title: CMIP6 Index
     desc: Build CMIP6 output index
     contents:
