### 2022-03-31 2fa513b
[test] Remove mockthat functions and clean code
