+++ R/dict.R
@@ -551,7 +551,7 @@ cmip6dict_fetch_mip_table <- function(tag, dir = tempdir()) {
     verbose("Fetching CMIP6 MIP Table...")
     url <- sprintf("http://proj.badc.rl.ac.uk/svn/exarch/CMIP6dreq/tags/%s/dreqPy/docs/CMIP6_MIP_tables.xlsx", tag)
     f <- file.path(dir, basename(url))
-    download.file(url, f, mode = "wb", quiet = TRUE)
+    utils::download.file(url, f, mode = "wb", quiet = TRUE)
 
     # parse version
     notes <- readxl::read_excel(f, "Notes", skip = 1)
@@ -674,243 +674,4 @@ cmip6dict_update <- function(cv_sha, req_tag) {
     verbose("CMIP6 Dictionary updated successfully at ", format(out$built_time), ".")
 
     out
-}
-
-cmip6dict_validate_query_input <- function(
-    DICT = NULL,
-    activity = NULL, experiment = NULL, sub_experiment = NULL,
-    institution = NULL, source = NULL, table = NULL, frequency = NULL,
-    grid_label = NULL, realm = NULL, source_type = NULL, resolution = NULL,
-    variable = NULL, variant = NULL)
-{
-    if (is.null(DICT) || !length(DICT)) {
-        stop("Current CMIP6Dict is empty. Please first run '$build()' to ",
-            "collection all CMIP6 Controlled Vocabularies (CVs) and MIP Table.",
-            call. = FALSE)
-    }
-
-    validate_input <- function(input) {
-        type <- deparse(substitute(input))
-        if (type == "variable") {
-            name <- "mipt_table"
-            colnm <- "variable_name"
-        } else {
-            name <- type
-            colnm <- type_to_std_name(type)
-        }
-
-        if (!checkmate::test_subset(input, empty.ok = TRUE, choices = DICT[[name]][[colnm]])) {
-            stop(sprintf(
-                "Invalid %s found: [%s]. Please run '$list_all(\"%s\")' to see all possible options.",
-                type, paste0("'", input, "'", collapse = ","), type
-            ))
-        }
-    }
-
-    validate_input(activity)
-    validate_input(experiment)
-    validate_input(sub_experiment)
-    validate_input(institution)
-    validate_input(source)
-    validate_input(table)
-    validate_input(frequency)
-    validate_input(grid_label)
-    validate_input(realm)
-    validate_input(source_type)
-    validate_input(resolution)
-    validate_input(variable)
-    assert_character(variant, any.missing = FALSE, pattern = "r\\d+i\\d+p\\d+f\\d+", null.ok = TRUE)
-}
-
-cmip6dict_esgf_query <- function(
-    DICT = NULL,
-    activity = NULL, experiment = NULL, sub_experiment = NULL,
-    institution = NULL, source = NULL, table = NULL, frequency = NULL,
-    grid_label = NULL, realm = NULL, source_type = NULL, resolution = NULL,
-    variable = NULL, variant = NULL,
-    count = FALSE, replica = FALSE, latest = TRUE, type = "Dataset", limit = 1000L, data_node = NULL)
-{
-    cmip6dict_validate_query_input(DICT,
-        activity = activity, experiment = experiment,
-        sub_experiment = sub_experiment, institution = institution,
-        source = source, table = table, frequency = frequency,
-        grid_label = grid_label, realm = realm, source_type = source_type,
-        resolution = resolution, variable = variable, variant = variant
-    )
-    assert_choice(type, c("Dataset", "File"))
-    assert_flag(count)
-    assert_flag(replica)
-    assert_flag(latest)
-    assert_count(limit, positive = TRUE)
[…]
