+++ man/print.CMIP6QueryFacets.Rd
@@ -1,18 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/dict.R
-\name{print.CMIP6QueryFacets}
-\alias{print.CMIP6QueryFacets}
-\title{Print Results of CMIP6 Query Facets}
-\usage{
-\method{print}{CMIP6QueryFacets}(x, max = 10L, ...)
-}
-\arguments{
-\item{x}{An \code{CMIP6QueryFacets} object.}
-
-\item{max}{How many elements to list in each facet field. Default: \code{10L}.}
-
-\item{...}{Further arguments passed to or from other methods.}
-}
-\description{
-Print Results of CMIP6 Query Facets
-}
