+++ NAMESPACE
@@ -1,6 +1,5 @@
 # Generated by roxygen2: do not edit by hand
 
-S3method(print,CMIP6QueryFacets)
 export(CMIP6Dict)
 export(cmip6_dict)
 export(esgf_query)
