### 2022-04-01 48b307f
[doc] Fix `CMIP6Dict` docs
