### 2022-04-01 3b4f0b5
[doc] Add pkgdown article about `CMIP6Dict`
