+++ vignettes/articles/cmip6_dict.Rmd
@@ -1,0 +1,231 @@
+---
+title: "Manage CMIP6 CVs and Data Request"
+author: "Hongyuan Jia"
+output:
+    rmarkdown::html_vignette:
+        df_print: "paged"
+        toc: true
+vignette: >
+    %\VignetteIndexEntry{Introduction to eplusr}
+    %\VignetteEngine{knitr::rmarkdown}
+    %\VignetteEncoding{UTF-8}
+description: >
+    This vignette introduces you how to use epwshiftr to fetch, parse and store
+    CMIP6 Controlled Vocabularies (CVs) and Data Request.
+---
+
+```{r, include = FALSE}
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>"
+)
+```
+
+```{r setup, echo = FALSE, message = FALSE}
+library(epwshiftr)
+
+# show verbose messages
+options(epwshiftr.verbose = TRUE)
+```
+
+# Controlled Vocabularies (CVs) and Data Request
+
+All CMIP6 outputs are written in netCDF files in conformance with the CMIP
+standards. The Controlled Vocabularies (CVs) and Data Request play a key role in
+ensuring uniformity in the description of data sets across all models.
+
+The CMIP6 CVs gives a well-defined set of global attributes that are recorded in
+each CMIP6 model output, providing information necessary for interpreting the
+data. The data of CVs for use in CMIP6 is stored as JSON files in a
+[GitHub Repo](https://github.com/WCRP-CMIP/CMIP6_CVs).
+
+The CMIP6 Data Request defines all the quantities from CMIP6 simulations that
+should be archived. This includes both quantities of general interest needed
+from most of the CMIP6-endorsed model intercomparison projects (MIPs) and
+quantities that are more specialized and only of interest to a single endorsed
+MIP. The Data Request data is stored a Microsoft Excel file
+(`CMIP6_MIP_tables.xlsx`) in a
+[Subversion repo](http://proj.badc.rl.ac.uk/svn/exarch/CMIP6dreq/trunk).
+
+For more information, please see:
+
+- [CMIP6 Global Attributes, DRS, Filenames, Directory Structure, and CV's](https://docs.google.com/document/d/1h0r8RZr_f3-8egBMMh7aqLwy3snpD6_MrDz1q8n5XUk/edit)
+- [CMIP6 Data Request](https://wcrp-cmip.github.io/WGCM_Infrastructure_Panel/CMIP6/data_request.html)
+
+# Introduce the `CMIP6Dict` class
+
+epwshiftr provides a `CMIP6Dict` class to help you fetch, parse and store CMIP6
+CVs and Data Request.
+
+## Create a `CMIP6Dict` object
+
+You can create a new `CMIP6Dict` object using the `cmip6_dict()` function. It
+takes no argument and simply returns an empty `CMIP6Dict`.
+
+> It is an [R6](https://r6.r-lib.org/) object with reference semantics. All
+> methods in an `R6` object can be invoked using the `$method_name()` style.
+
+```{r}
+dict <- cmip6_dict()
+dict
+```
+
+You can use `$is_empty()` to check if there are any contents in current
+`CMIP6Dict`.
+
+```{r}
+dict$is_empty()
+```
+
+## Fetch data of CVs and Data Request
+
+After created, you can use `$build()` to fetch all data of CMIP6 CVs and Data
+Request.
+
+Currently, the supproted CVs are:
+
+1. `activity_id`: Activity identifiers, e.g. `HighResMIP`, `ScenarioMIP`.
+1. `experiment_id`: Root experiment identifiers, e.g. `historical`, `ssp126`.
[…]
