+++ _pkgdown.yml
@@ -9,6 +9,29 @@ authors:
 development:
   model: auto
 
+navbar:
+  structure:
+    left:  [home, articles, reference, news]
+    right: [github]
+  components:
+    home:
+      icon: fas fa-home fa-lg
+      href: index.html
+    articles:
+      text: Vignettes
+      menu:
+      - text: "Introduction to eplusr"
+        href: articles/eplusr.html
+    reference:
+      text: Manual
+      href: reference/index.html
+    news:
+      text: News
+      href: news/index.html
+    github:
+      icon: fa-github fa-lg
+      href: https://github.com/ideas-lab-nus/epwshiftr
+
 reference:
   - title: CMIP6 Controlled Vocabularies (CVs) and Data Request
     contents:
