### 2022-04-01 6252910
[test] Skip dict tests on CRAN
