### 2022-04-01 c905dca
[refactor] Remove deep_clone fun
