### 2022-04-01 b9943af
[doc] Update NEWS
