### 2022-05-05 a87f8c0
[feat] Fall back to shift if alpha is too big
