### 2022-05-05 131b052
[feat] Fall back to shift if alpha is too big
