+++ R/epwshiftr-package.R
@@ -10,6 +10,12 @@
 #'
 #' * `epwshiftr.verbose`: If `TRUE`, more detailed message will be printed.
 #'   Default: `FALSE`.
+#' * `epwshiftr.threshold_alpha`: the threshold of the absolute value for alpha,
+#'   i.e. monthly-mean fractional change, when performing morphing operations.
+#'   The default value is set to `3`. If the morphing methods are set
+#'   `"stretch"` or `"combined"`, and the absolute alpha exceeds the threshold
+#'   value, warnings are issued and the morphing method is fallbacked to
+#'   `"shift"` to avoid unrealistic morphed values.
 #' * `epwshiftr.dir`: The directory to store package data, including CMIP6
 #'   model output file index and etc. If not set, the current user data
 #'   directory will be used.
@@ -26,7 +32,10 @@ EPWSHIFTR_ENV$index_db <- NULL
 # set package options
 # reference: https://github.com/Rdatatable/data.table/blob/master/R/onLoad.R
 .onLoad <- function (libname, pkgname) {
-    .opts  <-  c("epwshiftr.verbose" = "FALSE")
+    .opts  <-  c(
+        "epwshiftr.verbose" = "FALSE",
+        "epwshiftr.threshold_alpha" = "3"
+    )
     for (i in setdiff(names(.opts), names(options()))) {
         eval(parse(text = paste0("options(",i,"=",.opts[i],")")))
     }
