### 2022-05-20 #4030 @hongyuanjia
Hi @KristinaRiemer. Most of the feature-related issues have been solved, except for https://github.com/ideas-lab-nus/epwshiftr/issues/19, which may need more refactoring. Also, to solve https://github.com/ideas-lab-nus/epwshiftr/issues/38 in a more comprehensive way, I have decided to add a new feature to download and parse all available CMIP6 controlled variables and GCM outputs (https://github.com/ideas-lab-nus/epwshiftr/pull/53). This part has almost been completed. I should be able to complete them at the end of this month. Then other issues related to documentation and the paper structure next month. Apologies for the slow moving. I hope I could solve all issues by the mid of next month.
