### 2022-06-27 a726acf
[feat] Fall back to shift if alpha is too big (#55)
