+++ NEWS.md
@@ -70,6 +70,12 @@
   warnings will be issued for cases with input data less than a decade (10 years)
   . This is because using data that only covers a short period of time may not
   be able to capture the average of future climate (#41).
+* A new option `epwshiftr.threshold_alpha` has been added to set the threshold
+  of the absolute value for alpha, i.e. monthly-mean fractional change when
+  performing morphing operations. The default value is set to `3`. If the
+  morphing methods are set `"stretch"` or `"combined"`, and the absolute alpha
+  exceeds the threshold value, warnings are issued and the morphing method is
+  fallbacked to `"shift"` to avoid unrealistic morphed values (#54).
 
 ## Bug fixes
 
