### 2022-07-24 844926c
[refactor] Use DReq data from `PCMDI/cmip6-cmor-tables` repo
