+++ R/dict.R
@@ -1,23 +1,22 @@
 #' CMIP6 Controlled Vocabularies (CVs) and Data Request Dictionary
 #'
 #' The `CMIP6Dict` object provides functionalities to fetch the latested CMIP6
-#' Controlled Vocabularies (CVs) and Data Request information.
+#' Controlled Vocabularies (CVs) and Data Request (DReq) information.
 #'
 #' The CMIP6 CVs gives a well-defined set of global attributes that are recorded
 #' in each CMIP6 model output, providing information necessary for interpreting
-#' the data.
+#' the data. The data of CMIP6 CVs is stored as JSON files in the WCRP-CMIP
+#' [GitHub Repo](https://github.com/WCRP-CMIP/CMIP6_CVs).
 #'
-#' The CMIP6 Data Request defines all the quantities from CMIP6 simulations that
-#' should be archived. This includes both quantities of general interest needed
-#' from most of the CMIP6-endorsed model intercomparison projects (MIPs) and
+#' The CMIP6 DReq defines all the quantities from CMIP6 simulations that should
+#' be archived. This includes both quantities of general interest needed from
+#' most of the CMIP6-endorsed model intercomparison projects (MIPs) and
 #' quantities that are more specialized and only of interest to a single
-#' endorsed MIP.
-#'
-#' The data of CVs for use in CMIP6 is stored as JSON files in a
-#' [GitHub Repo](https://github.com/WCRP-CMIP/CMIP6_CVs), while the Data Request
-#' data is stored a Microsoft Excel file (`CMIP6_MIP_tables.xlsx`) in a
+#' endorsed MIP. The raw data of DReq is stored a Microsoft Excel file
+#' (`CMIP6_MIP_tables.xlsx`) in a
 #' [Subversion repo](http://proj.badc.rl.ac.uk/svn/exarch/CMIP6dreq/trunk).
-#' `CMIP6Dict` object is able to fetch those files and parse them.
+#' The `CMIP6Dict` object uses the parsed DReq data that is stored in the
+#' [GitHub Repo](https://github.com/PCMDI/cmip6-cmor-tables).
 #'
 #' For more information, please see:
 #'
@@ -45,22 +44,18 @@
 #' # get the time when the dict was built
 #' dict$built_time()
 #'
-#' # check if there is updates of CVs and Data Request
-#' dict$update()
-#'
-#' # list the data of CVs
-#' dict$list("activity")
-#' dict$list("experiment")
-#' dict$list("sub_experiment")
-#' dict$list("institution")
-#' dict$list("source")
-#' dict$list("table")
-#' dict$list("frequency")
-#' dict$list("grid_label")
-#' dict$list("realm")
-#' dict$list("source_type")
-#' dict$list("req_global_atts")
-#' dict$list("variable")
+#' # get the data of CVs and DReq
+#' dict$get("activity_id")
+#' dict$get("experiment_id")
+#' dict$get("sub_experiment_id")
+#' dict$get("institution_id")
+#' dict$get("source_id")
+#' dict$get("table_id")
+#' dict$get("frequency")
+#' dict$get("grid_label")
+#' dict$get("realm")
+#' dict$get("source_type")
+#' dict$get("dreq")
 #'
 #' # save the dict object for later usage
 #' # default location is the value of global option "epwshiftr.dir"
@@ -93,8 +88,8 @@ CMIP6Dict <- R6::R6Class("CMIP6Dict",
         #'
         #' @return A list of two element:
         #'
-        #' - `cv`: A [numeric_version] object giving the version of CVs
-        #' - `req`: A [numeric_version] object giving the version of Data
+        #' - `cvs`: A [numeric_version] object giving the version of CVs
[…]
