+++ NAMESPACE
@@ -1,5 +1,7 @@
 # Generated by roxygen2: do not edit by hand
 
+S3method(print,CMIP6CV)
+S3method(print,CMIP6DReq)
 export(CMIP6Dict)
 export(cmip6_dict)
 export(esgf_query)
@@ -33,6 +35,7 @@ importFrom(checkmate,assert_subset)
 importFrom(checkmate,test_directory_exists)
 importFrom(checkmate,test_file_exists)
 importFrom(checkmate,test_r6)
+importFrom(cli,cli_rule)
 importFrom(data.table,":=")
 importFrom(data.table,as.data.table)
 importFrom(data.table,copy)
@@ -51,11 +54,12 @@ importFrom(eplusr,read_epw)
 importFrom(fst,write_fst)
 importFrom(future.apply,future_Map)
 importFrom(future.apply,future_lapply)
+importFrom(jsonlite,fromJSON)
+importFrom(jsonlite,parse_json)
 importFrom(jsonlite,read_json)
 importFrom(progressr,with_progress)
 importFrom(psychrolib,GetTDewPointFromRelHum)
 importFrom(psychrolib,SetUnitSystem)
 importFrom(rappdirs,user_data_dir)
-importFrom(readxl,read_excel)
 importFrom(units,set_units)
 importFrom(utils,menu)
