+++ DESCRIPTION
@@ -47,13 +47,14 @@ Encoding: UTF-8
 URL: https://github.com/ideas-lab-nus/epwshiftr
 BugReports: https://github.com/ideas-lab-nus/epwshiftr/issues
 Roxygen: list(markdown = TRUE)
-RoxygenNote: 7.1.2
+RoxygenNote: 7.2.0
 Collate:
     'coord.R'
     'dict.R'
     'utils.R'
     'epwshiftr-package.R'
     'esgf.R'
+    'gh.R'
     'morph.R'
     'netcdf.R'
 VignetteBuilder: knitr

[…]
