### 2022-07-24 c789296
[fix] Add `cli` dep
