+++ DESCRIPTION
@@ -22,6 +22,7 @@ Description:
     using data from Global Climate Models ('GCM').
 Imports:
     checkmate,
+    cli,
     data.table (>= 1.12.4),
     eplusr,
     fst,
