### 2022-07-24 7a94dea
[fix] Clean up code
