+++ man/CMIP6Dict.Rd
@@ -178,9 +178,18 @@ A \link[=POSIXct]{DateTime}
 \subsection{Method \code{build()}}{
 Fetch and parse all data of CVs and Data Request
 \subsection{Usage}{
-\if{html}{\out{<div class="r">}}\preformatted{CMIP6Dict$build()}\if{html}{\out{</div>}}
+\if{html}{\out{<div class="r">}}\preformatted{CMIP6Dict$build(token = NULL)}\if{html}{\out{</div>}}
 }
 
+\subsection{Arguments}{
+\if{html}{\out{<div class="arguments">}}
+\describe{
+\item{\code{token}}{A string of GitHub token that is used to access GitHub
+REST APIs. If \code{NULL}, \code{GITHUB_PAT} or \code{GITHUB_TOKEN}
+environment variable will be used if exists. Default: \code{NULL}.}
+}
+\if{html}{\out{</div>}}
+}
 \subsection{Returns}{
 The updated \code{CMIP6Dict} itself.
 }
