### 2022-07-25 0de6dda
[doc] Fix NEWS
