+++ NEWS.md
@@ -77,8 +77,8 @@
   of the absolute value for alpha, i.e. monthly-mean fractional change when
   performing morphing operations. The default value is set to `3`. If the
   morphing methods are set `"stretch"` or `"combined"`, and the absolute alpha
-  exceeds the threshold value, the morphing method is fallbacked to `"shift"` to
-  avoid unrealistic morphed values (#54).
+  exceeds the threshold value, warnings are issued and the morphing method is
+  fallbacked to `"shift"` to avoid unrealistic morphed values (#54).
 
 ## Bug fixes
 
