### 2022-07-25 81b7206
[doc] Fix vignette
