+++ vignettes/articles/cmip6_dict.Rmd
@@ -54,13 +54,13 @@ For more information, please see:
 
 # Introduce the `CMIP6Dict` class
 
-epwshiftr provides a `CMIP6Dict` class to help you fetch, parse and store CMIP6
-CVs and Data Request.
+{epwshiftr} provides a `CMIP6Dict` class to help you fetch, parse and store
+CMIP6 CVs and Data Request.
 
 ## Create a `CMIP6Dict` object
 
 You can create a new `CMIP6Dict` object using the `cmip6_dict()` function. It
-takes no argument and simply returns an empty `CMIP6Dict`.
+takes no argument and simply returns an empty `CMIP6Dict` object.
 
 > It is an [R6](https://r6.r-lib.org/) object with reference semantics. All
 > methods in an `R6` object can be invoked using the `$method_name()` style.
@@ -82,21 +82,27 @@ dict$is_empty()
 After created, you can use `$build()` to fetch all data of CMIP6 CVs and Data
 Request.
 
-Currently, the supproted CVs are:
+The GitHub RESTful APIs are used to to fetch the latest tag of CVs and Data
+Request. You can specify your GitHub token by specifying the `token` argument in
+`$build()`. By default, it will uses `GITHUB_PAT` or `GITHUB_TOKEN` environment
+variable if exists.
 
+Currently, all supproted CVs are:
+
+1. `drs`: Data Reference Syntax (DRS).
 1. `activity_id`: Activity identifiers, e.g. `HighResMIP`, `ScenarioMIP`.
 1. `experiment_id`: Root experiment identifiers, e.g. `historical`, `ssp126`.
-1. `sub_experiment_id`: Description of sub-experiment, e.g. `none`, `s1960`.
-1. `institution_id`: Institution identifiers, e.g. `AWI`, `CAS`.
-1. `source_id`: Model identifiers, e.g. `GFDL-CM2-1`, `ACCESS-CM2`.
-1. `source_type`: Model configurations, e.g. `AGCM`, `OGCM`.
-1. `table_id`: Table identifiers, e.g. `Amon`, `Oday`.
 1. `frequency`: Sampling frequencies, e.g. `mon`, `day`.
 1. `grid_label`: Grid identifiers, e.g. `gn`, `gr`.
-1. `realm`: Realms where variables are defined, e.g. `atmos`, `ocean`.
+1. `institution_id`: Institution identifiers, e.g. `AWI`, `CAS`.
 1. `nominal_resolution`: Approximate horizontal resolutions, e.g. `50 km`, `100
    km`.
+1. `realm`: Realms where variables are defined, e.g. `atmos`, `ocean`.
 1. `required_global_attributes`: Names of required global attributes.
+1. `source_id`: Model identifiers, e.g. `GFDL-CM2-1`, `ACCESS-CM2`.
+1. `source_type`: Model configurations, e.g. `AGCM`, `OGCM`.
+1. `sub_experiment_id`: Description of sub-experiment, e.g. `none`, `s1960`.
+1. `table_id`: Table identifiers, e.g. `Amon`, `Oday`.
 
 ```{r}
 dict$build()
@@ -105,7 +111,7 @@ dict$is_empty()
 
 You can get the version of the CV collection and Data Request using `$version()`
 method. It returns a list of two `numeric_version`s giving you the current
-version of fetched CV collection (`cv`) and Data Request (`req`):
+version of fetched CV collection (`cvs`) and Data Request (`dreq`):
 
 ```{r}
 dict$version()
@@ -117,98 +123,70 @@ last_built <- dict$built_time()
 last_built
 ```
 
-The last modified time for each CV can be retrieved using the `$timestamp()`
-method:
[…]
