### 2022-07-25 cb5cbbe
[refactor] Clean up code
