+++ tests/testthat/test-morph.R
@@ -11,9 +11,9 @@ test_that("monthly_mean()", {
         epw$add_unit()
         d <- epw$data()
         expect_is(res1 <- monthly_mean(d, "dry_bulb_temperature"), "data.table")
-        expect_equal(nrow(res1), 12L)
-        expect_equal(names(res1), c("month", "val_mean", "val_max", "val_min"))
-        expect_equal(res1$month, 1:12)
+        expect_identical(nrow(res1), 12L)
+        expect_named(res1, c("month", "val_mean", "val_max", "val_min"))
+        expect_identical(res1$month, 1L:12L)
         expect_is(res1$val_mean, "units")
         expect_is(res1$val_max, "units")
         expect_is(res1$val_min, "units")
@@ -21,9 +21,9 @@ test_that("monthly_mean()", {
         epw$drop_unit()
         d <- epw$data()
         expect_is(res2 <- monthly_mean(d, "dry_bulb_temperature"), "data.table")
-        expect_equal(nrow(res2), 12L)
-        expect_equal(names(res2), c("month", "val_mean", "val_max", "val_min"))
-        expect_equal(res2$month, 1:12)
+        expect_identical(nrow(res2), 12L)
+        expect_named(res2, c("month", "val_mean", "val_max", "val_min"))
+        expect_identical(res2$month, 1L:12L)
         expect_is(res2$val_mean, "numeric")
         expect_is(res2$val_max, "numeric")
         expect_is(res2$val_min, "numeric")
@@ -40,26 +40,26 @@ test_that("preprocess_morphing()", {
         d <- fst::read_fst(path, as.data.table = TRUE)
 
         expect_warning(res <- preprocess_morphing(d, warning = TRUE), "less than a decade")
-        expect_equal(names(res),
+        expect_named(res,
             c("activity_drs", "institution_id", "source_id", "experiment_id",
               "member_id", "table_id", "lon", "lat", "dist", "units", "value", "month",
               "interval"
             )
         )
-        expect_equal(nrow(res), 12L)
+        expect_identical(nrow(res), 12L)
         expect_is(res$interval, "factor")
 
         # can stop if specified years were not found
-        expect_error(preprocess_morphing(d, years = 2059:2061), "does not contain any data")
+        expect_error(preprocess_morphing(d, years = 2059L:2061L), "does not contain any data")
 
-        expect_is(res <- preprocess_morphing(d, years = 2060, labels = as.factor("2060s")), "data.table")
-        expect_equal(names(res),
+        expect_is(res <- preprocess_morphing(d, years = 2060L, labels = as.factor("2060s")), "data.table")
+        expect_named(res,
             c("activity_drs", "institution_id", "source_id", "experiment_id",
               "member_id", "table_id", "lon", "lat", "dist", "units", "value", "month",
               "interval"
             )
         )
-        expect_equal(nrow(res), 12L)
+        expect_identical(nrow(res), 12L)
         expect_is(res$interval, "factor")
     }
 })
@@ -95,7 +95,7 @@ test_that("morphing_from_mean()", {
                 type = "shift"
             )
         )
-        expect_equal(names(res),
+        expect_named(res,
             c("activity_drs", "institution_id", "source_id", "experiment_id",
               "member_id", "table_id", "lon", "lat", "dist", "interval", "datetime",
               "year", "month", "day", "hour", "minute", "dry_bulb_temperature",
@@ -116,7 +116,7 @@ test_that("morphing_from_mean()", {
                 type = "shift"
             )
         )
-        expect_equal(names(res),
+        expect_named(res,
[…]
