+++ NAMESPACE
@@ -54,7 +54,6 @@ importFrom(eplusr,read_epw)
 importFrom(fst,write_fst)
 importFrom(future.apply,future_Map)
 importFrom(future.apply,future_lapply)
-importFrom(jsonlite,fromJSON)
 importFrom(jsonlite,parse_json)
 importFrom(jsonlite,read_json)
 importFrom(progressr,with_progress)
