### 2022-07-25 1e8bf24
[fix] Fix DReq loading
