### 2022-07-25 7701e2b
[refactor] Improve code coverage
