### 2022-07-25 6d87182
[doc] Fix ordered list in the vignette
