+++ tests/testthat/test-dict.R
@@ -1,0 +1,199 @@
+test_that("cmip6_dict()", {
+    expect_is(dict <- cmip6_dict(), "CMIP6Dict")
+    expect_is(EPWSHIFTR_ENV$dict, "CMIP6Dict")
+})
+
+test_that("$build() and $save()", {
+    skip_on_cran()
+
+    dict <- cmip6_dict()
+    expect_warning(dict$save())
+
+    expect_is(dict$build(), "CMIP6Dict")
+
+    expect_true(file.exists(dict$save(test_path())))
+    # can create dir if not exists
+    expect_true(file.exists(dict$save(file.path(tempdir(), "abc"))))
+})
+
+test_that("$load()", {
+    dict <- cmip6_dict()
+    expect_message(dict$load(file.path(tempdir(), "not_exists")))
+
+    empty_dir <- file.path(tempdir(), "empty")
+    if (!dir.exists(empty_dir)) dir.create(empty_dir)
+    empty <- file.path(empty_dir, "CMIP6DICT")
+    saveRDS(list(), empty)
+    expect_error(dict$load(dirname(empty)))
+})
+
+test_that("$version()", {
+    skip_on_cran()
+
+    dict <- cmip6_dict()
+    expect_null(dict$version())
+
+    expect_is(dict$load(test_path()), "CMIP6Dict")
+    expect_is(dict$version(), "list")
+    expect_named(dict$version(), c("cvs", "dreq"))
+    expect_is(dict$version()$cvs, "numeric_version")
+    expect_is(dict$version()$dreq, "numeric_version")
+})
+
+test_that("$is_empty()", {
+    skip_on_cran()
+
+    dict <- cmip6_dict()
+    expect_true(dict$is_empty())
+
+    expect_is(dict$load(test_path()), "CMIP6Dict")
+    expect_false(dict$is_empty())
+    expect_is(dict$version()$dreq, "numeric_version")
+})
+
+test_that("$timestamp()", {
+    skip_on_cran()
+
+    dict <- cmip6_dict()
+    expect_null(dict$timestamp())
+
+    expect_is(dict$load(test_path()), "CMIP6Dict")
+    expect_is(dict$timestamp(), "list")
+    expect_length(dict$timestamp(), 14L)
+    expect_named(dict$timestamp(),
+        c(
+            "cvs", "drs", "activity_id", "experiment_id", "frequency",
+            "grid_label", "institution_id", "nominal_resolution", "realm",
+            "required_global_attributes", "source_id", "source_type",
+            "sub_experiment_id", "table_id"
+        )
+    )
+})
+
+test_that("$built_time()", {
+    skip_on_cran()
+
+    dict <- cmip6_dict()
+    expect_null(dict$built_time())
+
+    expect_is(dict$load(test_path()), "CMIP6Dict")
+    expect_is(dict$built_time(), "POSIXct")
+    expect_length(dict$built_time(), 1L)
+})
+
+test_that("$get()", {
+    skip_on_cran()
+
+    dict <- cmip6_dict()
+
+    expect_null(dict$get("activity_id"))
+
+    expect_is(dict$load(test_path()), "CMIP6Dict")
+
+    expect_is(dict$get("drs"), "list")
+    expect_identical(
+        unique(vapply(dict$get("drs"), typeof, "", USE.NAMES = FALSE)),
+        "character"
+    )
+    expect_message(print(dict$get("drs")))
+
+    expect_is(dict$get("activity_id"), "list")
+    expect_identical(
+        unique(vapply(dict$get("activity_id"), typeof, "", USE.NAMES = FALSE)),
+        "character"
+    )
+    expect_message(print(dict$get("activity_id")))
+
+    expect_is(dict$get("experiment_id"), "data.table")
[…]
