+++ R/epwshiftr-package.R
@@ -27,6 +27,13 @@
 # package internal environment
 EPWSHIFTR_ENV <- new.env(parent = emptyenv())
 EPWSHIFTR_ENV$index_db <- NULL
+EPWSHIFTR_ENV$dict <- NULL
+
+## usethis namespace: start
+#' @importFrom checkmate assert_string
+#' @importFrom cli cli_rule
+#' @importFrom jsonlite parse_json
+## usethis namespace: end
 
 # nocov start
 # set package options
