### 2022-07-25 8dff716
[feat] Add functionality to parse CMIP6 Controlled Vocabularies (CVs) and Data Request (#53)
* [feat] Add CMIP6Dict class

* [feat] Add method `CMIP6Dict$is_empty()`

* [feat] Save the CMIP6 into package env

* [fix] Fix `CMIP6Dict$timestamp()` and `CMIP6Dict$built_time()`

* [doc] Add examples for `CMIP6Dict`

* [fix] Fix `CMIP6Dict$built_time()`

* [refactor] Return self in `CMIP6Dict$load()`

* [fix] Fix `cmip6dict_save()` and `cmip6dict_load()`

* [doc] Add `CMIP6Dict$print()` in examples

* [refactor] Add `dict` in package environment

* [test] Add CMIP6Dict tests

* [doc] Update NEWS

* 
[…]
