+++ R/dict.R
@@ -1,0 +1,946 @@
+#' CMIP6 Controlled Vocabularies (CVs) and Data Request Dictionary
+#'
+#' The `CMIP6Dict` object provides functionalities to fetch the latested CMIP6
+#' Controlled Vocabularies (CVs) and Data Request (DReq) information.
+#'
+#' The CMIP6 CVs gives a well-defined set of global attributes that are recorded
+#' in each CMIP6 model output, providing information necessary for interpreting
+#' the data. The data of CMIP6 CVs is stored as JSON files in the WCRP-CMIP
+#' [GitHub Repo](https://github.com/WCRP-CMIP/CMIP6_CVs).
+#'
+#' The CMIP6 DReq defines all the quantities from CMIP6 simulations that should
+#' be archived. This includes both quantities of general interest needed from
+#' most of the CMIP6-endorsed model intercomparison projects (MIPs) and
+#' quantities that are more specialized and only of interest to a single
+#' endorsed MIP. The raw data of DReq is stored a Microsoft Excel file
+#' (`CMIP6_MIP_tables.xlsx`) in a
+#' [Subversion repo](http://proj.badc.rl.ac.uk/svn/exarch/CMIP6dreq/trunk).
+#' The `CMIP6Dict` object uses the parsed DReq data that is stored in the
+#' [GitHub Repo](https://github.com/PCMDI/cmip6-cmor-tables).
+#'
+#' For more information, please see:
+#'
+#' - [CMIP6 Global Attributes, DRS, Filenames, Directory Structure, and CV's](https://docs.google.com/document/d/1h0r8RZr_f3-8egBMMh7aqLwy3snpD6_MrDz1q8n5XUk/edit)
+#' - [CMIP6 Data Request](https://wcrp-cmip.github.io/WGCM_Infrastructure_Panel/CMIP6/data_request.html)
+#'
+#' @examples
+#' \dontrun{
+#'
+#' # create a new CMIP6Dict object
+#' dict <- cmip6_dict()
+#'
+#' # by default, there is no data when the CMIP6Dict was created
+#' dict$is_empty()
+#'
+#' # fetch and parse all CVs and Data Request data
+#' dict$build()
+#'
+#' # get the version of CVs nand Data Request
+#' dict$version()
+#'
+#' # get the last modified time for each CV and Data Request
+#' dict$timestamp()
+#'
+#' # get the time when the dict was built
+#' dict$built_time()
+#'
+#' # get the data of CVs and DReq
+#' dict$get("activity_id")
+#' dict$get("experiment_id")
+#' dict$get("sub_experiment_id")
+#' dict$get("institution_id")
+#' dict$get("source_id")
+#' dict$get("table_id")
+#' dict$get("frequency")
+#' dict$get("grid_label")
+#' dict$get("realm")
+#' dict$get("source_type")
+#' dict$get("dreq")
+#'
+#' # save the dict object for later usage
+#' # default location is the value of global option "epwshiftr.dir"
+#' dict$save()
+#'
+#' # the saved dict object can be reloaded
+#' new_dict <- cmip6_dict()
+#' new_dict$load()
+#'
+#' # print will show the version summary and the last built time
+#' dict$print()
+#' }
+#' @author Hongyuan Jia
+#'
+#' @importFrom R6 R6Class
+#' @name CMIP6Dict
+#' @export
+cmip6_dict <- function() {
+    EPWSHIFTR_ENV$dict <- CMIP6Dict$new()
+    EPWSHIFTR_ENV$dict
+}
+
+#' @name CMIP6Dict
+#' @export
+CMIP6Dict <- R6::R6Class("CMIP6Dict",
+    cloneable = FALSE, lock_class = TRUE,
+    public = list(
+        #' @description
[…]
