+++ pkgdown.yml
@@ -1,7 +1,8 @@
 pandoc: 2.14.2
-pkgdown: 2.0.5
+pkgdown: 2.0.6
 pkgdown_sha: ~
 articles:
+  cmip6_dict: cmip6_dict.html
   paper: paper.html
-last_built: 2022-06-27T02:44Z
+last_built: 2022-07-25T02:12Z
 
