+++ pkgdown.yml
@@ -4,5 +4,5 @@ pkgdown_sha: ~
 articles:
   cmip6_dict: cmip6_dict.html
   paper: paper.html
-last_built: 2022-07-25T02:12Z
+last_built: 2022-09-21T14:45Z
 
