### 2022-09-21 8a50fc0
[feat] Allow `replica` to be `NULL`
