### 2022-09-21 0864a3b
[fix] Remove redundant code
