### 2022-09-21 04ce88c
[feat] Allow `replica` to be `NULL` for ESGF query (#71)
* [feat] Allow `replica` to be `NULL`

* [fix] Remove redundant code
