### 2022-09-21 002d34b
[feat] Allow `replica` to be `NULL` for ESGF query (#71)
* [feat] Allow `replica` to be `NULL`

* [fix] Remove redundant code
