+++ R/esgf.R
@@ -1,3 +1,19 @@
+RES_DATASET <- c(
+    "id",
+    "mip_era", "activity_drs", "institution_id", "source_id",
+    "experiment_id", "member_id", "table_id", "frequency", "grid_label",
+    "version", "nominal_resolution", "variable_id", "variable_long_name",
+    "variable_units", "data_node", "pid"
+)
+
+RES_FILE <- c(
+    "id", "dataset_id",
+    "mip_era", "activity_drs", "institution_id", "source_id",
+    "experiment_id", "member_id", "table_id", "frequency", "grid_label",
+    "version", "nominal_resolution", "variable_id", "variable_long_name",
+    "variable_units", "data_node", "size", "url", "tracking_id"
+)
+
 # esgf_query {{{
 #' Query CMIP6 data using ESGF search RESTful API
 #'
@@ -200,23 +216,22 @@
 #' @importFrom checkmate assert_character assert_choice assert_count assert_flag assert_subset
 #' @importFrom data.table data.table setattr
 #' @export
-esgf_query <- function (
-    activity = "ScenarioMIP",
-    variable = c("tas", "tasmax", "tasmin", "hurs", "hursmax", "hursmin", "pr", "rsds", "rlds", "psl", "sfcWind", "clt"),
-    frequency = "day",
-    experiment = c("ssp126", "ssp245", "ssp370", "ssp585"),
-    source = c("AWI-CM-1-1-MR", "BCC-CSM2-MR", "CESM2", "CESM2-WACCM",
-               "EC-Earth3", "EC-Earth3-Veg", "GFDL-ESM4", "INM-CM4-8",
-               "INM-CM5-0", "MPI-ESM1-2-HR", "MRI-ESM2-0"),
-    variant = "r1i1p1f1",
-    replica = FALSE,
-    latest = TRUE,
-    resolution = c("100 km", "50 km"),
-    type = "Dataset",
-    limit = 10000L,
-    data_node = NULL
-)
-{
+esgf_query <- function(activity = "ScenarioMIP",
+                       variable = c("tas", "tasmax", "tasmin", "hurs", "hursmax", "hursmin", "pr", "rsds", "rlds", "psl", "sfcWind", "clt"),
+                       frequency = "day",
+                       experiment = c("ssp126", "ssp245", "ssp370", "ssp585"),
+                       source = c(
+                           "AWI-CM-1-1-MR", "BCC-CSM2-MR", "CESM2", "CESM2-WACCM",
+                           "EC-Earth3", "EC-Earth3-Veg", "GFDL-ESM4", "INM-CM4-8",
+                           "INM-CM5-0", "MPI-ESM1-2-HR", "MRI-ESM2-0"
+                       ),
+                       variant = "r1i1p1f1",
+                       replica = FALSE,
+                       latest = TRUE,
+                       resolution = c("100 km", "50 km"),
+                       type = "Dataset",
+                       limit = 10000L,
+                       data_node = NULL) {
     assert_subset(activity, empty.ok = FALSE, choices = c(
         "AerChemMIP", "C4MIP", "CDRMIP", "CFMIP", "CMIP", "CORDEX", "DAMIP",
         "DCPP", "DynVarMIP", "FAFMIP", "GMMIP", "GeoMIP", "HighResMIP",
@@ -240,18 +255,22 @@ esgf_query <- function (
 
     url_base <- "http://esgf-node.llnl.gov/esg-search/search/?"
 
-    dict <- c(activity = "activity_id",
-              experiment = "experiment_id",
-              source = "source_id",
-              variable = "variable_id",
-              resolution = "nominal_resolution",
-              variant = "variant_label")
+    dict <- c(
+        activity = "activity_id",
[…]
