### 2022-10-04 41e92da
[refactor] Change `EPWSHIFTR_ENV` to `this`
