+++ R/epwshiftr-package.R
@@ -25,9 +25,10 @@
 "_PACKAGE"
 
 # package internal environment
-EPWSHIFTR_ENV <- new.env(parent = emptyenv())
-EPWSHIFTR_ENV$index_db <- NULL
-EPWSHIFTR_ENV$dict <- NULL
+this <- new.env(parent = emptyenv())
+this$index_db <- NULL
+this$dict <- NULL
+this$cache <- list()
 
 ## usethis namespace: start
 #' @importFrom checkmate assert_string
