### 2022-10-08 079f16a
[refactor] Use `query_param_encode()` in `esgf_query()`
