### 2022-10-08 c2dcb51
[feat] Add `EsgfQuery` class
[feat] Add `EsgfQuery` class

[feat] Complete `EsgfQuery` class

[feat] Complete `EsgfQuery` class
