+++ R/query.R
@@ -1,0 +1,1371 @@
+new_query_param <- function(name, value) {
+    assert_string(name)
+
+    if (!is.list(value)) value <- list(value = value, negate = FALSE)
+    if (is.null(value$value)) return(NULL)
+
+    assert_list(value,
+        types = c("logical", "numeric", "character"),
+        any.missing = TRUE, names = "unique"
+    )
+    assert_names(names(value), must.include = c("value", "negate"))
+
+    # do not allow NAs
+    for (val in value) checkmate::assert_atomic(val, any.missing = FALSE)
+
+    value$name <- name
+    structure(value, class = c("EsgfQueryParam", "list"))
+}
+
+is.query_param <- function(x) inherits(x, "EsgfQueryParam")
+
+#' @export
+print.EsgfQueryParam <- function(x, encode = FALSE, space = TRUE, ...) {
+    cat(format.EsgfQueryParam(x, encode = encode, space = space), sep = "\n")
+    invisible(x)
+}
+
+format.EsgfQueryParam <- function(x, encode = TRUE, space = FALSE, ...) {
+    if (is.logical(x$value)) {
+        res <- tolower(if (x$negate) !x$value else x$value)
+    } else if (is.numeric(x$value)) {
+        res <- as.character(x$value)
+    } else if (encode) {
+        res <- query_param_encode(x$value)
+    } else {
+        res <- x$value
+    }
+
+    spc <- if (space) " " else ""
+
+    if (x$negate && !is.logical(x$value)) {
+        paste0(x$name, spc, "!=", spc, res, collapse = paste0(spc, "&", spc))
+    } else {
+        paste0(x$name, spc, "=", spc, paste0(res, collapse = paste0(",", spc)))
+    }
+}
+
+#' Query CMIP6 data using ESGF search RESTful API
+#'
+#' @description
+#' The Earth System Grid Federation (ESGF) is an international collaboration for
+#' the software that powers most global climate change research, notably
+#' assessments by the Intergovernmental Panel on Climate Change (IPCC).
+#'
+#' The ESGF search service exposes a RESTful URL that can be used by clients to
+#' query the contents of the underlying search index, and return results
+#' matching the given constraints.
+#'
+#' `EsgfQuery` is the workhorse for dealing with ESGF search services.
+#'
+#' # `EsgfQuery` object
+#'
+#' `query_esgf()` returns an `EsgfQuery` object, which is an [R6][R6::R6Class]
+#' object with quite a few methods that can be classified into 3 categories:
+#'
+#' - **Value listing**: methods to list all possible values of facets, shards,
+#'   etc.
+#'
+#' - **Parameter getter & setter**: methods to get the query parameter values or
+#'   set them before sending the actual query to the ESGF search services.
+#'
+#' - **Query responses**: methods to collect results for the query response.
+#'
+#' ## Value listing
+#'
+#' When creating an `EsgfQuery` object, a
+#' [facet listing query](https://esgf.github.io/esg-search/ESGF_Search_RESTful_API.html#facet-listings)
+#' is sent to the index node to get all available facets and shards.
+#' `EsgfQuery` object provides three value-listing methods to extract data from
+#' the response of the facet listing query:
+#'
+#' - \href{#method-EsgfQuery-list_all_facets}{\code{EsgfQuery$list_all_facets()}}:
+#'   List all available facet names.
+#'
+#' - \href{#method-EsgfQuery-list_all_shards}{\code{EsgfQuery$list_all_shards()}}:
+#'   List all available shards.
+#'
+#' - \href{#method-EsgfQuery-list_all_values}{\code{EsgfQuery$list_all_values()}}:
+#'   List all available values of a specific facet.
+#'
[…]
