+++ DESCRIPTION
@@ -48,7 +48,7 @@ Encoding: UTF-8
 URL: https://github.com/ideas-lab-nus/epwshiftr
 BugReports: https://github.com/ideas-lab-nus/epwshiftr/issues
 Roxygen: list(markdown = TRUE)
-RoxygenNote: 7.2.0
+RoxygenNote: 7.2.1
 Collate:
     'coord.R'
     'dict.R'
@@ -58,5 +58,7 @@ Collate:
     'gh.R'
     'morph.R'
     'netcdf.R'
+    'query.R'
 VignetteBuilder: knitr
 Config/testthat/start-first: utils, esgf, coord, netcdf, morph
+Config/testthat/edition: 3

[…]
