+++ tests/testthat/test-netcdf.R
@@ -6,7 +6,7 @@ test_that("get_nc_meta()", {
     path <- file.path(cache, file)
 
     if (file.exists(path)) {
-        expect_is(meta <- get_nc_meta(path), "list")
+        expect_type(meta <- get_nc_meta(path), "list")
         expect_named(meta,
             c("mip_era", "activity_id", "institution_id", "source_id",
               "experiment_id", "variant_label", "table_id", "grid_label",
@@ -16,7 +16,7 @@ test_that("get_nc_meta()", {
         )
 
         con <- RNetCDF::open.nc(path)
-        expect_equivalent(get_nc_meta(con), meta)
+        expect_equal(get_nc_meta(con), meta, ignore_attr = TRUE)
         RNetCDF::close.nc(con)
     }
 })
@@ -29,13 +29,13 @@ test_that("get_nc_atts()", {
     path <- file.path(cache, file)
 
     if (file.exists(path)) {
-        expect_is(atts <- get_nc_atts(path), "data.table")
+        expect_s3_class(atts <- get_nc_atts(path), "data.table")
         expect_named(atts, c("id", "variable", "attribute", "value"))
         expect_true(all(unique(atts$variable) %in% c("height", "lat", "lon", "NC_GLOBAL", "tas", "time")))
         expect_identical(atts[variable == "NC_GLOBAL", unique(id)], -1L)
 
         con <- RNetCDF::open.nc(path)
-        expect_equivalent(get_nc_atts(con), atts)
+        expect_equal(get_nc_atts(con), atts, ignore_attr = TRUE)
         RNetCDF::close.nc(con)
     }
 })
@@ -48,11 +48,11 @@ test_that("get_nc_vars()", {
     path <- file.path(cache, file)
 
     if (file.exists(path)) {
-        expect_is(vars <- get_nc_vars(path), "data.table")
+        expect_s3_class(vars <- get_nc_vars(path), "data.table")
         expect_named(vars, c("id", "name", "type", "ndims", "natts"))
 
         con <- RNetCDF::open.nc(path)
-        expect_equivalent(get_nc_vars(con), vars)
+        expect_equal(get_nc_vars(con), vars, ignore_attr = TRUE)
         RNetCDF::close.nc(con)
     }
 })
@@ -65,11 +65,11 @@ test_that("get_nc_dims()", {
     path <- file.path(cache, file)
 
     if (file.exists(path)) {
-        expect_is(dims <- get_nc_dims(path), "data.table")
+        expect_s3_class(dims <- get_nc_dims(path), "data.table")
         expect_named(dims, c("id", "name", "length", "unlim"))
 
         con <- RNetCDF::open.nc(path)
-        expect_equivalent(get_nc_dims(con), dims)
+        expect_equal(get_nc_dims(con), dims, ignore_attr = TRUE)
         RNetCDF::close.nc(con)
     }
 })
@@ -82,11 +82,11 @@ test_that("get_nc_axes()", {
     path <- file.path(cache, file)
 
     if (file.exists(path)) {
-        expect_is(axes <- get_nc_axes(path), "data.table")
+        expect_s3_class(axes <- get_nc_axes(path), "data.table")
         expect_named(axes, c("axis", "variable", "dimension"))
 
         con <- RNetCDF::open.nc(path)
-        expect_equivalent(get_nc_axes(con), axes)
+        expect_equal(get_nc_axes(con), axes, ignore_attr = TRUE)
         RNetCDF::close.nc(con)
     }
 })
@@ -99,14 +99,14 @@ test_that("get_nc_time()", {
     path <- file.path(cache, file)
 
     if (file.exists(path)) {
[…]
