### 2022-10-08 8cf1139
[refactor] Rename `CMIP6Dict` to `Cmip6Dict`
