+++ man/CMIP6Dict.Rd
@@ -1,14 +1,14 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/dict.R
-\name{CMIP6Dict}
-\alias{CMIP6Dict}
+\name{Cmip6Dict}
+\alias{Cmip6Dict}
 \alias{cmip6_dict}
 \title{CMIP6 Controlled Vocabularies (CVs) and Data Request Dictionary}
 \usage{
 cmip6_dict()
 }
 \description{
-The \code{CMIP6Dict} object provides functionalities to fetch the latested CMIP6
+The \code{Cmip6Dict} object provides functionalities to fetch the latested CMIP6
 Controlled Vocabularies (CVs) and Data Request (DReq) information.
 }
 \details{
@@ -24,7 +24,7 @@ quantities that are more specialized and only of interest to a single
 endorsed MIP. The raw data of DReq is stored a Microsoft Excel file
 (\code{CMIP6_MIP_tables.xlsx}) in a
 \href{http://proj.badc.rl.ac.uk/svn/exarch/CMIP6dreq/trunk}{Subversion repo}.
-The \code{CMIP6Dict} object uses the parsed DReq data that is stored in the
+The \code{Cmip6Dict} object uses the parsed DReq data that is stored in the
 \href{https://github.com/PCMDI/cmip6-cmor-tables}{GitHub Repo}.
 
 For more information, please see:
@@ -36,10 +36,10 @@ For more information, please see:
 \examples{
 \dontrun{
 
-# create a new CMIP6Dict object
+# create a new Cmip6Dict object
 dict <- cmip6_dict()
 
-# by default, there is no data when the CMIP6Dict was created
+# by default, there is no data when the Cmip6Dict was created
 dict$is_empty()
 
 # fetch and parse all CVs and Data Request data
@@ -85,24 +85,24 @@ Hongyuan Jia
 \section{Methods}{
 \subsection{Public methods}{
 \itemize{
-\item \href{#method-CMIP6Dict-version}{\code{CMIP6Dict$version()}}
-\item \href{#method-CMIP6Dict-is_empty}{\code{CMIP6Dict$is_empty()}}
-\item \href{#method-CMIP6Dict-timestamp}{\code{CMIP6Dict$timestamp()}}
-\item \href{#method-CMIP6Dict-built_time}{\code{CMIP6Dict$built_time()}}
-\item \href{#method-CMIP6Dict-build}{\code{CMIP6Dict$build()}}
-\item \href{#method-CMIP6Dict-get}{\code{CMIP6Dict$get()}}
-\item \href{#method-CMIP6Dict-save}{\code{CMIP6Dict$save()}}
-\item \href{#method-CMIP6Dict-load}{\code{CMIP6Dict$load()}}
-\item \href{#method-CMIP6Dict-print}{\code{CMIP6Dict$print()}}
+\item \href{#method-Cmip6Dict-version}{\code{Cmip6Dict$version()}}
+\item \href{#method-Cmip6Dict-is_empty}{\code{Cmip6Dict$is_empty()}}
+\item \href{#method-Cmip6Dict-timestamp}{\code{Cmip6Dict$timestamp()}}
+\item \href{#method-Cmip6Dict-built_time}{\code{Cmip6Dict$built_time()}}
+\item \href{#method-Cmip6Dict-build}{\code{Cmip6Dict$build()}}
+\item \href{#method-Cmip6Dict-get}{\code{Cmip6Dict$get()}}
+\item \href{#method-Cmip6Dict-save}{\code{Cmip6Dict$save()}}
[…]
