+++ R/epwshiftr-package.R
@@ -30,7 +30,9 @@ this$index_db <- NULL
 this$dict <- NULL
 this$cache <- list()
 
+# nocov start
 attach_cache <- function(cache) this$cache <- cache
+# nocov end
 
 ## usethis namespace: start
 #' @importFrom checkmate assert_string
