### 2022-10-08 b147dd8
[doc] Update NEWS
