### 2022-10-08 cf10d0e
[doc] Update pkgdown entries
