### 2022-10-09 a2a4fed
[fix] Fix other parameter parsing logic
