+++ tests/testthat/helper-cache.R
@@ -13,14 +13,6 @@ get_cache <- function(path = Sys.getenv("EPWSHIFTR_CHECK_CACHE", NA), reset = FA
         eplusr::download_weather("SGP_Singapore.486980_IWEC", dir = cache, type = "epw", ask = FALSE, max_match = 1L)
     }
 
-    # ESGF facet listing query
-    facets <- file.path(cache, "facets")
-    if (!file.exists(facets)) {
-        query_build_facet_cache("https://esgf-node.llnl.gov/esg-search", NULL, TRUE)
-        saveRDS(epwshiftr:::this$cache, facets)
-        attach_cache(list())
-    }
-
     # download NetCDF
     withr::with_options(
         list(epwshiftr.dir = cache),
