### 2022-10-09 0d0ff9c
[fix] Replace localhost when listing Shards
