### 2022-10-09 cc62c23
[feat] Add `EsgfQuery` class to wrap ESGF search RESTful APIs (#73)
