+++ tests/testthat/test-morph.R
@@ -10,23 +10,23 @@ test_that("monthly_mean()", {
 
         epw$add_unit()
         d <- epw$data()
-        expect_is(res1 <- monthly_mean(d, "dry_bulb_temperature"), "data.table")
+        expect_s3_class(res1 <- monthly_mean(d, "dry_bulb_temperature"), "data.table")
         expect_identical(nrow(res1), 12L)
         expect_named(res1, c("month", "val_mean", "val_max", "val_min"))
         expect_identical(res1$month, 1L:12L)
-        expect_is(res1$val_mean, "units")
-        expect_is(res1$val_max, "units")
-        expect_is(res1$val_min, "units")
+        expect_s3_class(res1$val_mean, "units")
+        expect_s3_class(res1$val_max, "units")
+        expect_s3_class(res1$val_min, "units")
 
         epw$drop_unit()
         d <- epw$data()
-        expect_is(res2 <- monthly_mean(d, "dry_bulb_temperature"), "data.table")
+        expect_s3_class(res2 <- monthly_mean(d, "dry_bulb_temperature"), "data.table")
         expect_identical(nrow(res2), 12L)
         expect_named(res2, c("month", "val_mean", "val_max", "val_min"))
         expect_identical(res2$month, 1L:12L)
-        expect_is(res2$val_mean, "numeric")
-        expect_is(res2$val_max, "numeric")
-        expect_is(res2$val_min, "numeric")
+        expect_type(res2$val_mean, "double")
+        expect_type(res2$val_max, "double")
+        expect_type(res2$val_min, "double")
     }
 })
 
@@ -47,12 +47,12 @@ test_that("preprocess_morphing()", {
             )
         )
         expect_identical(nrow(res), 12L)
-        expect_is(res$interval, "factor")
+        expect_s3_class(res$interval, "factor")
 
         # can stop if specified years were not found
         expect_error(preprocess_morphing(d, years = 2059L:2061L), "does not contain any data")
 
-        expect_is(res <- preprocess_morphing(d, years = 2060L, labels = as.factor("2060s")), "data.table")
+        expect_s3_class(res <- preprocess_morphing(d, years = 2060L, labels = as.factor("2060s")), "data.table")
         expect_named(res,
             c("activity_drs", "institution_id", "source_id", "experiment_id",
               "member_id", "table_id", "lon", "lat", "dist", "units", "value", "month",
@@ -60,7 +60,7 @@ test_that("preprocess_morphing()", {
             )
         )
         expect_identical(nrow(res), 12L)
-        expect_is(res$interval, "factor")
+        expect_s3_class(res$interval, "factor")
     }
 })
 
@@ -87,7 +87,7 @@ test_that("morphing_from_mean()", {
         )
 
         # can use morphing using shift factor
-        expect_is(class = "data.table",
+        expect_s3_class(class = "data.table",
             res <- morphing_from_mean(
                 "dry_bulb_temperature",
                 data_epw, data_mean, NULL, NULL,
@@ -102,13 +102,13 @@ test_that("morphing_from_mean()", {
               "delta", "alpha"
             )
         )
-        expect_is(res$dry_bulb_temperature, "units")
-        expect_is(res$delta, "units")
-        expect_is(res$alpha, "units")
+        expect_s3_class(res$dry_bulb_temperature, "units")
+        expect_s3_class(res$delta, "units")
+        expect_s3_class(res$alpha, "units")
         expect_equal(res$dry_bulb_temperature - data_epw$dry_bulb_temperature, res$delta)
 
         # can use morphing using stretch factor
[…]
