+++ tests/testthat/_snaps/query.md
@@ -1,0 +1,40 @@
+# ESGF Query Parameter works
+
+    [1] "x=false"
+
+---
+
+    [1] "x!=1"
+
+---
+
+    [1] "x!=solr%2Bjson"
+
+---
+
+    [1] "x = true"
+
+---
+
+    [1] "x = 1"
+
+---
+
+    [1] "x = solr%2Bjson"
+
+---
+
+    x = false
+
+---
+
+    x != 1
+
+---
+
+    x != solr+json
+
+# ESGF Query works
+
+    
+
