### 2022-10-09 e161ab1
[feat] Add `EsgfQuery` class to wrap ESGF search RESTful APIs (#73)
