+++ NAMESPACE
@@ -2,7 +2,9 @@
 
 S3method(print,CMIP6CV)
 S3method(print,CMIP6DReq)
-export(CMIP6Dict)
+S3method(print,EsgfQueryParam)
+export(Cmip6Dict)
+export(EsgfQuery)
 export(cmip6_dict)
 export(esgf_query)
 export(extract_data)
@@ -13,6 +15,7 @@ export(init_cmip6_index)
 export(load_cmip6_index)
 export(match_coord)
 export(morphing_epw)
+export(query_esgf)
 export(set_cmip6_index)
 export(summary_database)
 importFrom(PCICt,as.PCICt)

[…]
