+++ pkgdown.yml
@@ -4,5 +4,5 @@ pkgdown_sha: ~
 articles:
   cmip6_dict: cmip6_dict.html
   paper: paper.html
-last_built: 2022-09-23T01:37Z
+last_built: 2022-10-09T08:19Z
 
