### 2022-10-09 f29a7c2
[feat] Support wget script URL in `EsgfQuery$url()`
