+++ man/EsgfQuery.Rd
@@ -10,7 +10,7 @@ query_esgf(host = "https://esgf-node.llnl.gov/esg-search")
 \arguments{
 \item{host}{The URL to the ESGF Search API service. This should be the URL of
 the ESGF search service excluding the final endpoint name. Usually
-this is \verb{http://<hostname>/esg-search}. Default is to ses the
+this is \verb{http://<hostname>/esg-search}. Default is set to the
 \href{http://esgf-node.llnl.gov}{LLNL (Lawrence Livermore National Laboratory) Index Node},
 which is \code{"https://esgf-node.llnl.gov/esg-search"}.}
 }
@@ -146,14 +146,15 @@ to the index node and stores the response internally. The response contains
 all available facets and shards and is used as a source for validating user
 input for parameter setters.
 \item \href{#method-EsgfQuery-url}{\code{EsgfQuery$url()}}: Returns the actual
-query URL.
+query URL or the wget script URL which can be used to download all files
+matching the given constraints..
 \item \href{#method-EsgfQuery-response}{\code{EsgfQuery$response()}}: Returns the
-acutal response of
+actual response of
 \href{#method-EsgfQuery-count}{\code{EsgfQuery$count()}} and
 \href{#method-EsgfQuery-collect}{\code{EsgfQuery$collect()}}. It is a named
 list generated from the JSON response using \code{\link[jsonlite:fromJSON]{jsonlite::fromJSON()}}.
 \item \href{#method-EsgfQuery-print}{\code{EsgfQuery$print()}}: Print a summary
-of the current \code{EsgfQuery} object including the host url, the built time of
+of the current \code{EsgfQuery} object including the host URL, the built time of
 facet cache and all query parameters.
 }
 }
@@ -529,6 +530,15 @@ q$params(NULL)$params()
 
 \dontrun{
 q$url()
+
+# get the wget script URL
+q$url(wget = TRUE)
+
+# You can download the wget script using the URL directly. For
+# example, the code below downloads the script and save it as
+# 'wget.sh' in R's temporary folder:
+download.file(q$url(TRUE), file.path(tempdir(), "wget.sh"), mode = "wb")
+
 }
 
 ## ------------------------------------------------
@@ -775,7 +785,7 @@ Get or set the \code{project} facet parameter.
 There are two options:
 \itemize{
 \item If \code{value} is not given, current value is returned.
-\item A character vector or \code{NULL}.  Note that you can put a preceding \code{!} to negate the facet contraints. For example, \code{$project(!c("CMIP5", "CMIP6"))} searches for all \code{project}s except for \code{CMIP5} and \code{CMIP6}.
+\item A character vector or \code{NULL}.  Note that you can put a preceding \code{!} to negate the facet constraints. For example, \code{$project(!c("CMIP5", "CMIP6"))} searches for all \code{project}s except for \code{CMIP5} and \code{CMIP6}.
 }}
 }
 \if{html}{\out{</div>}}
@@ -820,7 +830,7 @@ Get or set the \code{activity_id} facet parameter.
 There are two options:
 \itemize{
 \item If \code{value} is not given, current value is returned.
-\item A character vector or \code{NULL}.  Note that you can put a preceding \code{!} to negate the facet contraints. For example, \code{$activity_id(!c("C4MIP", "GeoMIP"))} searches for all \code{activity_id}s except for \code{C4MIP} and \code{GeoMIP}.
[…]
