### 2022-10-09 375fcb5
[feat] Generate URL of wget download script using ESGF RESTful API (#74)
