+++ NEWS.md
@@ -82,8 +82,8 @@
 * Now HDF5 format is supported (#60).
 * Now `replica` can be `NULL` in `esgf_query()` and `init_cmip6_index()`. In
   this case, both the master record and replicas are all returned (#61).
-* New class `EsgfQuery` is added to support more flexible. Please see
-  `?EsgfQuery` for details (#63).
+* New class `EsgfQuery` is added to support more flexible APIs for ESGF search
+  services. Please see `?EsgfQuery` for details (#63, #69).
 
 ## Bug fixes
 
