### 2022-10-09 b20db2d
[feat] Generate URL of wget download script using ESGF RESTful API (#74)
