+++ R/epwshiftr-package.R
@@ -14,7 +14,7 @@
 #'   i.e. monthly-mean fractional change, when performing morphing operations.
 #'   The default value is set to `3`. If the morphing methods are set
 #'   `"stretch"` or `"combined"`, and the absolute alpha exceeds the threshold
-#'   value, warnings are issued and the morphing method is fallbacked to
+#'   value, warnings are issued and the morphing method fallbacks to
 #'   `"shift"` to avoid unrealistic morphed values.
 #' * `epwshiftr.dir`: The directory to store package data, including CMIP6
 #'   model output file index and etc. If not set, the current user data
