### 2022-10-12 db0bc15
[refactor] Use `is_query_param()` instead of `is.query_param()`
