### 2022-10-12 f28bf8e
[test] Add helper for attach and clear facet cache
