### 2022-10-12 18e7684
[refactor] Make `EsgfQuery` clonable
