### 2022-10-12 b65bc54
[refactor] Return `self` in `EsgfQuery$new()`
