### 2022-10-12 2971ff3
[refactor] Dynamically set max query limit [doc] Update doc of `EsgfQuery$limit()` [doc] Update doc
