### 2022-10-12 65dd6f7
[refactor] Dynamically set max query limit [doc] Update doc of `EsgfQuery$limit()` [doc] Update doc
