### 2022-10-12 af5bc1a
[refactor] Remove `EsgfQuery$type()`
