### 2022-10-12 b5c2c36
[refactor] Remove `EsgfQuery$type()`
