### 2022-10-12 a0b8905
[refactor] Always use `type = Dataset` in `EsgfQuery`
