### 2022-10-12 c1b7f41
[refactor] Add a private method to get all params
