### 2022-10-12 342edc5
[refactor] Keep original params untouched in `EsgfQuery$count()`
