### 2022-10-12 aee1e97
[refactor] Remove duplications when creating new params
