### 2022-10-12 eb34012
[feat] Allow auto-pagination in `EsgfQuery$collect()` [doc]
