### 2022-10-12 5f848ea
[feat] Allow auto-pagination in `EsgfQuery$collect()` [doc]
