### 2022-10-12 e74cb34
[feat] Add `EsgfQueryResult` and subclasses
