+++ R/epwshiftr-package.R
@@ -49,9 +49,20 @@ attach_cache <- function(cache) this$cache <- cache
         "epwshiftr.verbose" = "FALSE",
         "epwshiftr.threshold_alpha" = "3"
     )
-    for (i in setdiff(names(.opts), names(options()))) {
-        eval(parse(text = paste0("options(",i,"=",.opts[i],")")))
+    for (name in setdiff(names(.opts), names(options()))) {
+        eval(parse(text = sprintf("options(%s = %s)", name, .opts[name])))
     }
+
+    # install IEC style Byte units
+    IEC <- c("Byte", "KiB", "MiB", "GiB", "TiB", "PiB", "EiB", "ZiB", "YiB")
+    big <- IEC[-1L]
+    small <- IEC[-length(IEC)]
+    for (i in seq_along(big)) {
+        try(units::install_unit(big[[i]], sprintf("1024 * %s", small[[i]])),
+            silent = TRUE
+        )
+    }
+
     invisible()
 }
 # nocov end
