### 2022-10-12 1d007d4
[feat] Add `EsgfQueryResult` and subclasses
