+++ R/query-result.R
@@ -1,0 +1,293 @@
+EsgfQueryResult <- R6::R6Class("EsgfQueryResult",
+    lock_class = TRUE,
+    public = list(
+        initialize = function(host, result) {
+            private$url_host <- host
+            private$result <- result
+            self
+        },
+        to_dt = function(fields = NULL, formatted = NULL) {
+            docs <- private$get_docs()
+
+            assert_subset(fields, names(docs))
+            assert_character(formatted, any.missing = FALSE, unique = TRUE, null.ok = TRUE)
+
+            if (length(fields)) {
+                docs <- docs[match(fields, names(docs))]
+            }
+            if (!is.null(formatted)) {
+                for (field in formatted) {
+                    if (is.null(docs[[field]])) next
+                    docs[[field]] <- self[[field]]
+                }
+            }
+
+            res <- data.table::setDT(
+                lapply(docs, function(doc) {
+                    if (typeof(doc) == "list" && all(lengths(doc) == 1L)) {
+                        doc <- unlst(doc)
+                    }
+                    doc
+                })
+            )
+
+            res
+        },
+        count = function() {
+            length(self$id)
+        }
+    ),
+
+    active = list(
+        id = function() {
+            private$get_field("id")
+        },
+        url = function() {
+            urls <- private$get_field("url")
+            if (!length(urls)) return(NULL)
+
+            lapply(urls, function(url) {
+                if (!length(url)) return(NULL) # nocov
+
+                s <- strsplit(url, "|", fixed = TRUE)
+                # nocov start
+                if (any(unreco <- lengths(s) != 3L)) s[unreco] <- NULL
+                if (!length(s)) return(NULL)
+                # nocov end
+
+                res <- data.table::setDT(data.table::transpose(s))
+                data.table::setcolorder(res, c(3L, 1L, 2L))
+                data.table::setnames(res, c("service", "url", "mime_type"))
+                res
+            })
+        },
+        size = function() {
+            size <- private$get_field("size")
+            set_size_units(size)
+        },
+        fields = function() {
+            sort(names(private$get_docs()))
+        }
+    ),
+    private = list(
+        url_host = NULL,
+        result = NULL,
+        response = NULL,
+        params = list(),
+        required_fields = c("id", "url", "size"),
+
+        get_docs = function() {
+            private$result$response$docs
+        },
+
+        get_field = function(field) {
+            val <- private$result$response$docs[[field]]
+            if (all(lengths(val) == 1L)) unlst(val) else val
+        },
+
+        has_facet_cache = function() {
+            !is.null(private$url_host) && !is.null(this$cache[[private$url_host]])
+        },
+
+        facet_cache = function() {
+            this$cache[[private$url_host]]
+        },
+
+        validate_fields = function(fields) {
+            if (is.null(fields)) return(NULL)
+            if ("*" %in% fields) fields <- "*"
+            choices <- unlst(private$facet_cache()$responseHeader$params$facet.field)
+            choices <- c("*", choices)
+
+            # add required fields
+            if (!"*" %in% fields && any(miss <- !private$required_fields %in% fields)) {
+                fields <- c(fields, private$required_fields[miss])
+            }
+
+            assert_subset(fields, empty.ok = TRUE, .var.name = "fields", choices = choices)
+            unique(fields)
+        },
+
+        validate_shards = function(shards) {
+            if (is.null(shards)) return(NULL)
+            choices <- query_esgf(private$url_host)$list_all_shards()
+            # suffix should be excluded when query
+            choices <- self$list_all_shards()
+            if (length(choices)) {
+                choices <- gsub("(?<=/solr).+", "", choices, perl = TRUE)
[…]
