+++ man/EsgfQuery.Rd
@@ -1217,7 +1217,9 @@ Get or set the \code{fields} parameter.
 
 By default, all available metadata fields are returned for each
 query. \verb{$facets()} can be used to limit the number of fields returned
-in the query response.
+in the query response. However, the following fields are always
+included in the results:
+\verb{id}, \verb{url}, \verb{size}, \verb{index_node}, \verb{number_of_fields}, \verb{access}.
 \subsection{Usage}{
 \if{html}{\out{<div class="r">}}\preformatted{EsgfQuery$fields(value = "*")}\if{html}{\out{</div>}}
 }
@@ -1707,9 +1709,12 @@ q$count(facets = c("activity_id", "source_id"))
 \subsection{Method \code{collect()}}{
 Send the actual query and fetch the results
 
-\verb{$collect()} sends the actual query to the ESGF search services and
-returns the results in a \link[data.table:data.table]{data.table::data.table}. The columns depend
-on the value of query type and \code{fields} parameter.
+\verb{$collect()} sends the actual query with \strong{\code{type=Dataset}} to the
+ESGF search services and returns the results as an
+\link{EsgfQueryResultDataset} object.
+The columns depend on the value of query type and \code{fields} parameter.
+However, the following fields are always included in the results:
+\verb{id}, \verb{url}, \verb{size}, \verb{index_node}, \verb{number_of_fields}, \verb{access}.
 \subsection{Usage}{
 \if{html}{\out{<div class="r">}}\preformatted{EsgfQuery$collect(all = FALSE, limit = TRUE)}\if{html}{\out{</div>}}
 }
