### 2022-10-13 e81e003
[fix] Fix required fields for `EsgfQueryResultDataset`
