### 2022-10-13 1ba9150
[fix] Fix required fields for `EsgfQueryResultDataset`
