### 2022-10-13 2b53eda
[feat] Include fields with constraints in `EsgfQuery$collect()`
