### 2022-10-13 c1277a9
[feat] Include fields with constraints in `EsgfQuery$collect()`
