### 2022-10-13 49397d8
[refactor] Do not return `self` invisibly in `EsgfQuery`
