### 2022-10-13 f01bd47
[refactor] Do not return `self` invisibly in `EsgfQuery`
