### 2022-10-13 c679411
[doc] Update doc for `EsgfQuery`
