### 2022-10-13 3d1f98c
[doc] Update doc for `EsgfQuery`
