### 2022-10-13 bb4a2f1
[refactor] Set `limit` range in `EsgfQuery$collect()` fix
