### 2022-10-13 05fe16a
[test] Rewrite `EsgfQuery` tests
