### 2022-10-13 f2189ed
[test] Rewrite `EsgfQuery` tests
