+++ tests/testthat/test-query.R
@@ -42,16 +42,39 @@ test_that("ESGF Query Parameter works", {
     ))
 })
 
-test_that("ESGF Query works", {
+test_that("query_esgf()", {
     skip_on_cran()
 
+    attach_facet_cache()
     expect_s3_class(q <- EsgfQuery$new(), "EsgfQuery")
     expect_s3_class(q <- query_esgf(), "EsgfQuery")
+})
 
-    # listing
+test_that("EsgfQuery$list_all_facets()", {
+    skip_on_cran()
+
+    expect_s3_class(q <- query_esgf(), "EsgfQuery")
     expect_type(q$list_all_facets(), "character")
+})
+
+test_that("EsgfQuery$list_all_shards()", {
+    skip_on_cran()
+
+    expect_s3_class(q <- query_esgf(), "EsgfQuery")
     expect_type(q$list_all_shards(), "character")
+})
+
+test_that("EsgfQuery$list_all_values()", {
+    skip_on_cran()
+
+    expect_s3_class(q <- query_esgf(), "EsgfQuery")
     expect_type(q$list_all_values("activity_id"), "character")
+})
+
+test_that("EsgfQuery$project() and other facet methods", {
+    skip_on_cran()
+
+    expect_s3_class(q <- query_esgf(), "EsgfQuery")
 
     # project
     expect_equal(q$project()$value, "CMIP6")
@@ -102,19 +125,34 @@ test_that("ESGF Query works", {
     expect_null(q$data_node())
     expect_equal(q$data_node("esg.lasg.ac.cn")$data_node()$value, "esg.lasg.ac.cn")
     expect_null(q$data_node(NULL)$data_node())
+})
 
-    # facets
+test_that("EsgfQuery$facets()", {
+    skip_on_cran()
+
+    expect_s3_class(q <- query_esgf(), "EsgfQuery")
+
     expect_null(q$facets())
     expect_equal(q$facets(c("activity_id", "source_id"))$facets()$value, c("activity_id", "source_id"))
     expect_null(q$facets(NULL)$facets())
+})
 
-    # fields
+test_that("EsgfQuery$fields()", {
+    skip_on_cran()
+
+    expect_s3_class(q <- query_esgf(), "EsgfQuery")
+
     expect_equal(q$fields()$value, "*")
     expect_equal(q$fields(c("activity_id", "source_id"))$fields()$value, c("activity_id", "source_id"))
     expect_equal(q$fields("*")$fields()$value, "*")
     expect_null(q$fields(NULL)$fields())
+})
 
-    # shards
+test_that("EsgfQuery$shards()", {
+    skip_on_cran()
+
+    expect_s3_class(q <- query_esgf(), "EsgfQuery")
+
     expect_null(q$shards())
     expect_false(q$distrib(FALSE)$distrib()$value)
     expect_error(q$shards("a"), "distrib")
@@ -125,36 +163,62 @@ test_that("ESGF Query works", {
         "esgf-node.llnl.gov:8985/solr"
     )
     expect_null(q$shards(NULL)$shards())
+})
 
-    # replica
+test_that("EsgfQuery$replica()", {
+    skip_on_cran()
+
+    expect_s3_class(q <- query_esgf(), "EsgfQuery")
+
     expect_null(q$replica())
     expect_equal(q$replica(TRUE)$replica()$value, TRUE)
     expect_equal(q$replica(FALSE)$replica()$value, FALSE)
     expect_null(q$replica(NULL)$replica())
+})
 
-    # latest
+test_that("EsgfQuery$latest()", {
+    skip_on_cran()
+
+    expect_s3_class(q <- query_esgf(), "EsgfQuery")
+
[…]
