### 2022-10-13 b30ba05
[test] Correct test descriptions for `Cmip6Dict`
