### 2022-10-14 0e8f84a
[chore] Update min required version of package dependencies (#76)
