### 2022-10-14 505fa0e
[doc] Fix errors in equations
