### 2022-10-15 1f29659
[doc] Correct doc for `EsgfQuery$fields()`
