### 2022-10-15 f86aebf
[doc] Add doc for `EsgfQueryResult` and subclasses
