+++ man/EsgfQueryResultAggregation.Rd
@@ -1,0 +1,52 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/query-result.R
+\name{EsgfQueryResultAggregation}
+\alias{EsgfQueryResultAggregation}
+\title{ESGF Query results for \code{Aggregation} type}
+\description{
+\code{EsgfQueryResultAggregation} is a class that represents query results for
+\code{Aggregation} type from ESGF search RESTful API.
+
+In general, there is no need to create an \code{EsgfQueryResultAggregation} manually.
+Usually, it is created by calling
+\href{#method-EsgfQueryResultDataset-collect}{\code{EsgfQueryResultDataset$collect()}}.
+}
+\author{
+Hongyuan Jia
+}
+\section{Super class}{
+\code{\link[epwshiftr:EsgfQueryResult]{epwshiftr::EsgfQueryResult}} -> \code{EsgfQueryResultAggregation}
+}
+\section{Methods}{
+\subsection{Public methods}{
+\itemize{
+\item \href{#method-EsgfQueryResultAggregation-clone}{\code{EsgfQueryResultAggregation$clone()}}
+}
+}
+\if{html}{\out{
+<details open><summary>Inherited methods</summary>
+<ul>
+<li><span class="pkg-link" data-pkg="epwshiftr" data-topic="EsgfQueryResult" data-id="count"><a href='../../epwshiftr/html/EsgfQueryResult.html#method-EsgfQueryResult-count'><code>epwshiftr::EsgfQueryResult$count()</code></a></span></li>
+<li><span class="pkg-link" data-pkg="epwshiftr" data-topic="EsgfQueryResult" data-id="initialize"><a href='../../epwshiftr/html/EsgfQueryResult.html#method-EsgfQueryResult-initialize'><code>epwshiftr::EsgfQueryResult$initialize()</code></a></span></li>
+<li><span class="pkg-link" data-pkg="epwshiftr" data-topic="EsgfQueryResult" data-id="to_dt"><a href='../../epwshiftr/html/EsgfQueryResult.html#method-EsgfQueryResult-to_dt'><code>epwshiftr::EsgfQueryResult$to_dt()</code></a></span></li>
+</ul>
+</details>
+}}
+\if{html}{\out{<hr>}}
+\if{html}{\out{<a id="method-EsgfQueryResultAggregation-clone"></a>}}
+\if{latex}{\out{\hypertarget{method-EsgfQueryResultAggregation-clone}{}}}
+\subsection{Method \code{clone()}}{
+The objects of this class are cloneable with this method.
+\subsection{Usage}{
+\if{html}{\out{<div class="r">}}\preformatted{EsgfQueryResultAggregation$clone(deep = FALSE)}\if{html}{\out{</div>}}
+}
+
+\subsection{Arguments}{
+\if{html}{\out{<div class="arguments">}}
+\describe{
+\item{\code{deep}}{Whether to make a deep clone.}
+}
+\if{html}{\out{</div>}}
+}
+}
+}
