+++ man/EsgfQueryResultFile.Rd
@@ -1,0 +1,68 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/query-result.R
+\name{EsgfQueryResultFile}
+\alias{EsgfQueryResultFile}
+\title{ESGF Query results for \code{File} type}
+\description{
+\code{EsgfQueryResultFile} is a class that represents query results for
+\code{File} type from ESGF search RESTful API.
+
+In general, there is no need to create an \code{EsgfQueryResultDataset} manually.
+Usually, it is created by calling
+\href{#method-EsgfQueryResultDataset-collect}{\code{EsgfQueryResultDataset$collect()}}.
+}
+\author{
+Hongyuan Jia
+}
+\section{Super class}{
+\code{\link[epwshiftr:EsgfQueryResult]{epwshiftr::EsgfQueryResult}} -> \code{EsgfQueryResultFile}
+}
+\section{Active bindings}{
+\if{html}{\out{<div class="r6-active-bindings">}}
+\describe{
+\item{\code{filename}}{A character vector indicating file names on the
+sever.}
+
+\item{\code{url_opendap}}{A character vector of the OPeNDAP URLs of the
+files.}
+
+\item{\code{url_download}}{A character vector of the download URLs of the
+files.}
+
+\item{\code{fields}}{A character vector indicating all fields in the results.}
+}
+\if{html}{\out{</div>}}
+}
+\section{Methods}{
+\subsection{Public methods}{
+\itemize{
+\item \href{#method-EsgfQueryResultFile-clone}{\code{EsgfQueryResultFile$clone()}}
+}
+}
+\if{html}{\out{
+<details open><summary>Inherited methods</summary>
+<ul>
+<li><span class="pkg-link" data-pkg="epwshiftr" data-topic="EsgfQueryResult" data-id="count"><a href='../../epwshiftr/html/EsgfQueryResult.html#method-EsgfQueryResult-count'><code>epwshiftr::EsgfQueryResult$count()</code></a></span></li>
+<li><span class="pkg-link" data-pkg="epwshiftr" data-topic="EsgfQueryResult" data-id="initialize"><a href='../../epwshiftr/html/EsgfQueryResult.html#method-EsgfQueryResult-initialize'><code>epwshiftr::EsgfQueryResult$initialize()</code></a></span></li>
+<li><span class="pkg-link" data-pkg="epwshiftr" data-topic="EsgfQueryResult" data-id="to_dt"><a href='../../epwshiftr/html/EsgfQueryResult.html#method-EsgfQueryResult-to_dt'><code>epwshiftr::EsgfQueryResult$to_dt()</code></a></span></li>
+</ul>
+</details>
+}}
+\if{html}{\out{<hr>}}
+\if{html}{\out{<a id="method-EsgfQueryResultFile-clone"></a>}}
+\if{latex}{\out{\hypertarget{method-EsgfQueryResultFile-clone}{}}}
+\subsection{Method \code{clone()}}{
+The objects of this class are cloneable with this method.
+\subsection{Usage}{
+\if{html}{\out{<div class="r">}}\preformatted{EsgfQueryResultFile$clone(deep = FALSE)}\if{html}{\out{</div>}}
+}
+
+\subsection{Arguments}{
+\if{html}{\out{<div class="arguments">}}
+\describe{
+\item{\code{deep}}{Whether to make a deep clone.}
+}
+\if{html}{\out{</div>}}
+}
+}
+}
