### 2022-10-15 8bfbe1c
[doc] Add doc for `EsgfQueryResult` and subclasses
