+++ R/query-result.R
@@ -1,11 +1,51 @@
+#' Base class for results for ESGF query
+#'
+#' @description
+#'
+#' `EsgfQueryResult` is a base class that represents basic query results from
+#' ESGF search RESTful API. It defines common fields and methods for results
+#' from all query types, including `Dataset`, `File` and `Aggregation`. Results
+#' from the three types are
+#'
+#' In general, there is no need to create an `EsgfQueryResult` manually.
+#'
+#' @author Hongyuan Jia
+#' @name EsgfQueryResult
+#' @importFrom R6 R6Class
 EsgfQueryResult <- R6::R6Class("EsgfQueryResult",
     lock_class = TRUE,
     public = list(
+        #' @description
+        #' Create a new EsgfQueryResult object
+        #'
+        #' @param host The URL to the ESGF Search API service. This should be
+        #'        the URL of the ESGF search service excluding the final
+        #'        endpoint name. It should be the same as the `host` for an
+        #'        [EsgfQuery] object that collects the query results.
+        #'
+        #' @param result The result of an query response.
+        #'
+        #' @return An `EsgfQueryResult` object.
+        #'
         initialize = function(host, result) {
             private$url_host <- host
             private$result <- result
             self
         },
+
+        #' @description
+        #' Convert the results into a [data.table][data.table::data.table()]
+        #'
+        #' @param fields A character vector indicating the fields to put into
+        #'        the `data.table`. If `NULL`, all fields in the query result
+        #'        will be used. Default: `NULL`.
+        #'
+        #' @param formatted A character vector indicating the fields whose
+        #'        values should be extracted using the corresponding
+        #'        `EsgfQueryResult$FIELD` active binding. Default: `NULL`.
+        #'
+        #' @return A [data.table][data.table::data.table()].
+        #'
         to_dt = function(fields = NULL, formatted = NULL) {
             docs <- private$get_docs()
 
@@ -33,15 +73,32 @@ EsgfQueryResult <- R6::R6Class("EsgfQueryResult",
 
             res
         },
+
+        #' @description
+        #' Count the number of matched records in current result
+        #'
+        #' @return An integer.
+        #'
         count = function() {
             length(self$id)
         }
     ),
 
     active = list(
+        #' @field id A character vector indicating globally unique record
+        #'        identifiers.
         id = function() {
             private$get_field("id")
         },
+
+        #' @field url A list of [data.table][data.table::data.table()] with 3
+        #'        columns:
+        #'        (a) `service` \\[`character`\\]: The service types, e.g.
+        #'        OPENDAP, HTTPServer, etc.;
+        #'        (b) `url` \\[`character`\\]: The actual URLs;
+        #'        (c) `mime_type` \\[`character`\\]: The MIME types indicating
+        #'        the nature and format of the corresponding document of the
+        #'        URLs.
         url = function() {
             urls <- private$get_field("url")
             if (!length(urls)) return(NULL)
@@ -61,14 +118,20 @@ EsgfQueryResult <- R6::R6Class("EsgfQueryResult",
                 res
             })
         },
+
+        #' @field size A vector of [units][units::as_units()] indicating the
+        #'        file sizes.
         size = function() {
             size <- private$get_field("size")
             set_size_units(size)
         },
+
+        #' @field fields A character vector indicating all fields in the results.
         fields = function() {
             sort(names(private$get_docs()))
         }
     ),
+
     private = list(
         url_host = NULL,
[…]
