+++ man/EsgfQueryResultDataset.Rd
@@ -1,0 +1,179 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/query-result.R
+\name{EsgfQueryResultDataset}
+\alias{EsgfQueryResultDataset}
+\title{ESGF Query results for \code{Dataset} type}
+\description{
+\code{EsgfQueryResultDataset} is a class that represents query results for
+\code{Dataset} type from ESGF search RESTful API.
+
+In general, there is no need to create an \code{EsgfQueryResultDataset} manually.
+Usually, it is created by calling
+\href{#method-EsgfQuery-collect}{\code{EsgfQuery$collect()}}.
+}
+\author{
+Hongyuan Jia
+}
+\section{Super class}{
+\code{\link[epwshiftr:EsgfQueryResult]{epwshiftr::EsgfQueryResult}} -> \code{EsgfQueryResultDataset}
+}
+\section{Methods}{
+\subsection{Public methods}{
+\itemize{
+\item \href{#method-EsgfQueryResultDataset-to_dt}{\code{EsgfQueryResultDataset$to_dt()}}
+\item \href{#method-EsgfQueryResultDataset-has_opendap}{\code{EsgfQueryResultDataset$has_opendap()}}
+\item \href{#method-EsgfQueryResultDataset-has_download}{\code{EsgfQueryResultDataset$has_download()}}
+\item \href{#method-EsgfQueryResultDataset-collect}{\code{EsgfQueryResultDataset$collect()}}
+\item \href{#method-EsgfQueryResultDataset-print}{\code{EsgfQueryResultDataset$print()}}
+\item \href{#method-EsgfQueryResultDataset-clone}{\code{EsgfQueryResultDataset$clone()}}
+}
+}
+\if{html}{\out{
+<details open><summary>Inherited methods</summary>
+<ul>
+<li><span class="pkg-link" data-pkg="epwshiftr" data-topic="EsgfQueryResult" data-id="count"><a href='../../epwshiftr/html/EsgfQueryResult.html#method-EsgfQueryResult-count'><code>epwshiftr::EsgfQueryResult$count()</code></a></span></li>
+<li><span class="pkg-link" data-pkg="epwshiftr" data-topic="EsgfQueryResult" data-id="initialize"><a href='../../epwshiftr/html/EsgfQueryResult.html#method-EsgfQueryResult-initialize'><code>epwshiftr::EsgfQueryResult$initialize()</code></a></span></li>
+</ul>
+</details>
+}}
+\if{html}{\out{<hr>}}
+\if{html}{\out{<a id="method-EsgfQueryResultDataset-to_dt"></a>}}
+\if{latex}{\out{\hypertarget{method-EsgfQueryResultDataset-to_dt}{}}}
+\subsection{Method \code{to_dt()}}{
+Convert the results into a \link[data.table:data.table]{data.table}
+\subsection{Usage}{
+\if{html}{\out{<div class="r">}}\preformatted{EsgfQueryResultDataset$to_dt(fields = NULL, formatted = FALSE)}\if{html}{\out{</div>}}
+}
+
+\subsection{Arguments}{
+\if{html}{\out{<div class="arguments">}}
+\describe{
+\item{\code{fields}}{A character vector indicating the fields to put into
+the \code{data.table}. If \code{NULL}, all fields in the query result
+will be used. Default: \code{NULL}.}
+
+\item{\code{formatted}}{Whether to use formatted values for special fields,
+including \code{url} and \code{size}. Default: \code{FALSE}.}
+}
+\if{html}{\out{</div>}}
+}
+\subsection{Returns}{
+A \link[data.table:data.table]{data.table}.
+}
+}
+\if{html}{\out{<hr>}}
[…]

[…]
