+++ R/epwshiftr-package.R
@@ -22,6 +22,11 @@
 #'
 #' @include utils.R
 #' @author Hongyuan Jia
+## usethis namespace: start
+#' @importFrom checkmate assert_string
+#' @importFrom cli cli_rule
+#' @importFrom jsonlite parse_json
+## usethis namespace: end
 "_PACKAGE"
 
 # package internal environment
@@ -33,36 +38,4 @@ this$data_max_limit <- 10000L
 
 # nocov start
 attach_cache <- function(cache) this$cache <- cache
-# nocov end
-
-## usethis namespace: start
-#' @importFrom checkmate assert_string
-#' @importFrom cli cli_rule
-#' @importFrom jsonlite parse_json
-## usethis namespace: end
-
-# nocov start
-# set package options
-# reference: https://github.com/Rdatatable/data.table/blob/master/R/onLoad.R
-.onLoad <- function (libname, pkgname) {
-    .opts  <-  c(
-        "epwshiftr.verbose" = "FALSE",
-        "epwshiftr.threshold_alpha" = "3"
-    )
-    for (name in setdiff(names(.opts), names(options()))) {
-        eval(parse(text = sprintf("options(%s = %s)", name, .opts[name])))
-    }
-
-    # install IEC style Byte units
-    IEC <- c("Byte", "KiB", "MiB", "GiB", "TiB", "PiB", "EiB", "ZiB", "YiB")
-    big <- IEC[-1L]
-    small <- IEC[-length(IEC)]
-    for (i in seq_along(big)) {
-        try(units::install_unit(big[[i]], sprintf("1024 * %s", small[[i]])),
-            silent = TRUE
-        )
-    }
-
-    invisible()
-}
 # nocov end
