### 2022-10-15 dc6e779
[doc] Fix typos in `EsgfQuery$collect()`
