### 2022-10-15 eb7909d
[feat] Add `all` and `limit` parameter in `EsgfQueryResultDataset$collect()`
