### 2022-10-15 01fa489
[feat] Add `all` and `limit` parameter in `EsgfQueryResultDataset$collect()`
