### 2022-10-16 d2ed038
[fix] Support 0 bytes in `set_size_units()`
