### 2022-10-16 c4f009f
[refactor] Avoid using `cli::cli_warn()` and `cli::cli_abort()`
