+++ R/dict.R
@@ -216,7 +216,7 @@ Cmip6Dict <- R6::R6Class("Cmip6Dict",
         #' @return A single string giving the full path of the RDS file.
         save = function(dir = getOption("epwshiftr.dir", ".")) {
             if (self$is_empty()) {
-                cli::cli_warn(c("!" = "Saving an empty CMIP6 Dictionary object. You may want to build the dictionary first by running `$build()` and call `$save()` again."))
+                warning("Saving an empty CMIP6 Dictionary object. You may want to build the dictionary first by running `$build()` and call `$save()` again.")
             }
 
             cmip6dict_save(
@@ -939,7 +939,7 @@ cmip6dict_load <- function(dir = getOption("epwshiftr.dir", ".")) {
     val <- readRDS(path)
 
     if (!identical(names(val), c("cvs", "dreq", "built_time"))) {
-        cli::cli_abort(c(x = "Malformed format of {.file CMIP6DICT} found."))
+        stop(cli::format_inline("Malformed format of {.file CMIP6DICT} found."))
     }
 
     for (nm in names(val$cvs)) {
