### 2022-10-16 286f1a5
[refactor] Refactor `EsgfQueryResult` and its subclasses
