+++ R/query-result.R
@@ -64,8 +64,12 @@ EsgfQueryResult <- R6::R6Class("EsgfQueryResult",
 
             res <- data.table::setDT(
                 lapply(docs, function(doc) {
-                    if (typeof(doc) == "list" && all(lengths(doc) == 1L)) {
-                        doc <- unlst(doc)
+                    if (typeof(doc) == "list") {
+                        len <- lengths(doc)
+                        if (all(len <= 1L)) {
+                            doc[len == 0L] <- list(NA)
+                            doc <- unlst(doc)
+                        }
                     }
                     doc
                 })
@@ -128,6 +132,7 @@ EsgfQueryResult <- R6::R6Class("EsgfQueryResult",
 
         #' @field fields A character vector indicating all fields in the results.
         fields = function() {
+            if (!self$count()) return(NULL)
             sort(names(private$get_docs()))
         }
     ),
@@ -135,9 +140,8 @@ EsgfQueryResult <- R6::R6Class("EsgfQueryResult",
     private = list(
         url_host = NULL,
         result = NULL,
-        response = NULL,
-        params = list(),
-        required_fields = c("id", "url", "size"),
+
+        required_fields = c("id", "size", "url"),
 
         get_docs = function() {
             private$result$response$docs
@@ -148,56 +152,25 @@ EsgfQueryResult <- R6::R6Class("EsgfQueryResult",
             if (all(lengths(val) == 1L)) unlst(val) else val
         },
 
-        has_facet_cache = function() {
-            !is.null(private$url_host) && !is.null(this$cache[[private$url_host]])
-        },
-
         facet_cache = function() {
             this$cache[[private$url_host]]
         },
 
-        validate_fields = function(fields) {
-            if (is.null(fields)) return(NULL)
-            if ("*" %in% fields) fields <- "*"
-            choices <- unlst(private$facet_cache()$responseHeader$params$facet.field)
-            choices <- c("*", choices)
+        get_url = function(type, name = type) {
+            vapply(self$url, function(dt_url) {
+                if (!length(dt_url)) return(NA_character_)
 
-            # add required fields
-            if (!"*" %in% fields && any(miss <- !private$required_fields %in% fields)) {
-                fields <- c(fields, private$required_fields[miss])
-            }
+                res <- dt_url$url[dt_url$service == type]
+                if (!length(res)) return(NA_character_)
+                # nocov start
+                if (length(res) > 1L) {
+                    warning(sprintf("Multiple %s URL found. Only the first is returned.", name))
+                    res <- res[[1L]]
+                }
+                # nocov end
 
-            assert_subset(fields, empty.ok = TRUE, .var.name = "fields", choices = choices)
-            unique(fields)
-        },
-
-        validate_shards = function(shards) {
-            if (is.null(shards)) return(NULL)
-            choices <- query_esgf(private$url_host)$list_all_shards()
-            if (length(choices)) {
-                choices <- gsub("(?<=/solr).+", "", choices, perl = TRUE)
-            }
-            assert_subset(shards, empty.ok = TRUE, .var.name = "shards", choices = choices)
-            unique(shards)
-        },
-
-        build_params = function(fields = NULL, shards = NULL, replica = NULL, latest = TRUE, type = "File") {
-            assert_flag(replica, null.ok = TRUE)
-            assert_flag(latest)
-            assert_choice(type, c("File", "Aggregation"))
-            fields <- private$validate_fields(fields)
-            shards <- private$validate_shards(shards)
-
-            private$params <- list(
-                dataset_id = self$id,
-                fields = fields,
-                shards = shards,
-                replica = replica,
-                latest = latest,
-                type = type,
-                format = "application/solr+json"
-            )
-            private$params
+                res
+            }, character(1L))
[…]
