### 2022-10-16 b6c6960
[refactor] Refactor `EsgfQueryResult` and its subclasses
