### 2022-10-16 0719f9c
[test] Add tests for `EsgfQueryResult` and its subclasses
