+++ tests/testthat/test-query-result.R
@@ -1,0 +1,203 @@
+test_that("ESGF Query Result Dataset works", {
+    skip_on_cran()
+
+    attach_facet_cache()
+    q <- query_esgf()$
+        activity_id("ScenarioMIP")$
+        source_id("AWI-CM-1-1-MR")$
+        frequency("day")$
+        variable_id("tas")$
+        experiment_id("ssp585")$
+        variant_label("r1i1p1f1")$
+        fields(c("source_id", "experiment_id", "frequency"))$
+        limit(2)
+
+    # can create a new result dataset from EsgfQuery$collect
+    expect_s3_class(datasets <- q$collect(), "EsgfQueryResultDataset")
+
+    # $to_dt(): can extract the data into a data.table
+    expect_s3_class(datasets$to_dt(), "data.table")
+    # $to_dt(): can only keep selected fields
+    expect_s3_class(datasets$to_dt(c("source_id", "frequency")), "data.table")
+    expect_equal(names(datasets$to_dt(c("source_id", "frequency"))), c("source_id", "frequency"))
+    # $to_dt(): can keep the special format of certain fields
+    expect_s3_class(datasets$to_dt(formatted = TRUE)$size, "units")
+    expect_s3_class(datasets$to_dt(formatted = TRUE)$url[[1L]], "data.table")
+
+    # $has_opendap(): can test accessibility
+    expect_type(datasets$has_opendap(), "logical")
+    expect_length(datasets$has_opendap(), 2L)
+
+    # $has_download(): can test accessibility
+    expect_type(datasets$has_download(), "logical")
+    expect_length(datasets$has_download(), 2L)
+
+    # $count(): can get length
+    expect_equal(datasets$count(), 2L)
+
+    # $field: can add active bindings for all fields
+    expect_equal(
+        sort(datasets$fields),
+        c(
+            "access", "activity_id", "experiment_id", "frequency", "id",
+            "index_node", "number_of_aggregations", "number_of_files",
+            "project", "size", "source_id", "url", "variable_id",
+            "variant_label"
+        )
+    )
+
+    # $id
+    expect_type(datasets$id, "character")
+    expect_length(datasets$id, 2L)
+
+    # $url
+    expect_type(datasets$url, "list")
+    expect_length(datasets$url, 2L)
+    expect_s3_class(datasets$url[[1L]], "data.table")
+
+    # $size
+    expect_s3_class(datasets$size, "units")
+    expect_length(datasets$size, 2L)
+
+    # $index_node
+    expect_type(datasets$index_node, "character")
+    expect_length(datasets$index_node, 2L)
+
+    # $access
+    expect_type(datasets$access, "list")
+    expect_length(datasets$access, 2L)
+    expect_type(datasets$access[[1]], "character")
+
+    # $number_of_aggregations
+    expect_type(datasets$number_of_aggregations, "integer")
+    expect_length(datasets$number_of_aggregations, 2L)
+
+    # $number_of_files
+    expect_type(datasets$number_of_files, "integer")
+    expect_length(datasets$number_of_files, 2L)
+
+    # $collect():
+    ## $collect(): can limit fields and record number
+    expect_s3_class(files <- datasets$collect(fields = "id", limit = 1), "EsgfQueryResultFile")
+    expect_length(files$fields, 12L)
+    ## $collect(): can collect all fields
+    expect_s3_class(files <- datasets$collect(fields = "id", all = TRUE), "EsgfQueryResultFile")
+    expect_equal(files$count(), sum(datasets$number_of_files))
+    ## $collect(): can specify specific parameters
[…]
