### 2022-10-17 fefd71e
[test] Add tests for utils
