### 2022-10-17 6b8fb91
[fix] Fix the error when `type` appears in `EsgfQuery$params()`
