### 2022-10-17 b5214a6
[test] Do not use snapshot for `EsgfQuery$print()`
This is because `EsgfQuery$print()` will print the built time of the facet cache.
