### 2022-10-17 ea67342
[test] Add tests for `EsgfQuery$build_cache()`
