+++ R/esgf.R
@@ -211,9 +211,6 @@ RES_FILE <- c(
 #' esgf_query(variable = "rss", experiment = "ssp126", type = "File", limit = 1)
 #' }
 #'
-#' @importFrom jsonlite read_json
-#' @importFrom checkmate assert_character assert_choice assert_count assert_flag assert_subset
-#' @importFrom data.table data.table setattr
 #' @export
 esgf_query <- function(activity = "ScenarioMIP",
                        variable = c("tas", "tasmax", "tasmin", "hurs", "hursmax", "hursmin", "pr", "rsds", "rlds", "psl", "sfcWind", "clt"),
@@ -231,26 +228,26 @@ esgf_query <- function(activity = "ScenarioMIP",
                        type = "Dataset",
                        limit = 10000L,
                        data_node = NULL) {
-    assert_subset(activity, empty.ok = FALSE, choices = c(
+    checkmate::assert_subset(activity, empty.ok = FALSE, choices = c(
         "AerChemMIP", "C4MIP", "CDRMIP", "CFMIP", "CMIP", "CORDEX", "DAMIP",
         "DCPP", "DynVarMIP", "FAFMIP", "GMMIP", "GeoMIP", "HighResMIP",
         "ISMIP6", "LS3MIP", "LUMIP", "OMIP", "PAMIP", "PMIP", "RFMIP", "SIMIP",
         "ScenarioMIP", "VIACSAB", "VolMIP"
     ))
-    assert_character(variable, any.missing = FALSE, null.ok = TRUE)
-    assert_subset(frequency, empty.ok = TRUE, choices = c(
+    checkmate::assert_character(variable, any.missing = FALSE, null.ok = TRUE)
+    checkmate::assert_subset(frequency, empty.ok = TRUE, choices = c(
         "1hr", "1hrCM", "1hrPt", "3hr", "3hrPt", "6hr", "6hrPt", "day", "dec",
         "fx", "mon", "monC", "monPt", "subhrPt", "yr", "yrPt"
     ))
-    assert_character(experiment, any.missing = FALSE, null.ok = TRUE)
-    assert_character(source, any.missing = FALSE, null.ok = TRUE)
-    assert_character(variant, any.missing = FALSE, pattern = "r\\d+i\\d+p\\d+f\\d+", null.ok = TRUE)
-    assert_character(resolution, any.missing = FALSE, null.ok = TRUE)
-    assert_flag(replica, null.ok = TRUE)
-    assert_flag(latest)
-    assert_count(limit, positive = TRUE)
-    assert_choice(type, choices = c("Dataset", "File"))
-    assert_character(data_node, any.missing = FALSE, null.ok = TRUE)
+    checkmate::assert_character(experiment, any.missing = FALSE, null.ok = TRUE)
+    checkmate::assert_character(source, any.missing = FALSE, null.ok = TRUE)
+    checkmate::assert_character(variant, any.missing = FALSE, pattern = "r\\d+i\\d+p\\d+f\\d+", null.ok = TRUE)
+    checkmate::assert_character(resolution, any.missing = FALSE, null.ok = TRUE)
+    checkmate::assert_flag(replica, null.ok = TRUE)
+    checkmate::assert_flag(latest)
+    checkmate::assert_count(limit, positive = TRUE)
+    checkmate::assert_choice(type, choices = c("Dataset", "File"))
+    checkmate::assert_character(data_node, any.missing = FALSE, null.ok = TRUE)
 
     url_base <- "http://esgf-node.llnl.gov/esg-search/search/?"
 
@@ -328,7 +325,7 @@ esgf_query <- function(activity = "ScenarioMIP",
         pair(fields) %and%
         pair(format)
 
-    q <- tryCatch(jsonlite::read_json(q), warning = function(w) w, error = function(e) e)
[…]
