+++ NAMESPACE
@@ -21,26 +21,9 @@ export(summary_database)
 importFrom(PCICt,as.PCICt)
 importFrom(R6,R6Class)
 importFrom(RNetCDF,utcal.nc)
-importFrom(checkmate,assert_character)
-importFrom(checkmate,assert_choice)
-importFrom(checkmate,assert_class)
 importFrom(checkmate,assert_count)
-importFrom(checkmate,assert_data_table)
-importFrom(checkmate,assert_directory_exists)
-importFrom(checkmate,assert_flag)
-importFrom(checkmate,assert_integerish)
-importFrom(checkmate,assert_list)
-importFrom(checkmate,assert_names)
-importFrom(checkmate,assert_number)
-importFrom(checkmate,assert_scalar)
-importFrom(checkmate,assert_string)
-importFrom(checkmate,assert_subset)
-importFrom(checkmate,test_directory_exists)
-importFrom(checkmate,test_file_exists)
-importFrom(checkmate,test_r6)
 importFrom(cli,cli_rule)
 importFrom(data.table,":=")
-importFrom(data.table,as.data.table)
 importFrom(data.table,copy)
 importFrom(data.table,data.table)
 importFrom(data.table,fifelse)
@@ -48,17 +31,13 @@ importFrom(data.table,fread)
 importFrom(data.table,fwrite)
 importFrom(data.table,rbindlist)
 importFrom(data.table,set)
-importFrom(data.table,setDT)
-importFrom(data.table,setattr)
 importFrom(data.table,setcolorder)
 importFrom(data.table,setnames)
 importFrom(data.table,tstrsplit)
 importFrom(eplusr,read_epw)
 importFrom(fst,write_fst)
-importFrom(future.apply,future_Map)
 importFrom(future.apply,future_lapply)
-importFrom(jsonlite,parse_json)
-importFrom(jsonlite,read_json)
+importFrom(jsonlite,fromJSON)
 importFrom(progressr,with_progress)
 importFrom(psychrolib,GetTDewPointFromRelHum)
 importFrom(psychrolib,SetUnitSystem)
