### 2022-11-01 4e3c4e3
[fix] Fix errors caused by using `jsonlite::fromJSON()`
Unlike `jsonlite::read_json()`, `jsonlite::fromJSON()` by default will coerce nested list into a data.frame.
