+++ R/zzz.R
@@ -1,6 +1,6 @@
 # nocov start
 # reference: https://github.com/Rdatatable/data.table/blob/master/R/onLoad.R
-.onLoad <- function (libname, pkgname) {
+.onLoad <- function(libname, pkgname) {
     # set package options
     .opts  <-  c(
         "epwshiftr.verbose" = "FALSE",
