### 2022-11-01 f68e971
[fix] Fix errors caused by using `jsonlite::fromJSON()`
Unlike `jsonlite::read_json()`, `jsonlite::fromJSON()` by default will coerce nested list into a data.frame.
