### 2022-11-01 8f12a88
[test] Add more tests for query results
