### 2022-11-01 3e732dc
[test] Add more tests for query results
