+++ tests/testthat.R
@@ -2,3 +2,5 @@ library(testthat)
 library(epwshiftr)
 
 test_check("epwshiftr")
+
+# TODO: update the vignette
