### 2023-01-25 ed6be36
[feat] support query without cache
