+++ R/query.R
@@ -1,7 +1,7 @@
 #' @include query-result.R
 NULL
 
-# TODO: update doc about tye query type
+# TODO: update doc about query type
 # TODO: collect all results with auto-pagination
 # TODO: automatically add fields that are specified via methods
 read_json_response <- function(url, ...) {
@@ -233,9 +233,11 @@ format.EsgfQueryParam <- function(x, encode = TRUE, space = FALSE, ...) {
 #'        [LLNL (Lawrence Livermore National Laboratory) Index Node](http://esgf-node.llnl.gov),
 #'        which is `"https://esgf-node.llnl.gov/esg-search"`.
 #'
+#' @param build Whether to send the facet listing query. Default: `TRUE`.
+#'
 #' @export
-query_esgf <- function(host = "https://esgf-node.llnl.gov/esg-search") {
-    EsgfQuery$new(host = host)
+query_esgf <- function(host = "https://esgf-node.llnl.gov/esg-search", build = TRUE) {
+    EsgfQuery$new(host = host, build = build)
 }
 
 #' @name EsgfQuery
@@ -246,11 +248,12 @@ EsgfQuery <- R6::R6Class("EsgfQuery",
         #' @description
         #' Create a new EsgfQuery object
         #'
-        #' When initialization, a
+        #' When initialization, by default, a
         #' [facet listing query](https://esgf.github.io/esg-search/ESGF_Search_RESTful_API.html#facet-listings)
         #' is sent to the index node to get all available facets and shards.
         #' This information will be used to validate inputs for `activity_id`,
-        #' `scource_id` facets and etc.
+        #' `scource_id` facets and etc. You can disable sending this query by
+        #' setting `build` to `FALSE`.
         #'
         #' @param host The URL to the ESGF Search API service. This should be
         #'        the URL of the ESGF search service excluding the final
@@ -259,6 +262,8 @@ EsgfQuery <- R6::R6Class("EsgfQuery",
         #'        Laboratory)](http://esgf-node.llnl.gov) Index Node, which is
         #'        `"https://esgf-node.llnl.gov/esg-search"`.
         #'
+        #' @param build Whether to send the facet listing query. Default: `TRUE`.
+        #'
         #' @return An `EsgfQuery` object.
         #'
         #' @examples
@@ -266,17 +271,22 @@ EsgfQuery <- R6::R6Class("EsgfQuery",
         #' q <- EsgfQuery$new(host = "https://esgf-node.llnl.gov/esg-search")
         #' q
         #' }
-        initialize = function(host = "https://esgf-node.llnl.gov/esg-search") {
+        initialize = function(host = "https://esgf-node.llnl.gov/esg-search", build = TRUE) {
             checkmate::assert_string(host)
+            checkmate::assert_flag(build)
+
             # in case of a encoded URL
             private$url_host <- utils::URLdecode(host)
+            if (grepl("/$", private$url_host)) {
+                private$url_host <- substring(private$url_host, 1L, nchar(private$url_host) - 1L)
+            }
 
             if (private$has_facet_cache()) {
                 vb(cli::cli_alert_success(paste(
                     "Loaded facet query cache for host {.var {private$url_host}}",
                     "built at {.var {private$facet_cache()$timestamp}}."
                 )))
-            } else {
+            } else if (build) {
                 self$build_cache()
             }
 
@@ -326,39 +336,45 @@ EsgfQuery <- R6::R6Class("EsgfQuery",
         #' @description
         #' List all available facet names
         #'
-        #' @return A character vector.
+        #' @return A character vector or `NULL` if no facet cache is found.
         #'
         #' @examples
         #' \dontrun{
         #' q$list_all_facets()
         #' }
         list_all_facets = function() {
+            if (!private$has_facet_cache()) return(NULL)
+
[…]
