+++ tests/testthat/test-query.R
@@ -1,3 +1,5 @@
+host <- "https://esgf.ceda.ac.uk/esg-search/"
+
 test_that("ESGF Query Parameter works", {
     # can create new query parameter
     expect_s3_class(
@@ -5,23 +7,22 @@ test_that("ESGF Query Parameter works", {
         "EsgfQueryParam"
     )
 
-    expect_snapshot_output(format(new_query_param("x", list(value = TRUE, negate = TRUE))))
-    expect_snapshot_output(format(new_query_param("x", list(value = 1.0, negate = TRUE))))
+    expect_snapshot(format(new_query_param("x", list(value = TRUE, negate = TRUE))))
+    expect_snapshot(format(new_query_param("x", list(value = 1.0, negate = TRUE))))
     # can encode parameter
-    expect_snapshot_output(format(new_query_param("x", list(value = "solr+json", negate = TRUE))))
+    expect_snapshot(format(new_query_param("x", list(value = "solr+json", negate = TRUE))))
 
     # can add spaces
-    expect_snapshot_output(format(new_query_param("x", TRUE), space = TRUE))
-    expect_snapshot_output(format(new_query_param("x", 1.0), space = TRUE))
-    expect_snapshot_output(format(new_query_param("x", "solr+json"), space = TRUE))
+    expect_snapshot(format(new_query_param("x", TRUE), space = TRUE))
+    expect_snapshot(format(new_query_param("x", 1.0), space = TRUE))
+    expect_snapshot(format(new_query_param("x", "solr+json"), space = TRUE))
 
     # can print query parameter
-    expect_snapshot_output(print(new_query_param("x", list(value = TRUE, negate = TRUE))))
-    expect_snapshot_output(print(new_query_param("x", list(value = 1.0, negate = TRUE))))
-    expect_snapshot_output(print(new_query_param("x", list(value = "solr+json", negate = TRUE))))
+    expect_snapshot(print(new_query_param("x", list(value = TRUE, negate = TRUE))))
+    expect_snapshot(print(new_query_param("x", list(value = 1.0, negate = TRUE))))
+    expect_snapshot(print(new_query_param("x", list(value = "solr+json", negate = TRUE))))
 
     # can build query url
-    host <- "https://esgf-node.llnl.gov/esg-search"
     expect_null(query_build(host, list(project = NULL)))
     expect_true(grepl("CMIP5", query_build(host, list(project = "CMIP6", others = list(project = "CMIP5")))))
     expect_true(grepl(
@@ -43,44 +44,107 @@ test_that("ESGF Query Parameter works", {
 })
 
 test_that("query_esgf()", {
+    expect_s3_class(EsgfQuery$new(build = FALSE), "EsgfQuery")
+    expect_s3_class(query_esgf(build = FALSE), "EsgfQuery")
+
     skip_on_cran()
 
-    expect_s3_class(q <- EsgfQuery$new(), "EsgfQuery")
-    expect_s3_class(q <- query_esgf(), "EsgfQuery")
+    expect_s3_class(q <- EsgfQuery$new(host), "EsgfQuery")
+    expect_s3_class(q <- query_esgf(host), "EsgfQuery")
 })
 
 test_that("EsgfQuery$list_all_facets()", {
+    expect_null(query_esgf(build = FALSE)$list_all_facets())
+
     skip_on_cran()
 
-    expect_s3_class(q <- query_esgf(), "EsgfQuery")
+    expect_s3_class(q <- query_esgf(host), "EsgfQuery")
[…]

[…]
