### 2023-01-28 69c9dfc
[test] fix EsgfQuery tests
