### 2023-01-28 29c6e41
[test] fix EsgfQuery tests
