### 2023-01-29 f127108
[refactor] stop by default if no json response
