### 2023-01-29 909497c
[refactor] Use cli_bullets in `CMIP6Dict$print()`
