### 2023-01-29 06df010
[refactor] Use cli_bullets in `CMIP6Dict$print()`
