+++ tests/testthat/test-dict.R
@@ -85,6 +85,10 @@ test_that("Cmip6Dict$built_time()", {
 test_that("Cmip6Dict$get()", {
     skip_on_cran()
 
+    trans <- function(out) {
+        gsub("^\\* (.+) Modified:.+$", "* \\1 Modified: [yyyy-mm-dd HH:MM:SS UTC]", out)
+    }
+
     dict <- cmip6_dict()
 
     expect_null(dict$get("activity_id"))
@@ -96,14 +100,14 @@ test_that("Cmip6Dict$get()", {
         unique(vapply(dict$get("drs"), typeof, "", USE.NAMES = FALSE)),
         "character"
     )
-    expect_message(print(dict$get("drs")))
+    expect_snapshot(print(dict$get("drs")), transform = trans)
 
     expect_s3_class(dict$get("activity_id"), "list")
     expect_identical(
         unique(vapply(dict$get("activity_id"), typeof, "", USE.NAMES = FALSE)),
         "character"
     )
-    expect_message(print(dict$get("activity_id")))
+    expect_snapshot(print(dict$get("activity_id")), transform = trans)
 
     expect_s3_class(dict$get("experiment_id"), "data.table")
     expect_identical(
@@ -117,45 +121,45 @@ test_that("Cmip6Dict$get()", {
             additional_allowed_model_components = "list"
         )
     )
-    expect_message(print(dict$get("experiment_id")))
+    expect_snapshot(print(dict$get("experiment_id")), transform = trans)
 
     expect_s3_class(dict$get("frequency"), "list")
     expect_identical(
         unique(vapply(dict$get("frequency"), typeof, "", USE.NAMES = FALSE)),
         "character"
     )
-    expect_message(print(dict$get("frequency")))
+    expect_snapshot(print(dict$get("frequency")), transform = trans)
 
     expect_s3_class(dict$get("grid_label"), "list")
     expect_identical(
         unique(vapply(dict$get("grid_label"), typeof, "", USE.NAMES = FALSE)),
         "character"
     )
-    expect_message(print(dict$get("grid_label")))
+    expect_snapshot(print(dict$get("grid_label")), transform = trans)
 
     expect_s3_class(dict$get("institution_id"), "list")
     expect_identical(
         unique(vapply(dict$get("institution_id"), typeof, "", USE.NAMES = FALSE)),
         "character"
     )
-    expect_message(print(dict$get("institution_id")))
+    expect_snapshot(print(dict$get("institution_id")), transform = trans)
 
     expect_s3_class(dict$get("nominal_resolution"), "character")
     expect_identical(
         unique(vapply(dict$get("nominal_resolution"), typeof, "", USE.NAMES = FALSE)),
         "character"
     )
-    expect_message(print(dict$get("nominal_resolution")))
+    expect_snapshot(print(dict$get("nominal_resolution")), transform = trans)
 
     expect_s3_class(dict$get("realm"), "list")
     expect_identical(
         unique(vapply(dict$get("realm"), typeof, "", USE.NAMES = FALSE)),
         "character"
     )
-    expect_message(print(dict$get("realm")))
+    expect_snapshot(print(dict$get("realm")), transform = trans)
 
     expect_s3_class(dict$get("required_global_attributes"), "character")
-    expect_message(print(dict$get("required_global_attributes")))
+    expect_snapshot(print(dict$get("required_global_attributes")), transform = trans)
 
     expect_s3_class(dict$get("source_id"), "data.table")
     expect_identical(
@@ -166,35 +170,42 @@ test_that("Cmip6Dict$get()", {
             activity_participation = "list", model_component = "list", license_info = "list"
         )
     )
[…]
