+++ tests/testthat/_snaps/dict.md
@@ -1,0 +1,632 @@
+# Cmip6Dict$get()
+
+    Code
+      print(dict$get("drs"))
+    Message <cliMessage>
+      == CMIP6CV Data Reference Syntax (DRS) ===================== CMIP6 Dictionary ==
+      
+      -- <VERSION METADATA> ----------------------------------------------------------
+      * CV Version: `6.2.58.49`
+      * CV Modified: [yyyy-mm-dd HH:MM:SS UTC]
+      * DRS Modified: [yyyy-mm-dd HH:MM:SS UTC]
+      * DRS Note: "Add CMIP6 Data Reference Syntax (DRS) templates"
+      
+      -- <STORED TYPE> ---------------------------------------------------------------
+      * Stored type: <list>
+      
+      -- <VALUES> --------------------------------------------------------------------
+      * Directory path example:
+        "CMIP6/CMIP/MOHC/HadGEM3-GC31-MM/historical/r1i1p1f3/Amon/tas/gn/v20191207/"
+      * Directory path sub experiment example:
+        "CMIP6/DCPP/MOHC/HadGEM3-GC31-MM/dcppA-hindcast/s1960-r1i1p1f2/Amon/tas/gn/v20200417/"
+      * Directory path template:
+        "<mip_era>/<activity_id>/<institution_id>/<source_id>/<experiment_id>/<member_id>/<table_id>/<variable_id>/<grid_label>/<version>"
+      * Filename example:
+        "tas_Amon_HadGEM3-GC31-MM_historical_r1i1p1f3_gn_185001-186912.nc"
+      * Filename sub experiment example:
+        "tas_Amon_HadGEM3-GC31-MM_dcppA-hindcast_s1960-r1i1p1f2_gn_196011-196012.nc"
+      # ... with 1 more item
+
+---
+
+    Code
+      print(dict$get("activity_id"))
+    Message <cliMessage>
+      == CMIP6CV Activity ID ===================================== CMIP6 Dictionary ==
+      
+      -- <VERSION METADATA> ----------------------------------------------------------
+      * CV Version: `6.2.58.49`
+      * CV Modified: [yyyy-mm-dd HH:MM:SS UTC]
+      * ActivityId Modified: [yyyy-mm-dd HH:MM:SS UTC]
+      * ActivityId Note: "Update activity_id to include CDRMIP and PAMIP"
+      
+      -- <STORED TYPE> ---------------------------------------------------------------
+      * Stored type: <list>
+      
+      -- <VALUES> --------------------------------------------------------------------
+      * AerChemMIP: "Aerosols and Chemistry Model Intercomparison Project"
+      * C4MIP: "Coupled Climate Carbon Cycle Model Intercomparison Project"
+      * CDRMIP: "Carbon Dioxide Removal Model Intercomparison Project"
+      * CFMIP: "Cloud Feedback Model Intercomparison Project"
+      * CMIP: "CMIP DECK: 1pctCO2, abrupt4xCO2, amip, esm-piControl,
+        esm-historical, historical, and piControl experiments"
+      # ... with 19 more items
+
+---
+
+    Code
+      print(dict$get("experiment_id"))
+    Message <cliMessage>
+      == CMIP6CV Experiment ID =================================== CMIP6 Dictionary ==
+      
+      -- <VERSION METADATA> ----------------------------------------------------------
+      * CV Version: `6.2.58.49`
+      * CV Modified: [yyyy-mm-dd HH:MM:SS UTC]
+      * ExperimentId Modified: [yyyy-mm-dd HH:MM:SS UTC]
+      * ExperimentId Note: "Revise experiment_id historical parent experiments"
+      
+      -- <STORED TYPE> ---------------------------------------------------------------
[…]
