### 2023-02-01 238a4cc
[feat] Add `host` parameter in `init_cmip6_index()`
