### 2023-02-01 6e79b84
[feat] Add `host` parameter in `init_cmip6_index()`
