### 2023-02-01 624ec77
[fix] Make sure `host` is respected in `init_cmip6_index()`
