### 2023-02-01 20e6b7d
[fix] Treat big integers as characters when parsing JSON
