+++ R/query.R
@@ -5,7 +5,7 @@ NULL
 # TODO: collect all results with auto-pagination
 # TODO: automatically add fields that are specified via methods
 read_json_response <- function(url, strict = TRUE, ...) {
-    q <- tryCatch(jsonlite::fromJSON(url, ...), warning = function(w) w, error = function(e) e)
+    q <- tryCatch(jsonlite::fromJSON(url, bigint_as_char = TRUE, ...), warning = function(w) w, error = function(e) e)
 
     # nocov start
     if (inherits(q, "warning") || inherits(q, "error")) {
@@ -292,12 +292,11 @@ EsgfQuery <- R6::R6Class("EsgfQuery",
             }
 
             # init parameter values
-            params <- setdiff(grep("^param_", names(private), value = TRUE), "param_others")
-            for (param in params) {
-                if (!is.null(.subset2(private, param))) {
-                    private[[param]] <- new_query_param(
-                        name = sub("param_", "", param, fixed = TRUE),
-                        value = .subset2(private, param)
+            params <- setdiff(names(private$parameters), "others")
+            for (name in params) {
+                if (!is.null(.subset2(private$parameters, name))) {
+                    private$parameters[[name]] <- new_query_param(
+                        name = name, value = .subset2(private$parameters, name)
                     )
                 }
             }
@@ -330,7 +329,7 @@ EsgfQuery <- R6::R6Class("EsgfQuery",
                 "Building facet cache for host {.var {private$url_host}}...",
                 "Built facet query cache for host {.var {private$url_host}} built at {private$facet_cache()$timestamp}."
             ))
-            query_build_facet_cache(private$url_host, private$param_project)
+            query_build_facet_cache(private$url_host, private$parameters$project)
             self
         },
 
@@ -448,8 +447,8 @@ EsgfQuery <- R6::R6Class("EsgfQuery",
         #' q$project(NULL)
         #' }
         project = function(value = "CMIP6") {
-            if (missing(value)) return(private$param_project)
-            private$param_project <- private$new_facet_param("project", value)
+            if (missing(value)) return(private$parameters$project)
+            private$parameters$project <- private$new_facet_param("project", value)
             self
         },
 
@@ -477,8 +476,8 @@ EsgfQuery <- R6::R6Class("EsgfQuery",
         #' q$activity_id(NULL)
         #' }
         activity_id = function(value) {
-            if (missing(value)) return(private$param_activity_id)
-            private$param_activity_id <- private$new_facet_param("activity_id", value)
+            if (missing(value)) return(private$parameters$activity_id)
+            private$parameters$activity_id <- private$new_facet_param("activity_id", value)
             self
         },
 
@@ -506,8 +505,8 @@ EsgfQuery <- R6::R6Class("EsgfQuery",
         #' q$experiment_id(NULL)
         #' }
         experiment_id = function(value) {
-            if (missing(value)) return(private$param_experiment_id)
-            private$param_experiment_id <- private$new_facet_param("experiment_id", value)
+            if (missing(value)) return(private$parameters$experiment_id)
+            private$parameters$experiment_id <- private$new_facet_param("experiment_id", value)
             self
         },
 
@@ -535,8 +534,8 @@ EsgfQuery <- R6::R6Class("EsgfQuery",
         #' q$source_id(NULL)
         #' }
         source_id = function(value) {
-            if (missing(value)) return(private$param_source_id)
-            private$param_source_id <- private$new_facet_param("source_id", value)
+            if (missing(value)) return(private$parameters$source_id)
+            private$parameters$source_id <- private$new_facet_param("source_id", value)
             self
         },
 
@@ -564,8 +563,8 @@ EsgfQuery <- R6::R6Class("EsgfQuery",
[…]
