### 2023-02-02 4b1d61a
[refactor] Store query parameters in `EsgfQueryResult`
