### 2023-02-02 3f956fd
[feat] Store timestamp when reading query responses
