### 2023-02-02 5f47b5b
[feat] Update printing methods
