### 2023-02-02 2a32e9f
[feat] Update printing methods
