+++ man/EsgfQueryResultFile.Rd
@@ -37,6 +37,7 @@ files.}
 \subsection{Public methods}{
 \itemize{
 \item \href{#method-EsgfQueryResultFile-to_dt}{\code{EsgfQueryResultFile$to_dt()}}
+\item \href{#method-EsgfQueryResultFile-print}{\code{EsgfQueryResultFile$print()}}
 \item \href{#method-EsgfQueryResultFile-clone}{\code{EsgfQueryResultFile$clone()}}
 }
 }
@@ -71,6 +72,27 @@ including \code{url} and \code{size}. Default: \code{FALSE}.}
 }
 \subsection{Returns}{
 A \link[data.table:data.table]{data.table}.
+}
+}
+\if{html}{\out{<hr>}}
+\if{html}{\out{<a id="method-EsgfQueryResultFile-print"></a>}}
+\if{latex}{\out{\hypertarget{method-EsgfQueryResultFile-print}{}}}
+\subsection{Method \code{print()}}{
+Print a summary of the current dataset
+\subsection{Usage}{
+\if{html}{\out{<div class="r">}}\preformatted{EsgfQueryResultFile$print(n = 10L)}\if{html}{\out{</div>}}
+}
+
+\subsection{Arguments}{
+\if{html}{\out{<div class="arguments">}}
+\describe{
+\item{\code{n}}{An integer indicating how many items to print. If \code{NULL},
+all items will be printed. Default: \code{10L}.}
+}
+\if{html}{\out{</div>}}
+}
+\subsection{Returns}{
+The \code{EsgfQueryResultFile} object itself, invisibly.
 }
 }
 \if{html}{\out{<hr>}}
