### 2023-02-02 be0dde0
[feat] Add support to select datasets to collect
