### 2023-02-02 bfe9721
[feat] Add support to select datasets to collect
