### 2023-02-02 ede9042
[refactor] Rename class name of CMIP6 dictionary
