+++ R/dict.R
@@ -451,7 +451,7 @@ cmip6dict_parse_cv_vec <- function(file, subclass = NULL) {
         cmip6dict_parse_cv_version_metadata(json$version_metadata)
     )
 
-    structure(res, class = c(subclass, "CMIP6CV", typeof(res)))
+    structure(res, class = c(subclass, "Cmip6CV", typeof(res)))
 }
 
 cmip6dict_parse_cv_list <- function(file, subclass = NULL) {
@@ -461,15 +461,15 @@ cmip6dict_parse_cv_list <- function(file, subclass = NULL) {
         cmip6dict_parse_cv_version_metadata(json$version_metadata)
     )
 
-    structure(res, class = c(subclass, "CMIP6CV", "list"))
+    structure(res, class = c(subclass, "Cmip6CV", "list"))
 }
 
 cmip6dict_parse_cv_drs <- function(file) {
-    cmip6dict_parse_cv_list(file, "CMIP6CV_DRS")
+    cmip6dict_parse_cv_list(file, "Cmip6CV_DRS")
 }
 
 cmip6dict_parse_cv_activity_id <- function(file) {
-    cmip6dict_parse_cv_list(file, "CMIP6CV_ActivityId")
+    cmip6dict_parse_cv_list(file, "Cmip6CV_ActivityId")
 }
 
 cmip6dict_parse_cv_experiment_id <- function(file) {
@@ -506,31 +506,31 @@ cmip6dict_parse_cv_experiment_id <- function(file) {
     ))
     data.table::setattr(d, "version", cmip6dict_parse_cv_version_metadata(json$version_metadata))
 
-    structure(d, class = c("CMIP6CV_ExperimentId", "CMIP6CV", class(d)))
+    structure(d, class = c("Cmip6CV_ExperimentId", "Cmip6CV", class(d)))
 }
 
 cmip6dict_parse_cv_frequency <- function(file) {
-    cmip6dict_parse_cv_list(file, "CMIP6CV_Frequency")
+    cmip6dict_parse_cv_list(file, "Cmip6CV_Frequency")
 }
 
 cmip6dict_parse_cv_grid_label <- function(file) {
-    cmip6dict_parse_cv_list(file, "CMIP6CV_GridLabel")
+    cmip6dict_parse_cv_list(file, "Cmip6CV_GridLabel")
 }
 
 cmip6dict_parse_cv_institution_id <- function(file) {
-    cmip6dict_parse_cv_list(file, "CMIP6CV_InstitutionId")
+    cmip6dict_parse_cv_list(file, "Cmip6CV_InstitutionId")
 }
 
 cmip6dict_parse_cv_nominal_resolution <- function(file) {
-    cmip6dict_parse_cv_vec(file, "CMIP6CV_Resolution")
+    cmip6dict_parse_cv_vec(file, "Cmip6CV_Resolution")
 }
 
 cmip6dict_parse_cv_realm <- function(file) {
-    cmip6dict_parse_cv_list(file, "CMIP6CV_Realm")
+    cmip6dict_parse_cv_list(file, "Cmip6CV_Realm")
 }
 
 cmip6dict_parse_cv_required_global_attributes <- function(file) {
-    cmip6dict_parse_cv_vec(file, "CMIP6CV_ReqGlobAttr")
+    cmip6dict_parse_cv_vec(file, "Cmip6CV_ReqGlobAttr")
 }
 
 cmip6dict_parse_cv_source_id <- function(file) {
@@ -561,34 +561,35 @@ cmip6dict_parse_cv_source_id <- function(file) {
     ))
     data.table::setattr(d, "version", cmip6dict_parse_cv_version_metadata(json$version_metadata))
 
-    structure(d, class = c("CMIP6CV_SourceId", "CMIP6CV", class(d)))
[…]
