+++ NAMESPACE
@@ -1,7 +1,7 @@
 # Generated by roxygen2: do not edit by hand
 
-S3method(print,CMIP6CV)
-S3method(print,CMIP6DReq)
+S3method(print,Cmip6CV)
+S3method(print,Cmip6DReq)
 S3method(print,EsgfQueryParam)
 export(Cmip6Dict)
 export(EsgfQuery)
