### 2023-02-02 b8b1842
[feat] Add `epwshiftr.cache` option
