### 2023-02-02 0574717
[refactor] Always show unknown size for aggregations
