### 2023-02-02 c7d812b
[refactor] Always show unknown size for aggregations
