### 2023-02-02 d4c994e
[fix] Only use facet cache when possible when collecting results
