### 2023-02-02 052abdb
[test] Simplify dict printing method
