+++ tests/testthat/_snaps/dict.md
@@ -3,18 +3,13 @@
     Code
       print(dict$get("drs"))
     Message <cliMessage>
-      == CMIP6CV Data Reference Syntax (DRS) ===================== CMIP6 Dictionary ==
-      
-      -- <VERSION METADATA> ----------------------------------------------------------
+      == Cmip6CV Data Reference Syntax (DRS) ===================== CMIP6 Dictionary ==
       * CV Version: `6.2.58.49`
-      * CV Modified: [yyyy-mm-dd HH:MM:SS UTC]
-      * DRS Modified: [yyyy-mm-dd HH:MM:SS UTC]
+      * CV Modified: yyyy-mm-dd HH:MM:SS UTC
+      * DRS Modified: yyyy-mm-dd HH:MM:SS UTC
       * DRS Note: "Add CMIP6 Data Reference Syntax (DRS) templates"
       
-      -- <STORED TYPE> ---------------------------------------------------------------
-      * Stored type: <list>
-      
-      -- <VALUES> --------------------------------------------------------------------
+      -- Values <list> ---------------------------------------------------------------
       * Directory path example:
         "CMIP6/CMIP/MOHC/HadGEM3-GC31-MM/historical/r1i1p1f3/Amon/tas/gn/v20191207/"
       * Directory path sub experiment example:
@@ -32,18 +27,13 @@
     Code
       print(dict$get("activity_id"))
     Message <cliMessage>
-      == CMIP6CV Activity ID ===================================== CMIP6 Dictionary ==
-      
-      -- <VERSION METADATA> ----------------------------------------------------------
+      == Cmip6CV Activity ID ===================================== CMIP6 Dictionary ==
       * CV Version: `6.2.58.49`
-      * CV Modified: [yyyy-mm-dd HH:MM:SS UTC]
-      * ActivityId Modified: [yyyy-mm-dd HH:MM:SS UTC]
+      * CV Modified: yyyy-mm-dd HH:MM:SS UTC
+      * ActivityId Modified: yyyy-mm-dd HH:MM:SS UTC
       * ActivityId Note: "Update activity_id to include CDRMIP and PAMIP"
       
-      -- <STORED TYPE> ---------------------------------------------------------------
-      * Stored type: <list>
-      
-      -- <VALUES> --------------------------------------------------------------------
+      -- Values <list> ---------------------------------------------------------------
       * AerChemMIP: "Aerosols and Chemistry Model Intercomparison Project"
       * C4MIP: "Coupled Climate Carbon Cycle Model Intercomparison Project"
       * CDRMIP: "Carbon Dioxide Removal Model Intercomparison Project"
@@ -57,18 +47,13 @@
     Code
       print(dict$get("experiment_id"))
     Message <cliMessage>
-      == CMIP6CV Experiment ID =================================== CMIP6 Dictionary ==
-      
-      -- <VERSION METADATA> ----------------------------------------------------------
+      == Cmip6CV Experiment ID =================================== CMIP6 Dictionary ==
       * CV Version: `6.2.58.49`
-      * CV Modified: [yyyy-mm-dd HH:MM:SS UTC]
-      * ExperimentId Modified: [yyyy-mm-dd HH:MM:SS UTC]
+      * CV Modified: yyyy-mm-dd HH:MM:SS UTC
+      * ExperimentId Modified: yyyy-mm-dd HH:MM:SS UTC
       * ExperimentId Note: "Revise experiment_id historical parent experiments"
       
-      -- <STORED TYPE> ---------------------------------------------------------------
-      * Stored type: <data.table>
-      
-      -- <VALUES> --------------------------------------------------------------------
+      -- Values <data.table> ---------------------------------------------------------
       
[…]
