### 2023-02-02 7e9f4d1
[test] Simplify dict printing method
