### 2023-02-03 9a85042
[fix] Clean up `query_collect()`
