### 2023-02-03 5374c1d
[refactor] Do not build facet cache by default
