### 2023-02-03 891f76b
[refactor] Do not build facet cache by default
