+++ tests/testthat/test-query.R
@@ -51,8 +51,8 @@ test_that("query_esgf()", {
 
     # NOTE: attached facet cache is used throughout tests
     attach_facet_cache()
-    expect_s3_class(q <- EsgfQuery$new(host), "EsgfQuery")
-    expect_s3_class(q <- query_esgf(host), "EsgfQuery")
+    expect_s3_class(q <- EsgfQuery$new(host, TRUE), "EsgfQuery")
+    expect_s3_class(q <- query_esgf(host, TRUE), "EsgfQuery")
 })
 
 test_that("EsgfQuery$list_all_facets()", {
@@ -60,7 +60,7 @@ test_that("EsgfQuery$list_all_facets()", {
 
     skip_on_cran()
 
-    expect_s3_class(q <- query_esgf(host), "EsgfQuery")
+    expect_s3_class(q <- query_esgf(host, TRUE), "EsgfQuery")
     expect_type(q$list_all_facets(), "character")
 })
 
@@ -69,7 +69,7 @@ test_that("EsgfQuery$list_all_fields()", {
 
     skip_on_cran()
 
-    expect_s3_class(q <- query_esgf(host), "EsgfQuery")
+    expect_s3_class(q <- query_esgf(host, TRUE), "EsgfQuery")
     expect_type(q$list_all_fields(), "character")
 })
 
@@ -78,7 +78,7 @@ test_that("EsgfQuery$list_all_shards()", {
 
     skip_on_cran()
 
-    expect_s3_class(q <- query_esgf(host), "EsgfQuery")
+    expect_s3_class(q <- query_esgf(host, TRUE), "EsgfQuery")
     expect_type(q$list_all_shards(), "character")
 })
 
@@ -87,7 +87,7 @@ test_that("EsgfQuery$list_all_values()", {
 
     skip_on_cran()
 
-    expect_s3_class(q <- query_esgf(host), "EsgfQuery")
+    expect_s3_class(q <- query_esgf(host, TRUE), "EsgfQuery")
     expect_type(q$list_all_values("activity_id"), "character")
 })
 
@@ -146,7 +146,7 @@ test_that("EsgfQuery$project() and other facet methods", {
 
     skip_on_cran()
 
-    expect_s3_class(q <- query_esgf(host), "EsgfQuery")
+    expect_s3_class(q <- query_esgf(host, TRUE), "EsgfQuery")
 
     # project
     expect_equal(q$project()$value, "CMIP6")
@@ -206,7 +206,7 @@ test_that("EsgfQuery$facets()", {
 
     skip_on_cran()
 
-    expect_s3_class(q <- query_esgf(host), "EsgfQuery")
+    expect_s3_class(q <- query_esgf(host, TRUE), "EsgfQuery")
 
     expect_null(q$facets())
     expect_equal(q$facets(c("activity_id", "source_id"))$facets()$value, c("activity_id", "source_id"))
@@ -221,7 +221,7 @@ test_that("EsgfQuery$fields()", {
 
     skip_on_cran()
 
-    expect_s3_class(q <- query_esgf(host), "EsgfQuery")
+    expect_s3_class(q <- query_esgf(host, TRUE), "EsgfQuery")
 
     expect_equal(q$fields()$value, "*")
     expect_equal(q$fields(c("activity_id", "source_id"))$fields()$value, c("activity_id", "source_id"))
@@ -245,7 +245,7 @@ test_that("EsgfQuery$shards()", {
 
     skip_on_cran()
 
-    expect_s3_class(q <- query_esgf(host), "EsgfQuery")
+    expect_s3_class(q <- query_esgf(host, TRUE), "EsgfQuery")
 
     expect_null(q$shards())
[…]
