### 2023-02-03 2828266
[refactor] Store facet listing in each `EsgfQuery`
