+++ tests/testthat/test-query.R
@@ -44,28 +44,29 @@ test_that("ESGF Query Parameter works", {
 })
 
 test_that("query_esgf()", {
-    expect_s3_class(EsgfQuery$new(build = FALSE), "EsgfQuery")
-    expect_s3_class(query_esgf(build = FALSE), "EsgfQuery")
+    expect_s3_class(EsgfQuery$new(listing = FALSE), "EsgfQuery")
+    expect_s3_class(query_esgf(listing = FALSE), "EsgfQuery")
 
     skip_on_cran()
 
-    # NOTE: attached facet cache is used throughout tests
+    # NOTE: attached facet listing cache is used throughout tests
     attach_facet_cache()
     expect_s3_class(q <- EsgfQuery$new(host, TRUE), "EsgfQuery")
     expect_s3_class(q <- query_esgf(host, TRUE), "EsgfQuery")
 })
 
 test_that("EsgfQuery$list_all_facets()", {
-    expect_null(query_esgf(build = FALSE)$list_all_facets())
+    expect_null(query_esgf(listing = FALSE)$list_all_facets())
 
     skip_on_cran()
 
+    options("epwshiftr.verbose" = TRUE)
     expect_s3_class(q <- query_esgf(host, TRUE), "EsgfQuery")
     expect_type(q$list_all_facets(), "character")
 })
 
 test_that("EsgfQuery$list_all_fields()", {
-    expect_null(query_esgf(build = FALSE)$list_all_fields())
+    expect_null(query_esgf(listing = FALSE)$list_all_fields())
 
     skip_on_cran()
 
@@ -74,7 +75,7 @@ test_that("EsgfQuery$list_all_fields()", {
 })
 
 test_that("EsgfQuery$list_all_shards()", {
-    expect_null(query_esgf(build = FALSE)$list_all_shards())
+    expect_null(query_esgf(listing = FALSE)$list_all_shards())
 
     skip_on_cran()
 
@@ -83,7 +84,7 @@ test_that("EsgfQuery$list_all_shards()", {
 })
 
 test_that("EsgfQuery$list_all_values()", {
-    expect_null(query_esgf(build = FALSE)$list_all_values())
+    expect_null(query_esgf(listing = FALSE)$list_all_values("activity_id"))
 
     skip_on_cran()
 
@@ -92,7 +93,7 @@ test_that("EsgfQuery$list_all_values()", {
 })
 
 test_that("EsgfQuery$project() and other facet methods", {
-    q <- query_esgf(build = FALSE)
+    q <- query_esgf(listing = FALSE)
 
     # project
     expect_equal(q$project()$value, "CMIP6")
@@ -200,9 +201,9 @@ test_that("EsgfQuery$project() and other facet methods", {
 })
 
 test_that("EsgfQuery$facets()", {
-    expect_null(query_esgf(build = FALSE)$facets())
-    expect_equal(query_esgf(build = FALSE)$facets(c("activity_id", "source_id"))$facets()$value, c("activity_id", "source_id"))
-    expect_null(query_esgf(build = FALSE)$facets(NULL)$facets())
+    expect_null(query_esgf(listing = FALSE)$facets())
+    expect_equal(query_esgf(listing = FALSE)$facets(c("activity_id", "source_id"))$facets()$value, c("activity_id", "source_id"))
+    expect_null(query_esgf(listing = FALSE)$facets(NULL)$facets())
 
     skip_on_cran()
 
@@ -214,10 +215,10 @@ test_that("EsgfQuery$facets()", {
 })
 
 test_that("EsgfQuery$fields()", {
-    expect_equal(query_esgf(build = FALSE)$fields()$value, "*")
[…]
