### 2023-02-03 6b61a03
[refactor] Store facet listing in each `EsgfQuery`
