+++ R/query.R
@@ -214,9 +214,9 @@ format.EsgfQueryParam <- function(x, encode = TRUE, space = FALSE, ...) {
 #'
 #' `EsgfQuery` object also provide several other helper functions:
 #'
-#' - \href{#method-EsgfQuery-build_cache}{\code{EsgfQuery$build_cache()}}:
-#'   By default, `EsgfQuery$build_cache()` is not called when initialize a new
-#'   `EsgfQuery` object. Basically, `EsgfQuery$build_cahce()` sends a
+#' - \href{#method-EsgfQuery-build_listing}{\code{EsgfQuery$build_listing()}}:
+#'   By default, `EsgfQuery$build_listing()` is not called when initialize a new
+#'   `EsgfQuery` object. Basically, `EsgfQuery$build_listing()` sends a
 #'   [facet listing query](https://esgf.github.io/esg-search/ESGF_Search_RESTful_API.html#facet-listings)
 #'   to the index node and stores the response internally. The response contains
 #'   all available facets and shards and is used as a source for validating user
@@ -234,7 +234,7 @@ format.EsgfQueryParam <- function(x, encode = TRUE, space = FALSE, ...) {
 #'
 #' - \href{#method-EsgfQuery-print}{\code{EsgfQuery$print()}}: Print a summary
 #'   of the current `EsgfQuery` object including the host URL, the built time of
-#'   facet cache and all query parameters.
+#'   facet listing and all query parameters.
 #'
 #' @author Hongyuan Jia
 #'
@@ -246,11 +246,12 @@ format.EsgfQueryParam <- function(x, encode = TRUE, space = FALSE, ...) {
 #'        [LLNL (Lawrence Livermore National Laboratory) Index Node](http://esgf-node.llnl.gov),
 #'        which is `"https://esgf-node.llnl.gov/esg-search"`.
 #'
-#' @param build Whether to send the facet listing query. Default: `FALSE`.
+#' @param listing Whether to send the facet listing query used to validate
+#'        inputs. Default: `FALSE`.
 #'
 #' @export
-query_esgf <- function(host = "https://esgf-node.llnl.gov/esg-search", build = FALSE) {
-    EsgfQuery$new(host = host, build = build)
+query_esgf <- function(host = "https://esgf-node.llnl.gov/esg-search", listing = FALSE) {
+    EsgfQuery$new(host = host, listing = listing)
 }
 
 #' @name EsgfQuery
@@ -268,7 +269,7 @@ EsgfQuery <- R6::R6Class("EsgfQuery",
         #'        Laboratory)](http://esgf-node.llnl.gov) Index Node, which is
         #'        `"https://esgf-node.llnl.gov/esg-search"`.
         #'
-        #' @param build Whether to send the facet listing query.
+        #' @param listing Whether to send the facet listing query.
         #'        A [facet listing query](https://esgf.github.io/esg-search/ESGF_Search_RESTful_API.html#facet-listings)
         #'        can be sent to the index node to get all available facets and
         #'        shards. When available, this information will be automatically
@@ -284,9 +285,9 @@ EsgfQuery <- R6::R6Class("EsgfQuery",
         #' q <- EsgfQuery$new(host = "https://esgf-node.llnl.gov/esg-search")
         #' q
         #' }
-        initialize = function(host = "https://esgf-node.llnl.gov/esg-search", build = FALSE) {
+        initialize = function(host = "https://esgf-node.llnl.gov/esg-search", listing = FALSE) {
             checkmate::assert_string(host)
-            checkmate::assert_flag(build)
+            checkmate::assert_flag(listing)
 
             # in case of a encoded URL
             private$url_host <- utils::URLdecode(host)
@@ -294,14 +295,7 @@ EsgfQuery <- R6::R6Class("EsgfQuery",
[…]
