### 2023-02-21 a9737f5
[feat] Add `EsgfQuery$save()` and `$load()`
1. Remove `EsgfQuery$response()`
2. Add support for validating data using predefined schema
