+++ man/EsgfQuery.Rd
@@ -143,9 +143,9 @@ query results and format it into an \link{EsgfQueryResultDataset} object.
 
 \code{EsgfQuery} object also provide several other helper functions:
 \itemize{
-\item \href{#method-EsgfQuery-build_cache}{\code{EsgfQuery$build_cache()}}:
-By default, \code{EsgfQuery$build_cache()} is not called when initialize a new
-\code{EsgfQuery} object. Basically, \code{EsgfQuery$build_cahce()} sends a
+\item \href{#method-EsgfQuery-build_listing}{\code{EsgfQuery$build_listing()}}:
+By default, \code{EsgfQuery$build_listing()} is not called when initialize a new
+\code{EsgfQuery} object. Basically, \code{EsgfQuery$build_listing()} sends a
 \href{https://esgf.github.io/esg-search/ESGF_Search_RESTful_API.html#facet-listings}{facet listing query}
 to the index node and stores the response internally. The response contains
 all available facets and shards and is used as a source for validating user
@@ -153,14 +153,9 @@ input for parameter setters.
 \item \href{#method-EsgfQuery-url}{\code{EsgfQuery$url()}}: Returns the actual
 query URL or the wget script URL which can be used to download all files
 matching the given constraints..
-\item \href{#method-EsgfQuery-response}{\code{EsgfQuery$response()}}: Returns the
-actual response of
-\href{#method-EsgfQuery-count}{\code{EsgfQuery$count()}} and
-\href{#method-EsgfQuery-collect}{\code{EsgfQuery$collect()}}. It is a named
-list generated from the JSON response using \code{\link[jsonlite:fromJSON]{jsonlite::fromJSON()}}.
 \item \href{#method-EsgfQuery-print}{\code{EsgfQuery$print()}}: Print a summary
 of the current \code{EsgfQuery} object including the host URL, the built time of
-facet cache and all query parameters.
+facet listing and all query parameters.
 }
 }
 }
@@ -177,11 +172,11 @@ q
 }
 
 ## ------------------------------------------------
-## Method `EsgfQuery$build_cache`
+## Method `EsgfQuery$build_listing`
 ## ------------------------------------------------
 
 \dontrun{
-q$build_cache()
+q$build_listing()
 }
 
 ## ------------------------------------------------
@@ -586,11 +581,23 @@ identical(res2$count(), res4$count())
 }
 
 ## ------------------------------------------------
-## Method `EsgfQuery$response`
+## Method `EsgfQuery$save`
 ## ------------------------------------------------
 
 \dontrun{
-q$response()
+q$save(tempfile(fileext = ".json"))
+}
+
+## ------------------------------------------------
+## Method `EsgfQuery$load`
+## ------------------------------------------------
+
+\dontrun{
+f <- tempfile(fileext = "json")
+
+q <- query_esgf()
+json <- q$save(f)
+q$load(f)
 }
 
 ## ------------------------------------------------
@@ -608,7 +615,7 @@ Hongyuan Jia
 \subsection{Public methods}{
 \itemize{
 \item \href{#method-EsgfQuery-new}{\code{EsgfQuery$new()}}
-\item \href{#method-EsgfQuery-build_cache}{\code{EsgfQuery$build_cache()}}
+\item \href{#method-EsgfQuery-build_listing}{\code{EsgfQuery$build_listing()}}
 \item \href{#method-EsgfQuery-list_all_facets}{\code{EsgfQuery$list_all_facets()}}
[…]
