### 2023-02-21 6b082e0
[feat] Add `EsgfQuery$save()` and `$load()`
1. Remove `EsgfQuery$response()`
2. Add support for validating data using predefined schema
