+++ R/query.R
@@ -2,7 +2,8 @@
 NULL
 
 # TODO: update doc about query type
-# TODO: automatically add fields that are specified via methods
+# TODO: be able to import query parameters?
+# read_json_response {{{
 read_json_response <- function(url, strict = TRUE, cache = getOption("epwshiftr.cache", TRUE), ...) {
     checkmate::assert(checkmate::check_flag(cache), checkmate::check_count(cache, positive = TRUE))
     if (cache && url %in% names(this$cache$query)) return(this$cache$query[[url]])
@@ -14,7 +15,8 @@ read_json_response <- function(url, strict = TRUE, cache = getOption("epwshiftr.
     if (inherits(res, "warning") || inherits(res, "error")) {
         cond_fun <- if (strict) stop else warning
         cond_fun(
-            "No matched data. Please check network connection and the availability of ESGF index/data node. Details: \n",
+            "No matched data.",
+            "Please check network connection and the availability of ESGF index/data node. Details: \n",
             conditionMessage(res)
         )
         res <- NULL
@@ -37,7 +39,9 @@ read_json_response <- function(url, strict = TRUE, cache = getOption("epwshiftr.
 
     res
 }
+# }}}
 
+# new_query_param {{{
 new_query_param <- function(name, value) {
     checkmate::assert_string(name)
 
@@ -56,6 +60,7 @@ new_query_param <- function(name, value) {
     value$name <- name
     structure(value, class = c("EsgfQueryParam", "list"))
 }
+# }}}
 
 is_query_param <- function(x) inherits(x, "EsgfQueryParam")
 
@@ -65,6 +70,7 @@ print.EsgfQueryParam <- function(x, encode = FALSE, space = TRUE, ...) {
     invisible(x)
 }
 
+# format.EsgfQueryParam {{{
 format.EsgfQueryParam <- function(x, encode = TRUE, space = FALSE, ...) {
     if (is.logical(x$value)) {
         res <- tolower(if (x$negate) !x$value else x$value)
@@ -84,6 +90,9 @@ format.EsgfQueryParam <- function(x, encode = TRUE, space = FALSE, ...) {
         paste0(x$name, spc, "=", spc, paste0(res, collapse = paste0(",", spc)))
     }
 }
+# }}}
+
+# query_esgf {{{
 
 #' Query CMIP6 data using ESGF search RESTful API
 #'
@@ -226,12 +235,6 @@ format.EsgfQueryParam <- function(x, encode = TRUE, space = FALSE, ...) {
 #'   query URL or the wget script URL which can be used to download all files
 #'   matching the given constraints..
 #'
-#' - \href{#method-EsgfQuery-response}{\code{EsgfQuery$response()}}: Returns the
-#'   actual response of
-#'   \href{#method-EsgfQuery-count}{\code{EsgfQuery$count()}} and
-#'   \href{#method-EsgfQuery-collect}{\code{EsgfQuery$collect()}}. It is a named
-#'   list generated from the JSON response using [jsonlite::fromJSON()].
-#'
 #' - \href{#method-EsgfQuery-print}{\code{EsgfQuery$print()}}: Print a summary
 #'   of the current `EsgfQuery` object including the host URL, the built time of
 #'   facet listing and all query parameters.
@@ -253,12 +256,15 @@ format.EsgfQueryParam <- function(x, encode = TRUE, space = FALSE, ...) {
 query_esgf <- function(host = "https://esgf-node.llnl.gov/esg-search", listing = FALSE) {
     EsgfQuery$new(host = host, listing = listing)
 }
+# }}}
 
 #' @name EsgfQuery
 #' @export
+# EsgfQuery {{{
 EsgfQuery <- R6::R6Class("EsgfQuery",
     lock_class = TRUE,
     public = list(
+        # initialize {{{
[…]
