### 2023-05-11 9843c0b
[feat] Add `EsgfQueryResult$save()` and `$load()`
