+++ R/query-result.R
@@ -1,3 +1,4 @@
+# EsgfQueryResult {{{
 #' Base class for results for ESGF query
 #'
 #' @description
@@ -14,6 +15,7 @@
 EsgfQueryResult <- R6::R6Class("EsgfQueryResult",
     lock_class = TRUE,
     public = list(
+        # initialize {{{
         #' @description
         #' Create a new EsgfQueryResult object
         #'
@@ -24,17 +26,19 @@ EsgfQueryResult <- R6::R6Class("EsgfQueryResult",
         #'
         #' @param params A list of query parameters.
         #'
-        #' @param result The result of an query response.
+        #' @param response The result of an query response.
         #'
         #' @return An `EsgfQueryResult` object.
         #'
-        initialize = function(host, params, result) {
+        initialize = function(host, params, response) {
             private$url_host <- host
-            private$parameters <- params
-            private$result <- result
+            private$parameter <- params
+            private$response <- response
             self
         },
+        # }}}
 
+        # to_dt {{{
         #' @description
         #' Convert the results into a [data.table][data.table::data.table()]
         #'
@@ -81,7 +85,9 @@ EsgfQueryResult <- R6::R6Class("EsgfQueryResult",
 
             res
         },
+        # }}}
 
+        # count {{{
         #' @description
         #' Count the number of matched records in current result
         #'
@@ -89,24 +95,92 @@ EsgfQueryResult <- R6::R6Class("EsgfQueryResult",
         #'
         count = function() {
             length(self$id)
+        },
+        # }}}
+
+        # save {{{
+        #' @description
+        #' Save the result into a JSON file
+        #'
+        #' `$save()` puts main data of an `EsgfQueryResult` object into a JSON file
+        #' which can be loaded to restore the current state of result using
+        #' \href{#method-EsgfQueryResult-load}{\code{EsgfQueryResult$load()}}.
+        #'
+        #' @param file A string indicating the JSON file path to save the data
+        #'        to. Default: `result.json`.
+        #'
+        #' @param pretty Whether to add indentation whitespace to JSON output.
+        #'        For details, please see [jsonlite::toJSON()]. Default: `TRUE`.
+        #'
+        #' @return The full path of the output JSON file.
+        #'
+        save = function(file = "result.json", pretty = TRUE) {
+            query_save(
+                host = private$url_host,
+                parameter = private$parameter,
+                response = private$response,
+                last_result = private$last_result,
+                file = file, pretty = pretty
+            )
+        },
+        # }}}
+
+        # load {{{
+        #' @description
+        #' Restore the result state from an JSON file
+        #'
+        #' `$load()` reads data of an `EsgfQueryResult` object from a JSON file
+        #' created using
+        #' \href{#method-EsgfQueryResult-save}{\code{EsgfQueryResult$save()}}.
+        #'
+        #' @param file A string indicating the JSON file path to read the data
+        #'        from. Default: `result.json`.
+        #'
+        #' @return The modified `EsgfQueryResult` object itself.
+        #'
+        load = function(file = "result.json") {
+            q <- query_load(file, SCHEMA_RESULT_DATASET)
+
+            private$url_host <- q$host
+            private$parameter <- q$parameter
+            private$response <- q$response
+
+            if (length(q$last_result)) {
+                # restore parameters in last result
+                q$last_result$parameter <- restore_params(
+                    input = q$last_result$parameter,
+                    params = list(dataset_id = NULL, fields = NULL, latest = TRUE,
[…]
