### 2023-05-11 ff9e711
[feat] Add `EsgfQueryResult$save()` and `$load()`
