### 2023-05-15 22513a7
[fix] Fix `facet_listing` schema
