### 2023-05-15 75e75a4
[fix] Fix `facet_listing` JSON loading
