### 2023-05-15 de56308
[feat] Add `EsgfQuery$has_listing()`
