### 2023-05-15 33c289b
[feat] Add `EsgfQuery$has_listing()`
