+++ tests/testthat/test-query.R
@@ -1,5 +1,6 @@
 host <- "https://esgf.ceda.ac.uk/esg-search"
 
+# ESGF Query Parameter {{{
 test_that("ESGF Query Parameter works", {
     # can create new query parameter
     expect_s3_class(
@@ -42,7 +43,9 @@ test_that("ESGF Query Parameter works", {
         )
     ))
 })
+# }}}
 
+# query_esgf() {{{
 test_that("query_esgf()", {
     expect_s3_class(EsgfQuery$new(listing = FALSE), "EsgfQuery")
     expect_s3_class(query_esgf(listing = FALSE), "EsgfQuery")
@@ -54,7 +57,20 @@ test_that("query_esgf()", {
     expect_s3_class(q <- EsgfQuery$new(host, TRUE), "EsgfQuery")
     expect_s3_class(q <- query_esgf(host, TRUE), "EsgfQuery")
 })
+# }}}
 
+# EsgfQuery$build_listing() and EsgfQuery$has_listing() {{{
+test_that("EsgfQuery$build_listing() and EsgfQuery$has_listing()", {
+    expect_false(query_esgf()$has_listing())
+
+    skip_on_cran()
+
+    expect_s3_class(q <- query_esgf(host)$build_listing(), "EsgfQuery")
+    expect_true(q$has_listing())
+})
+# }}}
+
+# EsgfQuery$list_all_facets() {{{
 test_that("EsgfQuery$list_all_facets()", {
     expect_null(query_esgf(listing = FALSE)$list_all_facets())
 
@@ -64,7 +80,9 @@ test_that("EsgfQuery$list_all_facets()", {
     expect_s3_class(q <- query_esgf(host, TRUE), "EsgfQuery")
     expect_type(q$list_all_facets(), "character")
 })
+# }}}
 
+# EsgfQuery$list_all_fields() {{{
 test_that("EsgfQuery$list_all_fields()", {
     expect_null(query_esgf(listing = FALSE)$list_all_fields())
 
@@ -73,7 +91,9 @@ test_that("EsgfQuery$list_all_fields()", {
     expect_s3_class(q <- query_esgf(host, TRUE), "EsgfQuery")
     expect_type(q$list_all_fields(), "character")
 })
+# }}}
 
+# EsgfQuery$list_all_shards() {{{
 test_that("EsgfQuery$list_all_shards()", {
     expect_null(query_esgf(listing = FALSE)$list_all_shards())
 
@@ -82,7 +102,9 @@ test_that("EsgfQuery$list_all_shards()", {
     expect_s3_class(q <- query_esgf(host, TRUE), "EsgfQuery")
     expect_type(q$list_all_shards(), "character")
 })
+# }}}
 
+# EsgfQuery$list_all_values() {{{
 test_that("EsgfQuery$list_all_values()", {
     expect_null(query_esgf(listing = FALSE)$list_all_values("activity_id"))
 
@@ -91,7 +113,9 @@ test_that("EsgfQuery$list_all_values()", {
     expect_s3_class(q <- query_esgf(host, TRUE), "EsgfQuery")
     expect_type(q$list_all_values("activity_id"), "character")
 })
+# }}}
 
+# EsgfQuery$project() and other facet methods {{{
 test_that("EsgfQuery$project() and other facet methods", {
     q <- query_esgf(listing = FALSE)
 
@@ -199,7 +223,9 @@ test_that("EsgfQuery$project() and other facet methods", {
     expect_equal(q$data_node("esgf.ceda.ac.uk")$data_node()$value, "esgf.ceda.ac.uk")
     expect_null(q$data_node(NULL)$data_node())
 })
+# }}}
 
[…]
