### 2023-05-15 e3f292c
[feat] Add `result_esgf()`
