+++ man/result_esgf.Rd
@@ -1,0 +1,19 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/query-result.R
+\name{result_esgf}
+\alias{result_esgf}
+\title{Create empty query result object}
+\usage{
+result_esgf(type = c("dataset", "file", "aggregation"))
+}
+\arguments{
+\item{type}{A string indicating what type of ESGF query result should be
+created. Should be one of \code{"dataset"}, \code{"file"} or "aggregation"`.}
+}
+\value{
+An empty \code{EsgfQueryResult} object of given type.
+}
+\description{
+\code{result_esgf()} creates an empty query result object of input type, so that
+you can load the saved JSON file via \code{EsgfQueryResult$load()}.
+}
