### 2023-05-15 fe4ec39
[feat] Add `result_esgf()`
