+++ NAMESPACE
@@ -16,6 +16,7 @@ export(load_cmip6_index)
 export(match_coord)
 export(morphing_epw)
 export(query_esgf)
+export(result_esgf)
 export(set_cmip6_index)
 export(summary_database)
 importFrom(PCICt,as.PCICt)
