### 2023-05-15 ae4b939
[doc] Add ESGF query vignette
