+++ vignettes/articles/cmip6_query.Rmd
@@ -1,0 +1,370 @@
+---
+title: "Interacting with the ESGF search API"
+author: "Hongyuan Jia"
+output:
+    rmarkdown::html_vignette:
+        df_print: "paged"
+        toc: true
+vignette: >
+    %\VignetteIndexEntry{Interacting with the ESGF search API}
+    %\VignetteEngine{knitr::rmarkdown}
+    %\VignetteEncoding{UTF-8}
+description: >
+    This vignette introduces you how to use epwshiftr to search and fetch data
+    using the ESGF search API.
+---
+
+```{r, include = FALSE}
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>"
+)
+```
+
+```{r setup, echo = FALSE, message = FALSE}
+library(epwshiftr)
+
+# show verbose messages
+options(epwshiftr.verbose = TRUE)
+```
+
+# ESGF search RESTful API
+
+The Earth System Grid Federation (ESGF) is an international collaboration for
+the software that powers most global climate change research, notably
+assessments by the Intergovernmental Panel on Climate Change (IPCC).
+
+The ESGF search service exposes RESTful APIs that can be used by clients to
+query the contents of the underlying search index, and return results matching
+the given constraints. The documentation of the APIs can be found using this
+[link](https://esgf.github.io/esg-search/ESGF_Search_RESTful_API.html).
+
+# Query the ESGF portal
+
+epwshiftr provides `query_esgf()` to interact with the ESGF serach API. It
+returns an `EsgfQuery` object, which is an `R6` object. `query_esgf()` requires
+a `host` input, which is the URL to the ESGF Search API service. This should be
+the URL of the ESGF search service excluding the final endpoint name. Usually
+this is `http://<hostname>/esg-search`. Default is set to the
+[LLNL (Lawrence Livermore National Laboratory) Index Node](http://esgf-node.llnl.gov), i.e., `"https://esgf-node.llnl.gov/esg-search"`.
+
+```{r query}
+query <- query_esgf()
+
+class(query)
+```
+
+When created, an `EsgfQuery` object sets some default query parameters,
+including `project`, `fields`, `latest`, `type`, `offset`, `distrib`, `limit`
+and `format`. As you can see when print the `query` object we just created.
+
+```{r}
+query
+```
+
+The `EsgfQuery` class is the workhorse for dealing with ESGF search services
+with quite a few methods that can be classified into 3 categories:
+
+- **Value listing**: methods to list all possible values of facets, shards,
+  etc.
+- **Parameter getter & setter**: methods to get the query parameter values or
+  set them before sending the actual query to the ESGF search services.
+- **Query responses**: methods to collect results for the query response.
+
+## Value listing
+
+When creating an `EsgfQuery` object, a
+[facet listing query](https://esgf.github.io/esg-search/ESGF_Search_RESTful_API.html#facet-listings)
+can be sent to the index node to get all available facets and shards for the
+default project (CMIP6).
+
+```{r, eval = FALSE}
+query_esgf(listing = TRUE)
+```
+
+You can also call the `$build_listing()` method to send the facet listing after the `EsgfQuery` was created.
+
+> NOTE: Not all index nodes support facet listing query. You may encounter error
+> if your input host does not support it.
+
+```{r query-list}
+query$build_listing()
+```
+
+`EsgfQuery` class provides three value-listing methods to extract data from the
+response of the facet listing query:
+
+- `EsgfQuery$list_all_facets()`: List all available facet names.
+- `EsgfQuery$list_all_fields()`: List all available field names.
+- `EsgfQuery$list_all_shards()`: List all available shards.
[…]
