### 2023-05-16 2a80e68
[fix] Update `last_result` schema
