### 2023-05-16 3afbb89
[doc] Rename dataset in query vignette
